% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_content.R
\name{aboutContent}
\alias{aboutContent}
\alias{GenEstAuthors}
\alias{GenEstGUIauthors}
\alias{GenEstLicense}
\alias{GenEstAcknowledgements}
\alias{GenEstLogos}
\title{Create the Content for the About Pabel}
\usage{
aboutContent()

GenEstAuthors()

GenEstGUIauthors()

GenEstLicense()

GenEstAcknowledgements()

GenEstLogos()
}
\value{
HTML for the Help About main panel.

\code{GenEstAuthors}: HTML for the Help About panel Author text.

\code{GenEstGUIauthors}: HTML for the Help About panel GUI Author
  text.

\code{GenEstLicense}: HTML for the Help About panel License text.

\code{GenEstAcknowledgements}: HTML for the Help About panel 
  Acknowledgements text.

\code{GenEstLogos}: HTML for the Help About panel logos.
}
\description{
This set of functions create the HTML code for the Help About 
  main panel, which gives information about the product and project. \cr
  \cr \code{aboutContent} collates the authors (\code{GenEstAuthors} and
  \code{GenEstGUIAuthors}), license (\code{GenEstLicense}), 
  acknowledgements (\code{GenEstAcknowledgements}), and logo 
  (\code{GenEstLogos}) contents.

\code{GenEstAuthors} creates the HTML code for the Authors
  section.

\code{GenEstGUIauthors} creates the HTML code for the GUI
  Authors section.

\code{GenEstLicense} creates the HTML code for the License
  section.

\code{GenEstAcknowledgements} creates the HTML code for the
  Acknowledgements section.

\code{GenEstLogos} creates the HTML code for the Logos
  section.
}
