% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_about_functions.R
\name{aboutMainPanel}
\alias{aboutMainPanel}
\title{Help About main panel UI element}
\usage{
aboutMainPanel()
}
\value{
HTML for the Help About main panel
}
\description{
create the HTML code for the Help About main panel
}
