% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_about_functions.R
\name{aboutPanel}
\alias{aboutPanel}
\title{Help About Main Panel UI element}
\usage{
aboutPanel()
}
\value{
HTML for the Help About panel
}
\description{
create the HTML code for the Help About panel
}
