% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{AIC.cpmSetSize}
\alias{AIC.cpmSetSize}
\title{Create the AICc tables for a list of sets of searcher efficiency models}
\usage{
\method{AIC}{cpmSetSize}(x, ...)
}
\arguments{
\item{x}{List of sets of CP models fit to the same observations}
}
\value{
AICc table
}
\description{
S3 function to generate model comparison tables for lists of
 of sets of CP models of class \code{\link{pkmSetSize}}
}
\examples{
 cpmods <- cpmSetSize(formula_l = l ~ Visibility, data = wind_RP$data_CP,
   left = "LastPresent", right = "FirstAbsent", sizeclassCol = "Size")
 AIC(cpmods)

}
