% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R,
%   R/utility_functions.R
\name{AIC.pkm}
\alias{AIC.pkm}
\title{extract AIC value from pkm object}
\usage{
\method{AIC}{pkm}(x)

\method{AIC}{pkm}(x)
}
\arguments{
\item{x}{object of class \code{pkm}}

\item{x}{object of class \code{pkm}}
}
\value{
AIC and AICc of the model

AIC and AICc of the model
}
\description{
extract AIC value from pkm object

extract AIC value from pkm object
}
