% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{aicc.pkm}
\alias{aicc.pkm}
\title{extract AIC value from pkm object}
\usage{
\method{aicc}{pkm}(x, ...)
}
\arguments{
\item{x}{object of class \code{pkm}}

\item{...}{further arguments passed to or from other methods}
}
\value{
AIC and AICc of the model
}
\description{
extract AIC value from pkm object
}
