% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_analysis_functions.R
\name{analysisPanel}
\alias{analysisPanel}
\title{Analysis Main Panel UI element}
\usage{
analysisPanel()
}
\value{
HTML for the Analysis panel
}
\description{
create the HTML code for the Analysis panel
}
