% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{app_download_functions}
\alias{app_download_functions}
\alias{downloadCPFig}
\alias{downloadCPmod}
\alias{downloadSEFig}
\alias{downloadSEmod}
\alias{downloadgFig}
\alias{downloadMres}
\alias{downloadMFig}
\alias{downloadTable}
\alias{downloadgres}
\title{GenEst app download funtions}
\usage{
downloadCPFig(rv)

downloadCPmod(rv, input)

downloadSEFig(rv)

downloadSEmod(rv, input)

downloadgFig(rv, sc)

downloadMres(rv, input)

downloadMFig(rv, split = TRUE)

downloadTable(filename, tablename, csvformat)

downloadgres(rv, input)
}
\arguments{
\item{rv}{the reactive values list}

\item{input}{list of shiny input parameters}

\item{sc}{size class}

\item{split}{logical indicator to use the split or not}

\item{filename}{the name for the file writing out}

\item{tablename}{the name of the table in the rv list}

\item{csvformat}{format for .csv file: "" or NULL for comma-separated, 2
for semi-colon separated}
}
\value{
download handler functions
}
\description{
Handle downloading of app data and figures
}
