% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{app_msg_functions}
\alias{app_msg_functions}
\alias{msgList}
\alias{clearNotifications}
\alias{msgModRun}
\alias{msgModDone}
\alias{msgModPartialFail}
\alias{msgSampleSize}
\alias{msgModWarning}
\alias{msgModSENobs}
\alias{msgModFail}
\alias{msgSSavgFail}
\alias{msgSSinputFail}
\alias{msgSplitFail}
\alias{msgFracNote}
\title{GenEst App Messages}
\usage{
msgList()

clearNotifications(msgs = msgList(), clear = TRUE)

msgModRun(msgs, modelType, clear = TRUE)

msgModDone(msgs, rv, type = "SE", clear = TRUE)

msgModPartialFail(mods, type = "SE")

msgSampleSize(mods)

msgModWarning(mods, type = "SE", rv = NULL)

msgModSENobs(rv)

msgModFail(mods, type = "SE", special = NULL)

msgSSavgFail(msgs, rv, clear = TRUE)

msgSSinputFail(msgs, rv, clear = TRUE)

msgSplitFail(type = NULL)

msgFracNote(fracNote)
}
\arguments{
\item{msgs}{message list}

\item{clear}{logical indicator if clearing should happen.}

\item{modelType}{"SE", "CP", "g", or "M"}

\item{rv}{reactive values list}

\item{type}{"SE", "CP", "M", "split", or "g"}

\item{mods}{Set Size list of models}

\item{special}{indicator of a special type of message}

\item{fracNote}{the note regarding the input}
}
\description{
lists of messages used in the app
}
