% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_utilities.R
\name{app_output_utilities}
\alias{app_output_utilities}
\alias{classText}
\alias{estText}
\alias{setNotSuspending}
\alias{reNULL}
\alias{initialOutput}
\alias{setFigW}
\alias{setFigH}
\title{app utilities for formatting text, tables, figs, etc. for display}
\usage{
classText(rv, type = "SE")

estText(rv, type = "SE")

setNotSuspending(output, dontSuspend)

reNULL(x, toNULL)

initialOutput(rv, output)

setFigW(modelSet)

setFigH(modelSet, modType = "SE")
}
\arguments{
\item{rv}{Reactive values list for the GenEst GUI.}

\item{type}{Model type, either "SE" or "CP" or "g".}

\item{output}{\code{output} list to have elements \code{dontSuspend}
(re)set to having \code{suspendWhenHidden = FALSE}.}

\item{dontSuspend}{Names of elements in \code{output} to (re)set to
having \code{suspendWhenHidden = FALSE}.}

\item{x}{\code{list} object to have elements \code{toNULL} reset to
\code{NULL}.}

\item{toNULL}{Names of elements in \code{x} to reset to \code{NULL}.}

\item{modelSet}{Model set of class \code{cpmSet} or \code{pkmSet}.}

\item{modType}{"SE" or "CP"}
}
\description{
app utilities for formatting text, tables, figs, etc. for display
}
