% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_panels.R
\name{app_panels}
\alias{app_panels}
\alias{dataTabPanel}
\alias{selectedDataPanel}
\alias{modelOutputPanel}
\title{app panel utility functions}
\usage{
dataTabPanel(dataType)

selectedDataPanel(modType)

modelOutputPanel(outType)
}
\arguments{
\item{dataType}{Toggle control for the model type of the panel. One of
"SE", "CP", "SS", "DWP", or "CO".}

\item{modType}{Toggle control for the model type of the panel. One of
"SE", "CP", or "g".}

\item{outType}{Toggle control for the model type of the panel. One of
"SEFigures", "SEEstimates", "SEModComparison", "SEModSelection",
"CPFigures", "CPEstimates", "CPModComparison", "CPModSelection",
"gFigures", or "gSummary".}
}
\description{
functions for formatting and displaying app panels
}
