% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{app_ui}
\alias{app_ui}
\alias{GenEstUI}
\alias{dataInputPanel}
\alias{dataInputSidebar}
\alias{loadedDataPanel}
\alias{analysisPanel}
\alias{GeneralInputsPanel}
\alias{GeneralInputSidebar}
\alias{SEPanel}
\alias{SESidebar}
\alias{SEMainPanel}
\alias{CPPanel}
\alias{CPSidebar}
\alias{CPMainPanel}
\alias{MPanel}
\alias{MSidebar}
\alias{MMainPanel}
\alias{gPanel}
\alias{gSidebar}
\alias{gMainPanel}
\alias{helpPanel}
\alias{gettingStartedPanel}
\alias{downloadsPanel}
\alias{aboutPanel}
\alias{disclaimersPanel}
\title{Create the GenEst User Interface HTML}
\usage{
GenEstUI(appType = "base")

dataInputPanel()

dataInputSidebar()

loadedDataPanel()

analysisPanel()

GeneralInputsPanel()

GeneralInputSidebar()

SEPanel()

SESidebar()

SEMainPanel()

CPPanel()

CPSidebar()

CPMainPanel()

MPanel()

MSidebar()

MMainPanel()

gPanel()

gSidebar()

gMainPanel()

helpPanel(appType = "base")

gettingStartedPanel()

downloadsPanel()

aboutPanel()

disclaimersPanel(appType = "base")
}
\arguments{
\item{appType}{Toggle control for the app, \code{"base"} for local versions
or \code{"deploy"} for hosted version. Currently only differentiates the
disclaimer text.}
}
\value{
Each function returns a string of HTML code, either as a
  \code{"shiny.tag.list"} object (in the case of \code{GenEstUI}) or a
  \code{"shiny.tag"} object (in the case of the other functions). \cr \cr 
  \code{GenEstUI}: Full GenEst user interface.
}
\description{
This suite of functions in \code{app_ui.R} create the HTML code
 underlying the GenEst user interface (UI). See the "GenEst Graphic User
 Interface" vignette for a more complete detailing of the codebase underlying
  the GenEst UI. \cr \cr \code{GenEstUI}: whole application. Calls 
  \code{dataInputPanel}, \code{analysisPanel}, and \code{helpPanel}.
}
\details{
Currently there are few differences between the local and deployed
  versions of GenEst, and the \code{appType} toggle is only included as an 
  argument for functions that can produce different versions of the HTML.
  At this point, the only content that is different is the disclaimer text
  on the Help panel.
}
