% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_utilities.R
\name{app_ui_utilities}
\alias{app_ui_utilities}
\alias{navbar}
\alias{style}
\alias{ol}
\alias{ul}
\alias{li}
\alias{b}
\alias{u}
\alias{small}
\alias{big}
\alias{center}
\alias{GenEstInlineCSS}
\alias{GenEstShinyJS}
\alias{cButtonStyle}
\title{HTML parameters}
\usage{
navbar()

style(...)

ol(...)

ul(...)

li(...)

b(...)

u(...)

small(...)

big(text = NULL)

center(text = NULL)

GenEstInlineCSS(...)

GenEstShinyJS(...)

cButtonStyle(buttonType = "single")
}
\arguments{
\item{...}{attributes and children of the element}

\item{text}{text to wrap in the tag}

\item{buttonType}{"single" (for clearing a single component) or "all" (for
clearing everything).}
}
\description{
utility functions for simple HTML tasks
}
