% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{averageSS}
\alias{averageSS}
\title{Tabulate an average search schedule from a multi-unit SS data table}
\usage{
averageSS(data_SS, datesSearchedCol = "DateSearched", unitPrefix = "Unit")
}
\arguments{
\item{data_SS}{a multi-unit SS data table, for which the average interval 
will be tabulated}

\item{datesSearchedCol}{Column name for the date searched data}

\item{unitPrefix}{prefix for the unit identifiers, which correspond to 
columns in \code{data_ss}}
}
\value{
vector of the average search schedule
}
\description{
Tabulate an average search schedule from a multi-unit SS data table
}
