% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{b}
\alias{b}
\title{HTML b function}
\usage{
b(...)
}
\arguments{
\item{...}{attributes and children of the element}
}
\value{
HTML ordered list
}
\description{
Generate an HTML bolded text element. This simply pulls 
  the function definition from the tag environment in htmltools
}
