% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{big}
\alias{big}
\title{HTML big div}
\usage{
big(text)
}
\arguments{
\item{text}{text to wrap in \code{<big> ... </big>}}
}
\value{
HTML small text
}
\description{
Generate a big text HTML object.
}
