% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{calcMhat}
\alias{calcMhat}
\title{Calculate Mhat for a given Mtilde and DWP}
\usage{
calcMhat(Mtilde, DWP = 1)
}
\arguments{
\item{Mtilde}{Mtilde value}

\item{DWP}{Density weighted proportion associated with Mtilde}
}
\value{
Mhat
}
\description{
Calculate Mhat for a given Mtilde and DWP
}
\examples{
NA
}
