% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{center}
\alias{center}
\title{HTML center div}
\usage{
center(text)
}
\arguments{
\item{text}{text to wrap in \code{<center> ... </center>}}
}
\value{
HTML small text
}
\description{
Generate a center text HTML object.
}
