% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{checkSpecificModelCP}
\alias{checkSpecificModelCP}
\title{Error check a specific model selection for a CP plot}
\usage{
checkSpecificModelCP(modelSet, specificModel)
}
\arguments{
\item{modelSet}{cp model set of class cpmSet}

\item{specificModel}{the name of the specific model for the plot}
}
\value{
updated name of the model to use
}
\description{
Make sure it's available and good, update the name for usage
}
