% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{checkSpecificModelSE}
\alias{checkSpecificModelSE}
\title{Error check a specific model selection for an SE plot}
\usage{
checkSpecificModelSE(modelSet, specificModel)
}
\arguments{
\item{modelSet}{pk model set of class pkmSet}

\item{specificModel}{the name of the specific model for the plot}
}
\value{
updated name of the model to use
}
\description{
Make sure it's available and good, update the name for usage
}
