% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_utilities.R
\name{classText}
\alias{classText}
\title{Prepare class text header}
\usage{
classText(rv, type = "SE")
}
\arguments{
\item{rv}{Reactive values list for the GenEst GUI.}

\item{type}{Model type, either "SE" or "CP" or "g".}
}
\value{
Rendered text ready for export to \code{output} list.
}
\description{
Depending on the classes, prepare the header text.
}
