% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{clearNotifications}
\alias{clearNotifications}
\title{Clear all notifications}
\usage{
clearNotifications(msgs = msgList(), clear = TRUE)
}
\arguments{
\item{msgs}{message list}
}
\description{
Clear all messages in the message list
}
