% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{CO_DWP}
\alias{CO_DWP}
\title{Associate CO carcasses with appropriate DWP values (by unit and carcass class)}
\usage{
CO_DWP(dwpsim, data_CO, unitCol, sizeCol = NULL)
}
\arguments{
\item{dwpsim}{\code{rdwp} object}

\item{data_CO}{data frame with results from carcass surveys}

\item{unitCol}{name of the unit column in data_CO (required)}

\item{sizeCol}{name of the carcass class column in data_CO (optional).}
}
\value{
numeric DWP array
}
\description{
Calculate the conditional probability of observing a carcass
  at search oi as a function arrival interval (assuming carcass is not
  removed by scavengers before the time of the final search)
}
