% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{combine_factors}
\alias{combine_factors}
\title{Create the factor combination table for a searcher efficiency or carcass
 persistence analysis.}
\usage{
combine_factors(predictors, data)
}
\arguments{
\item{predictors}{Names of predictor variables to include.}

\item{data}{Searcher efficiency or carcass persistence data.}
}
\value{
Factor combination table.
}
\description{
Create the factor combination table for a searcher efficiency or carcass
 persistence analysis.
}
\examples{
NA
}
