% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{combine_factors_across_models}
\alias{combine_factors_across_models}
\title{Create the factor combination table for a set of searcher efficiency 
 and carcass persistence analyses.}
\usage{
combine_factors_across_models(cp_predictors, se_predictors, cp_data, se_data)
}
\arguments{
\item{cp_predictors}{Carcass persistence predictor variables.}

\item{se_predictors}{Seacher efficiency predictor variables.}

\item{cp_data}{Carcass persistence data.}

\item{se_data}{Seacher efficiency data.}
}
\value{
Factor combination table.
}
\description{
Create the factor combination table for a set of searcher efficiency 
 and carcass persistence analyses.
}
\examples{
NA
}
