% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{combinePreds}
\alias{combinePreds}
\title{Create the predictor combination table for a searcher efficiency or carcass
 persistence analysis.}
\usage{
combinePreds(preds, data)
}
\arguments{
\item{preds}{Names of predictor variables to include.}

\item{data}{Searcher efficiency or carcass persistence data.}
}
\value{
Predictor combination table.
}
\description{
Create the predictor combination table for a searcher efficiency or carcass
 persistence analysis.
}
\examples{
NA
}
