% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{condense_mhat}
\alias{condense_mhat}
\title{Condense the Mhat array across units and size classes.}
\usage{
condense_mhat(Mhatarray)
}
\arguments{
\item{Mhatarray}{Mhat array.}
}
\value{
Matrix of Mhat values (rows: iterations, columns: splits).
}
\description{
Condense the Mhat array across units and size classes.
}
\examples{
NA
}
