% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{countCarcs}
\alias{countCarcs}
\title{Count the minimum number of carcasses in the cells}
\usage{
countCarcs(mods)
}
\arguments{
\item{mods}{model output from the _SetSize version of a function}
}
\value{
the minimum number of carcasses in the cells
}
\description{
Count the minimum number of carcasses in the cells
}
