% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cp_aicc_table_create}
\alias{cp_aicc_table_create}
\title{Create the  AICc tables for the carcass persistence models}
\usage{
cp_aicc_table_create(models)
}
\arguments{
\item{models}{Carcass persistence models fit for each size class.}
}
\value{
List of AICc tables (one table per size class)
}
\description{
Create the  AICc tables for the carcass persistence models
}
\examples{
NA
}
