% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cp_model_fit}
\alias{cp_model_fit}
\title{Fit a single carcass persistence model for a single size class}
\usage{
cp_model_fit(data, surv_object, equation_choice, distribution_choice)
}
\arguments{
\item{data}{Carcass persistence data restricted to a single size class.}

\item{surv_object}{Survival object for a single size class.}

\item{equation_choice}{Equation to use.}

\item{distribution_choice}{Choice of distribution 
(exponential, weibull, loglogistic, lognormal).}
}
\value{
Model fit list.
}
\description{
Fit a single carcass persistence model for a single size class
}
\examples{
NA
}
