% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cp_model_set_across_sizes_fit}
\alias{cp_model_set_across_sizes_fit}
\title{Fit all possible carcass persistence models across all size classes}
\usage{
cp_model_set_across_sizes_fit(data, predictors, size_class_column,
  last_time_present_column, first_time_absent_column)
}
\arguments{
\item{data}{Carcass persistence data across all size classes.}

\item{predictors}{Names of predictor variables to include.}

\item{size_class_column}{Column header for the size class column.}

\item{last_time_present_column}{Column name of the last time present data.}

\item{first_time_absent_column}{Column name of the first time absent data.}
}
\value{
List of model fit lists.
}
\description{
Fit all possible carcass persistence models across all size classes
}
\examples{
NA
}
