% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cp_model_set_fit}
\alias{cp_model_set_fit}
\title{Fit all possible carcass persistence models for a single size class}
\usage{
cp_model_set_fit(data, predictors, last_time_present_column,
  first_time_absent_column)
}
\arguments{
\item{data}{Carcass persistence data restricted to a single size class.}

\item{predictors}{Names of predictor variables to include.}

\item{last_time_present_column}{Column name of the last time present data.}

\item{first_time_absent_column}{Column name of the first time absent data.}
}
\value{
List of model fit lists.
}
\description{
Fit all possible carcass persistence models for a single size class
}
\examples{
NA
}
