% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cp_theta_create}
\alias{cp_theta_create}
\title{Draw replicate samples of parameters from the carcass persistence models 
across all size classes}
\usage{
cp_theta_create(data, predictors, size_class_column, model_fits, replicates)
}
\arguments{
\item{data}{Carcass persistence data.}

\item{predictors}{Names of predictor variables to include.}

\item{size_class_column}{Column header for the size class column.}

\item{model_fits}{Carcass persistence model fits.}

\item{replicates}{Number of replicate samples to draw from the 
distribution.}
}
\value{
Multidimensional array (replicates, 2, factor combination 
        cells, models, size classes) of carcass persistence.
}
\description{
Draw replicate samples of parameters from the carcass persistence models 
across all size classes
}
\examples{
NA
}
