% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{CPcols}
\alias{CPcols}
\title{Produce a named vector of standard CP plot colors}
\usage{
CPcols()
}
\description{
Produce a named vector of standard CP plot colors
}
