% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{CPdistOptions}
\alias{CPdistOptions}
\title{Produce the options for the distributions in the CP model}
\usage{
CPdistOptions()
}
\value{
list with named elements of the distributions
}
\description{
Produce the options for the distributions in the CP model
}
