% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_CP_functions.R
\name{CPEstimatesPanel}
\alias{CPEstimatesPanel}
\title{Analysis Carcass Persistence main panel estimates UI element}
\usage{
CPEstimatesPanel()
}
\value{
HTML for the Analysis Carcass Persistence model estimates panel
}
\description{
create the HTML code for the Analysis Carcass Persistence
  main panel model estimates element
}
