% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_CP_functions.R
\name{CPFiguresPanel}
\alias{CPFiguresPanel}
\title{Analysis Carcass Persistence main panel figures UI element}
\usage{
CPFiguresPanel()
}
\value{
HTML for the Analysis Carcass Persistence figures panel
}
\description{
create the HTML code for the Analysis Carcass Persistence
  main panel figures element
}
