% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpLogLik}
\alias{cpLogLik}
\title{Calculate the negative log-likelihood of a carcass persistence model}
\usage{
cpLogLik(t1, t2, beta, nbeta_l, cellByCarc, cellMM, dataMM, dist)
}
\arguments{
\item{t1}{last times observed present}

\item{t2}{first times observed absent}

\item{beta}{Parameters to be optimized.}

\item{nbeta_l}{Number of parameters associated with l.}

\item{cellByCarc}{Which cell each observation belongs to.}

\item{cellMM}{Combined model matrix.}

\item{dataMM}{Combined model matrix expanded to the data.}

\item{dist}{Name of distribution.}
}
\value{
Negative log likelihood of the observations, given the parameters.
}
\description{
The function used to calculate the negative-loglikelihood of
  a given carcass persistence model (\code{\link{cpm}}) with a given data
  set
}
