% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_CP_functions.R
\name{CPMainPanel}
\alias{CPMainPanel}
\title{Analysis Carcass Persistence main panel UI element}
\usage{
CPMainPanel()
}
\value{
HTML for the Analysis Carcass Persistence main panel
}
\description{
create the HTML code for the Analysis Carcass Persistence
  main
}
