% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmAllFail}
\alias{cpmAllFail}
\title{Check if all of the cpm models fail}
\usage{
cpmAllFail(cpmToCheck)
}
\arguments{
\item{cpmToCheck}{A \code{cpm} model or a set of them or a suite of sets
associated with multiple sizes}
}
\value{
A single logical value indicating if all of the models failed
}
\description{
Check if all of the cpm models fail
}
