% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmCheck}
\alias{cpmCheck}
\title{Verify that a suite of carcass persistence models all fit successfully.}
\usage{
cpmCheck(cpmToCheck)
}
\arguments{
\item{cpmToCheck}{A \code{cpm} model or a set of them or a suite of sets
associated with multiple sizes}
}
\value{
A single (total) logcal
}
\description{
Verify that a suite of carcass persistence models all fit successfully.
}
