% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{cpmCPCellPlot}
\alias{cpmCPCellPlot}
\title{Plot cell-specific decay curve for carcass persistence}
\usage{
cpmCPCellPlot(model, specificCell, col, axis_y = TRUE, axis_x = TRUE)
}
\arguments{
\item{model}{model of class cpm}

\item{specificCell}{name of the specific cell to plot}

\item{col}{color to use}

\item{axis_y}{logical of whether or not to plot the y axis}

\item{axis_x}{logical of whether or not to plot the x axis}
}
\description{
Produce the figure panel for a specific cell (factor
  level combination) including the specific fitted decay curves.
}
