% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmFail}
\alias{cpmFail}
\title{Check if a CP model is well-fit}
\usage{
cpmFail(cpmod)
}
\arguments{
\item{cpmod}{A \code{\link{cpm}} object to test}
}
\value{
logical value indicating a failed fit (TRUE) or successful (FALSE)
}
\description{
Run a check the arg is a well-fit cpm object
}
