% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_CP_functions.R
\name{CPModComparisonPanel}
\alias{CPModComparisonPanel}
\title{Analysis Carcass Persistence main panel model comparison UI element}
\usage{
CPModComparisonPanel()
}
\value{
HTML for the Analysis Carcass Persistence model comparison panel
}
\description{
create the HTML code for the Analysis Carcass Persistence
  main panel model comparison element
}
