% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_CP_functions.R
\name{CPModSelectionPanel}
\alias{CPModSelectionPanel}
\title{Analysis Carcass Persistence main panel model selection UI element}
\usage{
CPModSelectionPanel()
}
\value{
HTML for the Analysis Carcass Persistence model selection panel
}
\description{
create the HTML code for the Analysis Carcass Persistence
  main panel model selection element
}
