% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmSet}
\alias{cpmSet}
\title{Run a set of cpm models based on predictor inputs}
\usage{
cpmSet(formula_l, formula_s = NULL, data, left = NULL, right = NULL,
  dists = c("exponential", "weibull", "lognormal", "loglogistic"),
  CL = 0.95, quiet = FALSE)
}
\arguments{
\item{formula_l}{Formula for location; an object of class 
"\code{\link{formula}}" (or one that can be coerced to that class):
a symbolic description of the model to be fitted. Details of model 
specification are given under 'Details'.}

\item{formula_s}{Formula for scale; an object of class 
"\code{\link{formula}}" (or one that can be coerced to that class):
a symbolic description of the model to be fitted. Details of model 
specification are given under 'Details'.}

\item{data}{Dataframe with results from carcass persistence trials and any
covariates included in \code{formula_l} or {formula_s} (required).}

\item{left}{Name of columns in \code{data} where the time of last present
observation is stored.}

\item{right}{Name of columns in \code{data} where the time of first absent
observation is stored.}

\item{dists}{Names of the distributions (from "exponential", "weibull", 
"loglogistic", and "lognormal") that are to be included}

\item{CL}{confidence level}

\item{quiet}{Logical indicator of whether or not to print messsages}
}
\value{
\code{cpmSet} returns a class-\code{cpmSet} list of objects, each
  of class "\code{cpm}", which each then a list whose components 
  characterize the fit of the specific model.
}
\description{
Run a set of \code{\link{cpm}} models based on all possible 
  models, given the predictor inputs. \code{cpmSet}'s inputs generally
  follow \code{\link{cpm}}, with all simpler models being run for all 
  included distributions and returned as a list of model objects
}
\examples{
  data(wind_RP)
  mod <- cpmSet(formula_l = l ~ Season, formula_s = s ~ Season,  
           data = wind_RP$CP, left = "Left", right = "Right"
          )

}
