% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmSetAICcTab}
\alias{cpmSetAICcTab}
\title{Create the  AICc tables for a set of carcass persistence models}
\usage{
cpmSetAICcTab(cpmset)
}
\arguments{
\item{cpmset}{Set of carcass persistence models fit to the same
observations}
}
\value{
AICc table
}
\description{
Create the  AICc tables for a set of carcass persistence models
}
\examples{
NA
}
