% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmSetFail}
\alias{cpmSetFail}
\title{Check if cpm models fail}
\usage{
cpmSetFail(cpmSetToCheck)
}
\arguments{
\item{cpmSetToCheck}{A \code{\link{cpmSet}} object to test}
}
\value{
A vector of logical values indicating if each of the models failed
}
\description{
Run a check on each model within a \code{\link{cpmSet}} object
  to determine if it failed or not
}
