% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmSetFailRemove}
\alias{cpmSetFailRemove}
\title{Remove failed pkm models}
\usage{
cpmSetFailRemove(cpmSetToTidy)
}
\arguments{
\item{cpmSetToTidy}{A \code{cpmSet} object to tidy}
}
\value{
A \code{cpmSet} object with failed models removed
}
\description{
Remove failed pkm models
}
