% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmSetSize}
\alias{cpmSetSize}
\title{Fit all possible carcass persistence models across all size classes}
\usage{
cpmSetSize(formula_l, formula_s = NULL, data, left = NULL,
  right = NULL, dists = c("exponential", "weibull", "lognormal",
  "loglogistic"), sizeclassCol = NULL, CL = 0.95, quiet = FALSE)
}
\arguments{
\item{formula_l}{Formula for location; an object of class 
"\code{\link{formula}}" (or one that can be coerced to that class):
a symbolic description of the model to be fitted. Details of model 
specification are given under 'Details'.}

\item{formula_s}{Formula for scale; an object of class 
"\code{\link{formula}}" (or one that can be coerced to that class):
a symbolic description of the model to be fitted. Details of model 
specification are given under 'Details'.}

\item{data}{Dataframe with results from carcass persistence trials and any
covariates included in \code{formula_l} or {formula_s} (required).}

\item{left}{Name of columns in \code{data} where the time of last present
observation is stored.}

\item{right}{Name of columns in \code{data} where the time of first absent
observation is stored.}

\item{dists}{Names of the distributions (from "exponential", "weibull", 
"loglogistic", and "lognormal") that are to be included}

\item{sizeclassCol}{Name of colum in \code{data} where the size classes
are recorded}

\item{CL}{confidence level}

\item{quiet}{Logical indicator of whether or not to print messsages}
}
\value{
\code{cpmSetSize} returns a class-\code{cpmSetSize} list of 
  objects, each of which is a class-\code{cpmSet} list of \code{cpm}" 
  outputs (each corresponding to the fit of a specific model
  within the set of \code{cpm} models fit for the given size class), that
  is of length equal to the total number of size classes
}
\description{
Run a set of \code{\link{cpmSet}} model set runs based on all 
  possible models for a suite of size classes. \code{cpmSetSize}'s inputs
  generally follow \code{\link{cpmSet}} and \code{\link{cpm}} but with an 
  additional size column input and calculation of the set of cpm models for 
  each of the size classes.
}
\examples{
  data(wind_RP)
  mod <- cpmSetSize(formula_l = l ~ Season, formula_s = s ~ Season,  
           data = wind_RP$CP, left = "Left", right = "Right",
           sizeclassCol = "Size"
          )

}
