% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmSetSizeFail}
\alias{cpmSetSizeFail}
\title{Check if all of the cpm models fail}
\usage{
cpmSetSizeFail(cpmSetSizeToCheck)
}
\arguments{
\item{cpmSetSizeToCheck}{A \code{cpmSetSize} object to test}
}
\value{
A list of vectors of logical values indicating if each of the 
  models failed
}
\description{
Check if all of the cpm models fail
}
