% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmSetSizeFailRemove}
\alias{cpmSetSizeFailRemove}
\title{Remove failed cpm models from a \code{\link{cpmSetSize}} object}
\usage{
cpmSetSizeFailRemove(cpmSetSizeToTidy)
}
\arguments{
\item{cpmSetSizeToTidy}{A list of \code{\link{cpmSetSize}} objects to tidy}
}
\value{
A list of \code{\link{cpmSet}} objects with failed models removed
}
\description{
Remove failed models from a \code{\link{cpmSetSize}} object
}
