% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{cpmSetSpecCPCellPlot}
\alias{cpmSetSpecCPCellPlot}
\title{Plot cell-specific decay curve for carcass persistence}
\usage{
cpmSetSpecCPCellPlot(modelSet, specificModel, fullModel, specificCell, col,
  modelMatches, axis_x, axis_y, n, seed)
}
\arguments{
\item{modelSet}{model set of class cpmSet}

\item{specificModel}{name of the specific model to plot and emphasize}

\item{fullModel}{name of the full model to create the cell structure}

\item{specificCell}{name of the specific cell to plot}

\item{col}{color to use}

\item{modelMatches}{list of matching models for the modelSet}

\item{axis_x}{logical of whether or not to plot the x axis}

\item{axis_y}{logical of whether or not to plot the y axis}

\item{n}{number of draws to use to characterize the distributions}

\item{seed}{random number seed to use}
}
\description{
Plot cell-specific decay curve for carcass persistence
}
