% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_CP_functions.R
\name{CPPanel}
\alias{CPPanel}
\title{Analysis Carcass Persistence Main Panel UI element}
\usage{
CPPanel()
}
\value{
HTML for the Analysis CP panel
}
\description{
create the HTML code for the Analysis CP panel
}
