% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_CP_functions.R
\name{CPSelectedDataPanel}
\alias{CPSelectedDataPanel}
\title{Analysis Carcass Persistence main panel selected data UI element}
\usage{
CPSelectedDataPanel()
}
\value{
HTML for the Analysis Carcass Persistence selected data panel
}
\description{
create the HTML code for the Analysis Carcass Persistence
  main panel selected data element
}
