% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_CP_functions.R
\name{CPSidebar}
\alias{CPSidebar}
\title{Analysis Carcass Persistence sidebar panel UI element}
\usage{
CPSidebar()
}
\value{
HTML for the Analysis Carcass Persistence sidebar
}
\description{
create the HTML code for the Analysis Carcass Persistence
  sidebar
}
