% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{create_DWP_table}
\alias{create_DWP_table}
\title{Determine the density weighted proportion (DWP) of area searched for
 each size by search schedule by unit combination.}
\usage{
create_DWP_table(data)
}
\arguments{
\item{data}{Search schedule data.}
}
\value{
Table of DWP values for each size, search schedule, and unit.
}
\description{
Determine the density weighted proportion (DWP) of area searched for
 each size by search schedule by unit combination.
}
\examples{
NA
}
