% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{create_g_table}
\alias{create_g_table}
\title{Summarize the replicates of detection probability according to each 
 size class, search schedule, and cell combination.}
\usage{
create_g_table(garray, confidence_level)
}
\arguments{
\item{garray}{Array of detection probabilities.}

\item{confidence_level}{Confidence level for the summary.}
}
\value{
Table of detection probabilities with means and confidence level
}
\description{
Summarize the replicates of detection probability according to each 
 size class, search schedule, and cell combination.
}
\examples{
NA
}
