% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{create_mhat_figure}
\alias{create_mhat_figure}
\title{Create the Mhat figure for a given split category.}
\usage{
create_mhat_figure(condensed_mhat_split, split_category_name,
  fraction_area_sampled)
}
\arguments{
\item{condensed_mhat_split}{Condensed Mhat table, for one split only.}

\item{split_category_name}{Name of the split category.}

\item{fraction_area_sampled}{Fraction of the area sampled.}
}
\value{
.
}
\description{
Create the Mhat figure for a given split category.
}
\examples{
NA
}
