% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{create_mhat_table}
\alias{create_mhat_table}
\title{Summarize the Mhat iterations with means and confidence intervals, 
 for searched area and expanded area, according to each split.}
\usage{
create_mhat_table(condensed_mhat, fraction_area_sampled, confidence_level)
}
\arguments{
\item{condensed_mhat}{Condensed mhat array.}

\item{fraction_area_sampled}{Fraction of area sampled.}

\item{confidence_level}{Confidence level for estimates.}
}
\value{
Table of mean (with confidence intervals) of fatalities for 
        the searched area and the expanded area.
}
\description{
Summarize the Mhat iterations with means and confidence intervals, 
 for searched area and expanded area, according to each split.
}
\examples{
NA
}
