% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{create_se_figure}
\alias{create_se_figure}
\title{Create the figures for the searcher efficiency models}
\usage{
create_se_figure(data, predictors, theta, observation_columns, replicates,
  size_class_column, r, j, cellwise)
}
\arguments{
\item{data}{Full searcher efficiency data set.}

\item{predictors}{Predictor variable names for the searcher efficiency 
models.}

\item{theta}{Theta array.}

\item{observation_columns}{Indicator vector of which columns in the data}

\item{replicates}{Number of replicate samples to draw from the 
distribution.}

\item{size_class_column}{Column header for the size class column.}

\item{r}{Size class index.}

\item{j}{Model index.}

\item{cellwise}{Index for the cellwise model.}
}
\value{
NA
}
\description{
Create the figures for the searcher efficiency models
}
\examples{
NA
}
