% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{create_ss_table}
\alias{create_ss_table}
\title{Create search schedule table for presentation.}
\usage{
create_ss_table(data)
}
\arguments{
\item{data}{Search schedule data.}
}
\value{
Search schedule table.
}
\description{
Create search schedule table for presentation.
}
\examples{
NA
}
