% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions_old.R
\name{create_ss_vec}
\alias{create_ss_vec}
\title{Create a vector of concatenated search schedules.}
\usage{
create_ss_vec(data)
}
\arguments{
\item{data}{Search schedule data.}
}
\value{
Vector of search schedules.
}
\description{
Create a vector of concatenated search schedules.
}
\examples{
NA
}
