% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_list_generation_functions.R
\name{createReactiveValues}
\alias{createReactiveValues}
\title{Create the main reactive value list for GenEst}
\usage{
createReactiveValues()
}
\value{
a reactive values list
}
\description{
Create a list of reactive values as used across the components
  of the GenEst application
}
