% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_content.R
\name{createvtext}
\alias{createvtext}
\title{Create the version text for GenEst}
\usage{
createvtext(type = "Full")
}
\arguments{
\item{type}{"Full" or "Short" or "Name" or "NameDate"}
}
\value{
version text
}
\description{
Create a text string of the version number and date
}
