% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_widgets.R
\name{dataDownloadWidget}
\alias{dataDownloadWidget}
\title{Create a Data Download Widget for the GenEst User Interface HTML}
\usage{
dataDownloadWidget(set)
}
\arguments{
\item{set}{Name of data set. One of "RP", "RPbat", "cleared", "powerTower",
"PV", "trough", "mock"}
}
\value{
HTML for the data download widget
}
\description{
This is a generalized function for creating a data download 
  widget (fluid row with name and button) used in the GenEst GUI, based on 
  the data set (\code{set}).
}
