% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_data_input_functions.R
\name{dataInputPanel}
\alias{dataInputPanel}
\title{Data Input Main Panel UI element}
\usage{
dataInputPanel()
}
\value{
HTML for the Data Input panel
}
\description{
create the HTML code for the Data Input panel
}
