% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_data_input_functions.R
\name{dataInputSidebar}
\alias{dataInputSidebar}
\title{Data Input sidebar panel UI element}
\usage{
dataInputSidebar()
}
\value{
HTML for the Data Input sidebar
}
\description{
create the HTML code for the Data Input sidebar
}
