% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataset2}
\alias{dataset2}
\title{Description of the dataset}
\format{A list with 7 items:
\describe{
  \item{SE_trials}{Searcher efficiency trial data}
  \item{CP_trials}{Carcass persistence trial data}
  \item{SS}{Search schedule data}
  \item{DWP_All60}{Density weighted proportion of area searched data}
  \item{DWP_RP120}{Density weighted proportion of area searched data}
  \item{CO_dataAll60}{Carcass observations}  
  \item{CO_dataRP120}{Carcass observations}  
}}
\source{
Source name
}
\usage{
dataset2
}
\description{
A dataset containing SE, CP, SS, DWP, and CO data.
}
\keyword{datasets}
