% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_data_input_functions.R
\name{dataTabPanel}
\alias{dataTabPanel}
\title{Data Input loaded data UI element for a single table}
\usage{
dataTabPanel(tabname, table)
}
\arguments{
\item{tabname}{shiny ID name of the table}

\item{table}{table's name in input object}
}
\value{
HTML for the Data Input data table ouput
}
\description{
create the HTML code for a single Data Input loaded data table
}
