% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{dateCols}
\alias{dateCols}
\title{Selects the date columns from a data table}
\usage{
dateCols(data)
}
\arguments{
\item{data}{data table}
}
\value{
column names of columns that can be coerced to dates
}
\description{
Simple function to facilitate selection of date columns
}
