% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_disclaimers_functions.R
\name{disclaimerDeploy}
\alias{disclaimerDeploy}
\title{Create the WEST disclaimer text for deployment}
\usage{
disclaimerDeploy(type)
}
\arguments{
\item{type}{"base" (for local versions) or "deploy" (for hosted version)}
}
\value{
HTML element of the WEST deployment disclaimer text
}
\description{
Create WEST deployment disclaimer text
}
