% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{disclaimers}
\alias{disclaimers}
\title{Print list of disclaimer statements}
\usage{
disclaimers()
}
\value{
prints disclaimers
}
\description{
Print list of disclaimer statements
}
