% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_content.R
\name{disclaimersContent}
\alias{disclaimersContent}
\alias{disclaimerUSGS}
\alias{disclaimerWEST}
\title{Create the Content for the Disclaimers Pabel}
\usage{
disclaimersContent(appType = "base")

disclaimerUSGS()

disclaimerWEST(appType)
}
\arguments{
\item{appType}{"base" (for local version) or "deploy" (for hosted version)}
}
\value{
HTML for the Help Disclaimers main panel.

\code{disclaimerUSGS}: text for the USGS disclaimer.

\code{disclaimerWEST}: text for the WEST, Inc. disclaimer.
}
\description{
Create the HTML code for the Help Disclaimers main panel,
  which gives the required legal disclaimers for using the application,
  based on \code{appType}. \cr \cr \code{disclaimersContent} collates the
  USGS (\code{disclaimerUSGS} and WEST, Inc. \code{disclaimerWEST})
  disclaimers.

\code{disclaimerUSGS} creates the text for the USGS 
  disclaimer.

\code{disclaimerWEST} creates the text for the WEST, Inc.  
  disclaimer.
}
