% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_disclaimers_functions.R
\name{disclaimersMainPanel}
\alias{disclaimersMainPanel}
\title{Help Disclaimers main panel UI element}
\usage{
disclaimersMainPanel(type)
}
\arguments{
\item{type}{"base" (for local versions) or "deploy" (for hosted version)}
}
\value{
HTML for the Help Disclaimers main panel
}
\description{
create the HTML code for the Help Disclaimers main panel
}
