% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_disclaimers_functions.R
\name{disclaimerUSGS}
\alias{disclaimerUSGS}
\title{Create USGS disclaimer text}
\usage{
disclaimerUSGS()
}
\value{
HTML element of the USGS disclaimer text
}
\description{
Create USGS disclaimer text
}
