% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_disclaimers_functions.R
\name{disclaimerWEST}
\alias{disclaimerWEST}
\title{Create WEST disclaimer text}
\usage{
disclaimerWEST()
}
\value{
HTML element of the WEST text
}
\description{
Create WEST disclaimer text
}
