% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_table_functions.R
\name{dlModTabCP}
\alias{dlModTabCP}
\title{Create the download version of the Carcass Persistence model table}
\usage{
dlModTabCP(modTabs, CL = 0.9)
}
\arguments{
\item{modTabs}{model tables}

\item{CL}{Confidence level}
}
\value{
download version of the SE model table
}
\description{
Format a user-friendly version of the parameter table from
  a Carcass Persistence model, based on confidence level of interest
}
