% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_table_functions.R
\name{dlModTabSE}
\alias{dlModTabSE}
\title{Create the download version of the Searcher Efficiency model table}
\usage{
dlModTabSE(modTab, CL = 0.9)
}
\arguments{
\item{modTab}{model table}

\item{CL}{Confidence level}
}
\value{
download version of the SE model table
}
\description{
Format a user-friendly version of the parameter table from
  a Searcher Efficiency model, based on confidence level of interest
}
