% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadCPFig}
\alias{downloadCPFig}
\title{Download the CP figure}
\usage{
downloadCPFig(rv)
}
\arguments{
\item{rv}{the reactive values list}
}
\value{
a download handler function
}
\description{
Handle the CP figure downloading
}
