% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadCPmod}
\alias{downloadCPmod}
\title{Download summary of CP model fitting}
\usage{
downloadCPmod(rv, input)
}
\arguments{
\item{rv}{the reactive values list,}

\item{input}{list of shiny input parameters}
}
\value{
a download handler function
}
\description{
Handle the CP model downloading
}
