% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadgFig}
\alias{downloadgFig}
\title{Download the g figure}
\usage{
downloadgFig(rv, sc)
}
\arguments{
\item{rv}{the reactive values list}

\item{sc}{size class}
}
\value{
a download handler function
}
\description{
Handle the g figure downloading
}
