% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadMFig}
\alias{downloadMFig}
\title{Download the M figure}
\usage{
downloadMFig(rv)
}
\arguments{
\item{rv}{the reactive values list}
}
\value{
an updated reactive values list
}
\description{
Handle the M figure downloading
}
