% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadMres}
\alias{downloadMres}
\title{Download M results (including SE and CP modeling)}
\usage{
downloadMres(rv, input)
}
\arguments{
\item{rv}{the reactive values list}

\item{input}{shiny input data}
}
\value{
a download handler function
}
\description{
Handle the downloading of results
}
