% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadSEmod}
\alias{downloadSEmod}
\title{Download summary of SE model fitting}
\usage{
downloadSEmod(rv, input)
}
\arguments{
\item{rv}{the reactive values list}

\item{input}{the shiny input data}
}
\value{
a download handler function
}
\description{
Handle the SE model downloading
}
