% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_downloads_functions.R
\name{downloadsMainPanel}
\alias{downloadsMainPanel}
\title{Help Downloads main panel UI element}
\usage{
downloadsMainPanel()
}
\value{
HTML for the Help Downloads main panel
}
\description{
create the HTML code for the Help Downloads main panel
}
