% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_downloads_functions.R
\name{downloadsRow}
\alias{downloadsRow}
\title{Help Downloads specific download UI element}
\usage{
downloadsRow(set)
}
\arguments{
\item{set}{name of data set "RP", "RPbat", "cleared", "powerTower", "PV",
"trough", "mock", or "mock2"}
}
\value{
HTML for the Help Downloads specific download row
}
\description{
create the HTML code for a specific download row within 
  the Help Downloads main panel
}
