% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{DWPCols}
\alias{DWPCols}
\title{Select the DWP-ok columns from a data table}
\usage{
DWPCols(data)
}
\arguments{
\item{data}{data table}
}
\value{
column names of columns that can be DWP values
}
\description{
Simple function to facilitate selection of columns that could
  be DWP values from a data table
}
