% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Ecbinom}
\alias{Ecbinom}
\title{Expected value of a continuous binomial}
\usage{
Ecbinom(prob)
}
\arguments{
\item{prob}{probability}
}
\value{
mean
}
\description{
Calculates the expected value of a continuous binomial

uses internal-only data
}
