% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{estgGenericCell}
\alias{estgGenericCell}
\title{Generic g estimation for a single cell of a combination of SE model and 
  CP model under a given search schedule}
\usage{
estgGenericCell(SS, param_SE, param_CP, dist, kFill)
}
\arguments{
\item{SS}{Search schedule (vector of days searched)}

\item{param_SE}{Searcher efficiency parameters (p and k}

\item{param_CP}{Carcass persistence parameters (a and b)}

\item{dist}{distribution for the CP model}

\item{kFill}{value to fill in for missing k when not existing in the model}
}
\description{
The g estimated by \code{gGenericCell} is a generic aggregate detection 
  probability and represents the probability of detecting a carcass that 
  arrives at a (uniform) random time during the period monitored, for one 
  cell across both the SE and CP models. This is somethat different from 
  the GenEst estimation of g when the purpose is to estimate total
  mortality (M), in which case the detection probability varies with 
  carcass arrival interval and is difficult to summarize statistically.
  The \code{gGenericCell} estimate is a useful "big picture" summary
  of detection probability, but would be difficult to work with for 
  estimating M with precision.
}
