% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{estimate_g}
\alias{estimate_g}
\title{Calculate detection probability (g) for a specific set of CP and SE 
 parameters and a specific search schedule.}
\usage{
estimate_g(days, CPab, distibution_choice, seef, k = NULL)
}
\arguments{
\item{days}{Specific search schedule.}

\item{CPab}{Matrix of parameters from CP theta
(rows: replicates, columns: a, b).}

\item{distibution_choice}{Distribution to use.}

\item{seef}{Matrix of parameters from SE theta 
(rows: replicates, columns: p, k).}

\item{k}{k.}
}
\value{
Simulated detection probabilities (length = replicates).
}
\description{
Calculate detection probability (g) for a specific set of CP and SE 
 parameters and a specific search schedule.
}
\examples{
NA
}
