% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{estimate_g_across_sizes}
\alias{estimate_g_across_sizes}
\title{Calculate detection probability (g) across all sizes.}
\usage{
estimate_g_across_sizes(cp_data, se_data, ss_data, replicates, cp_predictors,
  se_predictors, cp_theta, se_theta, cp_models, cp_models_to_use,
  se_models_to_use)
}
\arguments{
\item{cp_data}{Carcass persistence data.}

\item{se_data}{Seacher efficiency data.}

\item{ss_data}{Search schedule data.}

\item{replicates}{Number or iterations.}

\item{cp_predictors}{Carcass persistence predictor variables.}

\item{se_predictors}{Seacher efficiency predictor variables.}

\item{cp_theta}{Carcass persistence theta.}

\item{se_theta}{Seacher efficiency theta.}

\item{cp_models}{Carcass persistence model list.}

\item{cp_models_to_use}{Carcass persistence model to use.}

\item{se_models_to_use}{Seacher efficiency model to use.}
}
\value{
Array of g values 
        (replicates, search schedules, cell combinations, size classes).
}
\description{
Calculate detection probability (g) across all sizes.
}
\examples{
NA
}
