% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{estimate_mhat}
\alias{estimate_mhat}
\title{Calculate Mhat values across size classes, split categories, units,
 and cells.}
\usage{
estimate_mhat(co_data, ss_data, cp_data, se_data, size_class_column,
  split_column, unit_column, df_column, cp_predictors, se_predictors,
  replicates, garray)
}
\arguments{
\item{co_data}{Carcass observation data.}

\item{ss_data}{Search schedule data.}

\item{cp_data}{Carcass persistence data.}

\item{se_data}{Seacher efficiency data.}

\item{size_class_column}{Column header for the size class column.}

\item{split_column}{Column header for the split column}

\item{unit_column}{Column header for the unit column}

\item{df_column}{Column header for the date found column}

\item{cp_predictors}{Carcass persistence predictor variables.}

\item{se_predictors}{Seacher efficiency predictor variables.}

\item{replicates}{Number or iterations.}

\item{garray}{garray.}
}
\value{
Array of Mhat values 
 (iterations, units, split categories, size classes).
}
\description{
Calculate Mhat values across size classes, split categories, units,
 and cells.
}
\examples{
NA
}
