% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_functions.R
\name{estM}
\alias{estM}
\title{Estimate mortality}
\usage{
estM(nsim = 1, data_CO, data_SS, data_DWP, frac = 1, model_SE, model_CP,
  seed_SE = NULL, seed_CP = NULL, seed_g = NULL, seed_M = NULL,
  kFill = NULL, unitCol = "Unit", dateFoundCol = "DateFound",
  dateSearchedCol = "DateSearched", DWPCol = NULL, sizeclassCol = NULL)
}
\arguments{
\item{nsim}{the number of simulation draws}

\item{data_CO}{Carcass Observation data}

\item{data_SS}{Search Schedule data}

\item{data_DWP}{Survey unit (rows) by size (columns) density weighted 
proportion table}

\item{frac}{fraction of facility (by units or by area) surveyed}

\item{model_SE}{Searcher Efficiency model (or list of models if there are 
multiple size classes)}

\item{model_CP}{Carcass Persistence model (or list of models if there are 
multiple size classes)}

\item{seed_SE}{seed for random draws of the SE model}

\item{seed_CP}{seed for random draws of the CP model}

\item{seed_g}{seed for random draws of the gs}

\item{seed_M}{seed for the random draws of the Mhats}

\item{kFill}{value to fill in for missing k when not existing in the model}

\item{unitCol}{Column name for the unit indicator}

\item{dateFoundCol}{Column name for the date found data}

\item{dateSearchedCol}{Column name for the date searched data}

\item{DWPCol}{Column name for the DWP values in the DWP table when no
size class is used}

\item{sizeclassCol}{Name of colum in \code{data_CO} where the size classes
are recorded. Optional.}
}
\value{
list of SE parameters (pk), CP parameters (ab), g, Aj, and Mhat
}
\description{
Given given fitted Searcher Efficiency and Carcass 
  Persistence models; Search Schedule, Density Weighted Proportion,
  and Carcass Observation data; and information about the fraction of the
  the facility that was surveyed. This version of the function works for a
  single size class
}
\examples{
NA
}
