% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_utilities.R
\name{estText}
\alias{estText}
\title{Prepare the text for the estimation table}
\usage{
estText(rv, type = "SE")
}
\arguments{
\item{rv}{Reactive values list for the GenEst GUI.}

\item{type}{Model type, either "SE" or "CP".}
}
\value{
Rendered text ready for export to \code{output} list.
}
\description{
Depending on the model type and CL, prepare the header text.
}
