% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Etcbinom}
\alias{Etcbinom}
\title{Expected value of a truncated continuous binomial with size = 1/g and
 truncated below at g.}
\usage{
Etcbinom(prob)
}
\arguments{
\item{prob}{Vector of probabilities}
}
\value{
mean
}
\description{
Calculates the expected value of a truncated continuous
 binomial. Uses internal-only data
}
