% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{expandModelSetCP}
\alias{expandModelSetCP}
\title{Expand a CP model set for plotting}
\usage{
expandModelSetCP(modelSet)
}
\arguments{
\item{modelSet}{cp model set of class cpmSet}
}
\value{
updated model set
}
\description{
Expand the exponential models across the other distributions
}
