% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{ftpLink}
\alias{ftpLink}
\title{Creates a link to the FTP document of interest}
\usage{
ftpLink(doc = "UserGuide")
}
\arguments{
\item{doc}{"UserGuide" or "Models"}
}
\value{
link to the document
}
\description{
create the FTP link for the User Guide or Models document
}
