% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{fullMod}
\alias{fullMod}
\title{Return the model with the greatest log-likelihood}
\usage{
fullMod(modelSet)
}
\arguments{
\item{modSet}{a list of fitted models with a \code{loglik} element. Models
may be \code{pkm}, \code{cpm}, \code{survreg} objects or any objects with a
\code{loglik} component.}
}
\value{
The model object with the greatest log-likelihood among
 the models in \code{modelSet}
}
\description{
Compares all fitted models in a list and returns the model
 with the greatest log-likelihood
}
