% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_analysis_functions.R
\name{GeneralInputSidebar}
\alias{GeneralInputSidebar}
\title{Analysis General Inputs sidebar panel UI element}
\usage{
GeneralInputSidebar()
}
\value{
HTML for the Analysis General Inputs sidebar
}
\description{
create the HTML code for the Analysis General Inputs sidebar
}
