% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_general_inputs_functions.R
\name{GeneralInputsPanel}
\alias{GeneralInputsPanel}
\title{Analysis General Inputs Main Panel UI element}
\usage{
GeneralInputsPanel()
}
\value{
HTML for the Analysis General Inputs panel
}
\description{
create the HTML code for the Analysis General Inputs panel
}
