% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenEst.R
\docType{package}
\name{GenEst}
\alias{GenEst}
\alias{GenEst-package}
\title{Generalized estimation of fatalities}
\description{
This package is designed to analyze searcher efficiency, 
  carcass persistence, and carcass observation data for the estimation of
  fatalities.
}
\keyword{package}
