% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_about_functions.R
\name{GenEstAcknowledgements}
\alias{GenEstAcknowledgements}
\title{Help About acknowledgements}
\usage{
GenEstAcknowledgements()
}
\value{
HTML for the acknowledgements on the About panel
}
\description{
create the HTML code for the Help About acknowledgements
}
