% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_about_functions.R
\name{GenEstAuthors}
\alias{GenEstAuthors}
\title{Help About authors}
\usage{
GenEstAuthors()
}
\value{
HTML for the authors on the About panel
}
\description{
create the HTML code for the Help About authors
}
