% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_about_functions.R
\name{GenEstGUIauthors}
\alias{GenEstGUIauthors}
\title{Help About GUI authors}
\usage{
GenEstGUIauthors()
}
\value{
HTML for the GUI authors on the About panel
}
\description{
create the HTML code for the Help About GUI authors
}
