% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{GenEstInlineCSS}
\alias{GenEstInlineCSS}
\title{Inline CSS definition}
\usage{
GenEstInlineCSS(...)
}
\arguments{
\item{...}{to be passed down}
}
\description{
Define the inline CSS code for GenEst.
}
