% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_about_functions.R
\name{GenEstLicense}
\alias{GenEstLicense}
\title{Help About license}
\usage{
GenEstLicense()
}
\value{
HTML for the license on the About panel
}
\description{
create the HTML code for the Help About license
}
