% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_about_functions.R
\name{GenEstLogos}
\alias{GenEstLogos}
\title{Help About logos}
\usage{
GenEstLogos()
}
\value{
HTML for the logos on the About panel
}
\description{
create the HTML code for the Help About logos
}
