% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{GenEstShinyJS}
\alias{GenEstShinyJS}
\title{Shiny Java Script (via shinyjs) enabler}
\usage{
GenEstShinyJS(...)
}
\arguments{
\item{...}{to be passed down}
}
\description{
Enable the use of the \code{shinyjs} package within GenEst.
  Simply a wrapper to bring function into namespace.
}
