% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui.R
\name{GenEstUI}
\alias{GenEstUI}
\alias{dataInputPanel}
\alias{dataInputSidebar}
\alias{loadedDataPanel}
\alias{analysisPanel}
\alias{GeneralInputsPanel}
\alias{GeneralInputSidebar}
\alias{SEPanel}
\alias{SESidebar}
\alias{SEMainPanel}
\alias{CPPanel}
\alias{CPSidebar}
\alias{CPMainPanel}
\alias{MPanel}
\alias{MSidebar}
\alias{MMainPanel}
\alias{gPanel}
\alias{gSidebar}
\alias{gMainPanel}
\alias{helpPanel}
\alias{gettingStartedPanel}
\alias{downloadsPanel}
\alias{aboutPanel}
\alias{disclaimersPanel}
\title{Create the GenEst User Interface HTML}
\usage{
GenEstUI(appType = "base")

dataInputPanel()

dataInputSidebar()

loadedDataPanel()

analysisPanel()

GeneralInputsPanel()

GeneralInputSidebar()

SEPanel()

SESidebar()

SEMainPanel()

CPPanel()

CPSidebar()

CPMainPanel()

MPanel()

MSidebar()

MMainPanel()

gPanel()

gSidebar()

gMainPanel()

helpPanel(appType = "base")

gettingStartedPanel()

downloadsPanel()

aboutPanel()

disclaimersPanel(appType = "base")
}
\arguments{
\item{appType}{Toggle control for the app, \code{"base"} for local versions
or \code{"deploy"} for hosted version. Currently only differentiates the
disclaimer text.}
}
\value{
Each function returns a string of HTML code: \cr \cr 
  \code{GenEstUI}: Full GenEst user interface.

\code{dataInputPanel}: Data Input panel.

\code{dataInputSidebar}: Data Input sidebar.

\code{loadedDataPanel}: Data Input data panel.

\code{AnalysisPanel}: Analysis panel.

\code{GeneralInputsPanel}: Analysis -> General Inputs panel.

\code{GeneralInputSidebar}: Analysis -> General Inputs sidebar.

\code{SEPanel}: Analysis -> Searcher Efficiency panel.

\code{SESidebar}: Analysis -> Searcher Efficiency sidebar.

\code{SEMainPanel}: Analysis -> Searcher Efficiency main panel.

\code{CPPanel}: Analysis -> Carcass Persistence panel.

\code{CPSidebar}: Analysis -> Carcass Persistence sidebar.

\code{CPMainPanel}: Analysis -> Carcass Persistence main panel.

\code{MPanel}: Analysis -> Mortality Estimation panel.

\code{MSidebar}: Analysis -> Mortality Estimation 
  sidebar.

\code{MMainPanel}: Analysis -> Mortality Estimation main panel.

\code{gPanel}: Analysis -> Detection Probability panel.

\code{gSidebar}: Analysis -> Detection Probability sidebar.

\code{gMainPanel}: Analysis -> Detection Probability main panel.

\code{helpPanel}: Help panel.

\code{gettingStartedPanel}: Help -> Getting Started panel.

\code{downloadsPanel}: Help -> Downloads panel.

\code{aboutPanel}: Help -> About panel.

\code{aboutPanel}: Help -> Disclaimers panel.
}
\description{
This suite of functions create the HTML code underlying the
  GenEst user interface (UI). \cr \cr 
  \code{GenEstUI} creates the HTML code for the whole page by calling 
  \code{dataInputPanel}, \code{analysisPanel}, and \code{helpPanel}.

\code{dataInputPanel} creates the HTML code for the Data Input
  panel in the GenEst UI by calling \code{dataInputSidebar} and 
  \code{loadedDataPanel}.

\code{dataInputSidebar} creates the HTML code for the Data
  Input panel's sidebar (where the data files are selected) by calling 
  \code{\link{dataInputWidget}} for each of the data file types.

\code{loadedDataPanel} creates the HTML code for the Data
  Input panel's main page (where the data files are displayed) by calling 
  \code{\link{dataTabPanel}} for each of the data file types.

\code{analysisPanel} creates the HTML code for the Analysis
  panel in the GenEst UI by calling \code{GeneralInputsPanel},
  \code{SEPanel}, \code{CPPanel}, \code{MPanel}, and \code{gPanel}.

\code{GeneralInputsPanel} creates the HTML code for the 
  Analysis panel's General Inputs panel by calling 
  \code{GeneralInputSidebar}.

\code{GeneralInputSidebar} creates the HTML code for the 
  Analysis panel's General Inputs sidebar (where the Number of Iterations,
  Confidence Level, and Size Class Column are selected) by calling 
  \code{\link{modelInputWidget}} for each of the inputs.

\code{SEPanel} creates the HTML code for the Analysis
  panel's Searcher Efficiency panel by calling 
  \code{SESidebar} and \code{SEMainPanel}.

\code{SESidebar} creates the HTML code for the Analysis 
  panel's Searcher Efficiency panel's sidebar (where the Observation 
  Columns, Predictor Columns, and fixed-k values are input and where the 
  Size Class, p formula, and k formula are selected for the outputs) by
  calling \code{\link{modelInputWidget}} for each of the inputs,
  \code{\link{modelRunWidget}} for running the model, and then 
  \code{\link{modelOutputWidget}} for the output controls.

\code{SEMainPanel} creates the HTML code for the Analysis 
  panel's Searcher Efficiency panel's main panel (where the Selected Data,
  Model Estimates, Figures, Model Comparison, and Model Selection are 
  displayed) by calling \code{selectedDataPanel} for the selected
  data and then \code{modelOutputPanel} for each of the outputs.

\code{CPPanel} creates the HTML code for the Analysis
  panel's Carcass Persistence panel by calling 
  \code{CPSidebar} and \code{CPMainPanel}.

\code{CPSidebar} creates the HTML code for the Analysis 
  panel's Carcass Persistence panel's sidebar (where the Observation 
  Columns, Predictor Columns, and Distributions are input and where the 
  Size Class, location formula, and scale formula are selected for the 
  outputs) by calling \code{\link{modelInputWidget}} for each of the 
  inputs, \code{\link{modelRunWidget}} for running the model, and then 
  \code{\link{modelOutputWidget}} for the output controls.

\code{CPMainPanel} creates the HTML code for the Analysis 
  panel's Carcass Persistence panel's main panel (where the Selected Data,
  Model Estimates, Figures, Model Comparison, and Model Selection are 
  displayed) by calling \code{selectedDataPanel} for the selected
  data and then \code{modelOutputPanel} for each of the outputs.

\code{MPanel} creates the HTML code for the Analysis
  panel's Mortality Estimation panel by calling 
  \code{MSidebar} and \code{MMainPanel}.

\code{MSidebar} creates the HTML code for the Analysis 
  panel's Mortality Estimation panel's sidebar (where the Search Schedule
  and the assummed k (if needed) are input and where the Size Class, 
  is selected for the outputs) by calling \code{\link{modelInputWidget}} 
  for each of the inputs, \code{\link{modelRunWidget}} for running the 
  model, and then \code{\link{modelOutputWidget}} for the output controls.

\code{MMainPanel} creates the HTML code for the Analysis 
  panel's Mortality Estimation panel's main panel (where the Figures, and
  Summary are displayed) by calling \code{modelOutputPanel} for each of 
  the outputs.

\code{gPanel} creates the HTML code for the Analysis
  panel's Detection Probability panel by calling 
  \code{gSidebar} and \code{gMainPanel}.

\code{gSidebar} creates the HTML code for the Analysis 
  panel's Detection Probability panel's sidebar (where the Search Schedule
  and the assummed k (if needed) are input and where the Size Class, 
  is selected for the outputs) by calling \code{\link{modelInputWidget}} 
  for each of the inputs, \code{\link{modelRunWidget}} for running the 
  model, and then \code{\link{modelOutputWidget}} for the output controls.

\code{gMainPanel} creates the HTML code for the Analysis 
  panel's Detection Probability panel's main panel (where the Search 
  Schedule, Figures, and Summary are displayed) by calling 
  \code{modelOutputPanel} for each of the outputs.

\code{helpPanel} creates the HTML code for the Help panel in
  the GenEst UI by calling \code{gettingStartedPanel},
  \code{downloadsPanel}, \code{aboutPanel}, and \code{disclaimersPanel}.

\code{gettingStartedPanel} creates the HTML code for the Help
  panel's Getting Started panel, which calls 
  \code{\link{gettingStartedContent}}, the function containing the raw
  content for the page (which is text heavy and so moved to its own 
  function).

\code{downloadsPanel} creates the HTML code for the Help
  panel's Downloads panel, which calls 
  \code{\link{dataDownloadWidget}} for each of the data sets.

\code{aboutPanel} creates the HTML code for the Help
  panel's About panel, which calls 
  \code{\link{aboutContent}}, the function containing the raw
  content for the page (which is text heavy and so moved to its own 
  function).

\code{disclaimersPanel} creates the HTML code for the Help
  panel's Disclaimers panel, which calls 
  \code{\link{disclaimersContent}}, the function containing the raw
  content for the page (which is text heavy and so moved to its own 
  function).
}
\details{
Currently there are few differences between the local and deployed
  versions of GenEst, and the \code{appType} toggle is only included as an 
  argument for functions that can produce different versions of the HTML.
  At this point, the only content that is different is the disclaimer text
  on the Help panel.
}
