% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_content.R
\name{gettingStartedContent}
\alias{gettingStartedContent}
\title{Create the Content for the Getting Started Pabel}
\usage{
gettingStartedContent()
}
\value{
HTML for the Help Getting Started main panel.
}
\description{
Create the HTML code for the Help Getting Started main panel,
  which describes basic usage of the app and directs users to additional
  documentation.
}
