% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_getting_started_functions.R
\name{gettingStartedMainPanel}
\alias{gettingStartedMainPanel}
\title{Help Getting Started main panel UI element}
\usage{
gettingStartedMainPanel()
}
\value{
HTML for the Help Getting Started main panel
}
\description{
create the HTML code for the Help Getting Started main panel
}
