% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_getting_started_functions.R
\name{gettingStartedPanel}
\alias{gettingStartedPanel}
\title{Help Getting Started Main Panel UI element}
\usage{
gettingStartedPanel()
}
\value{
HTML for the Help Getting Started panel
}
\description{
create the HTML code for the Help Getting Started panel
}
