% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_g_functions.R
\name{gFigurePanel}
\alias{gFigurePanel}
\title{Analysis Detection Probability main panel figure UI element}
\usage{
gFigurePanel()
}
\value{
HTML for the Analysis Detection Probability figure panel
}
\description{
create the HTML code for the Analysis Detection Probability
  main panel figure element
}
