% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_g_functions.R
\name{gMainPanel}
\alias{gMainPanel}
\title{Analysis Detection Probability main panel UI element}
\usage{
gMainPanel()
}
\value{
HTML for the Analysis Detection Probability main panel
}
\description{
create the HTML code for the Analysis Detection Probability 
  main panel
}
