% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_g_functions.R
\name{gPanel}
\alias{gPanel}
\title{Analysis Detection Probability Main Panel UI element}
\usage{
gPanel()
}
\value{
HTML for the Analysis Detection Probability panel
}
\description{
create the HTML code for the Analysis Detection Probability 
  panel
}
