% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_g_functions.R
\name{gSchedulePanel}
\alias{gSchedulePanel}
\title{Analysis Detection Probability main panel schedule UI element}
\usage{
gSchedulePanel()
}
\value{
HTML for the Analysis Detection Probability schedule panel
}
\description{
create the HTML code for the Analysis Detection Probability
  main panel schedule element
}
