% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_g_functions.R
\name{gSidebar}
\alias{gSidebar}
\title{Analysis Detection Probability sidebar panel UI element}
\usage{
gSidebar()
}
\value{
HTML for the Analysis Detection Probability sidebar
}
\description{
create the HTML code for the Analysis Detection Probability 
  sidebar
}
