% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_g_functions.R
\name{gSummaryPanel}
\alias{gSummaryPanel}
\title{Analysis Detection Probability main panel summary UI element}
\usage{
gSummaryPanel()
}
\value{
HTML for the Analysis Detection Probability summary panel
}
\description{
create the HTML code for the Analysis Detection Probability
  main panel summary element
}
