% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_help_functions.R
\name{helpPanel}
\alias{helpPanel}
\title{Help Main Panel UI element}
\usage{
helpPanel(type = "base")
}
\arguments{
\item{type}{"base" (for local versions) or "deploy" (for hosted version)}
}
\value{
HTML for the Help panel
}
\description{
create the HTML code for the Help panel
}
