% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{initialOutput}
\alias{initialOutput}
\title{Update the output list upon initiation of the app}
\usage{
initialOutput(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}
}
\value{
an updated output list
}
\description{
Update the output list when the app is started
}
