% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{initialReactiveValues}
\alias{initialReactiveValues}
\title{Create the main reactive value list for GenEst}
\usage{
initialReactiveValues()
}
\value{
a reactive values list
}
\description{
Create a list of reactive values as used across the components
  of the GenEst application
}
