% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{isNeverDecreasing}
\alias{isNeverDecreasing}
\title{Check if a vector is never decreasing}
\usage{
isNeverDecreasing(x, tiesOK = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{vector of values}

\item{tiesOK}{logical if ties are ok or not}

\item{na.rm}{logical if NAs are to be removed or not}
}
\value{
logical value
}
\description{
Check if a vector is never decreasing
}
