% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{isNeverIncreasing}
\alias{isNeverIncreasing}
\title{Check if a vector is never increasing}
\usage{
isNeverIncreasing(x, tiesOK = TRUE)
}
\arguments{
\item{x}{vector of values}

\item{tiesOK}{logical if ties are ok or not}
}
\value{
logical value
}
\description{
Check if a vector is never increasing
}
