% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{kFillPropose}
\alias{kFillPropose}
\title{Propose a k value if it is not in the model table}
\usage{
kFillPropose(model)
}
\arguments{
\item{model}{searcher efficiency model}
}
\value{
proposed k value
}
\description{
Based on the call to the pkm function
}
\examples{
NA
}
