% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{kSuggest}
\alias{kSuggest}
\title{Suggest a value for k based on a weighted mean of the observed decay}
\usage{
kSuggest(obsData)
}
\arguments{
\item{obsData}{a matrix of carcasses (rows) x searched (columns)}
}
\value{
the weighted mean of the observed proportional decay in p
}
\description{
Suggest a value for k based on a weighted mean of the observed decay
}
