% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{li}
\alias{li}
\title{HTML li function}
\usage{
li(...)
}
\arguments{
\item{...}{attributes and children of the element}
}
\value{
HTML ordered list
}
\description{
Generate an HTML list element. This simply pulls the
  function definition from the tag environment in htmltools
}
