% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_data_input_functions.R
\name{loadedDataPanel}
\alias{loadedDataPanel}
\title{Data Input loaded data UI element}
\usage{
loadedDataPanel()
}
\value{
HTML for the Data Input data table ouput
}
\description{
create the HTML code for the Data Input loaded data tables
}
