% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R, R/utility_functions_old.R
\name{logit}
\alias{logit}
\alias{logit}
\title{Compute the logit.}
\usage{
logit(x)

logit(x)
}
\arguments{
\item{x}{A probability (between 0 and 1, inclusive).}

\item{x}{A probability (between 0 and 1, inclusive).}
}
\value{
The logit of \code{x}.

The logit of \code{x}.
}
\description{
Compute the logit.

Compute the logit.
}
\examples{
logit(0.5)
logit(0.5)
}
