% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits_functions.R
\name{ltranspose}
\alias{ltranspose}
\title{Transpose a list of arrays}
\usage{
ltranspose(M)
}
\arguments{
\item{M}{a list of \code{n} \code{m} x \code{k} arrays}
}
\value{
a list of \code{m} \code{n} x \code{k} arrays
}
\description{
A list of \code{n} arrays, each with dimension \code{m} x 
  \code{k} is redimensioned to a list of \code{m} arrays, each with 
  dimension \code{m} x \code{k}. NOTE: Attributes are not preserved.
}
