% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{makeEstText}
\alias{makeEstText}
\title{Make the text for an estimate based on CL}
\usage{
makeEstText(CL)
}
\arguments{
\item{CL}{confidence Limit}
}
\value{
character for the estimate range
}
\description{
Make the character of text for an estimate based on CL
}
