% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{makekFixedInput}
\alias{makekFixedInput}
\title{Make a kFixed selection menu}
\usage{
makekFixedInput(rv)
}
\arguments{
\item{rv}{reactive values list}
}
\value{
rendered HTML k selection menu object
}
\description{
Produce a size-based kFixed selection menu object based on
  column selection
}
