% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{makekInput}
\alias{makekInput}
\title{Make a k selection menu}
\usage{
makekInput(sizeclasses)
}
\arguments{
\item{sizeclasses}{size class names}
}
\value{
rendered HTML k selection menu object
}
\description{
Produce a size-based k-selection menu object based on
  column selection
}
