% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{makeMenu}
\alias{makeMenu}
\title{Make a model menu}
\usage{
makeMenu(mods, sizeclasses, type)
}
\arguments{
\item{mods}{size-indexed list of model sets}

\item{sizeclasses}{size class names}

\item{type}{model type, either "SE" or "CP"}
}
\value{
rendered HTML model selection menu object
}
\description{
Produce a size-based model-selection menu object based on
  model inputs
}
