% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_M_functions.R
\name{MFigurePanel}
\alias{MFigurePanel}
\title{Analysis Mortality main panel figure UI element}
\usage{
MFigurePanel()
}
\value{
HTML for the Analysis Mortality figure panel
}
\description{
create the HTML code for the Analysis Mortality
  main panel figure element
}
