% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_M_functions.R
\name{MMainPanel}
\alias{MMainPanel}
\title{Analysis Mortality main panel UI element}
\usage{
MMainPanel()
}
\value{
HTML for the Analysis Mortality main panel
}
\description{
create the HTML code for the Analysis Mortality main panel
}
