% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mock}
\alias{mock}
\title{A mock example data set}
\format{A list with 5 items:
\describe{
  \item{SE}{Searcher efficiency trial data}
  \item{CP}{Carcass persistence trial data}
  \item{SS}{Search schedule data}
  \item{DWP}{Density weighted proportion of area searched data}
  \item{CO}{Carcass observations}   
}}
\source{
\code{mock}
}
\usage{
mock
}
\description{
A template dataset used for testing purposes.  Dataset containing SE, CP, SS,
 DWP, and CO data. Data are mostly random without patterns.
}
\keyword{datasets}
