% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_welcome_functions.R
\name{modalTextUSGS}
\alias{modalTextUSGS}
\title{Create USGS text for the modal welcome}
\usage{
modalTextUSGS()
}
\value{
HTML element of the USGS text
}
\description{
Create USGS text for the modal welcome
}
