% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_welcome_functions.R
\name{modalTextWEST}
\alias{modalTextWEST}
\title{Create WEST text for the modal welcome}
\usage{
modalTextWEST()
}
\value{
HTML element of the WEST text
}
\description{
Create WEST text for the modal welcome
}
