% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_welcome_functions.R
\name{modalWelcome}
\alias{modalWelcome}
\title{Create the modal welcome for GenEst}
\usage{
modalWelcome()
}
\value{
Nothing
}
\description{
Create a modal welcome and basic info screen for the GenEst
  application
}
