% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions_old.R
\name{model_name_reverser}
\alias{model_name_reverser}
\title{Create an equation from a reader-friendly model name.}
\usage{
model_name_reverser(model_name)
}
\arguments{
\item{model_name}{Model in reader-friendly format.}
}
\value{
Model in equation format.
}
\description{
Create an equation from a reader-friendly model name.
}
\examples{
NA
}
