% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{model_namer}
\alias{model_namer}
\title{Create a reader-friendly model name from an equation.}
\usage{
model_namer(model_equation)
}
\arguments{
\item{model_equation}{Model in equation format.}
}
\value{
Model in reader-friendly format.
}
\description{
Create a reader-friendly model name from an equation.
}
\examples{
NA
}
