% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{model_utility_functions}
\alias{model_utility_functions}
\alias{matchCells}
\alias{modelSetModelPredictors}
\alias{modelSetPredictors}
\alias{modelSetModelCells}
\alias{modelSetCells}
\title{model utility functions (not exported)}
\usage{
matchCells(specific, modelSet)

modelSetModelPredictors(modelSet)

modelSetPredictors(modelSet)

modelSetModelCells(modelSet)

modelSetCells(modelSet)
}
\arguments{
\item{specific}{specific model compared against the full set}

\item{modelSet}{full model set to compare to the specific}
}
\description{
model utility functions that are not exported
}
