% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{modelMenuHeader}
\alias{modelMenuHeader}
\title{Make a model menu header}
\usage{
modelMenuHeader(n)
}
\arguments{
\item{n}{number of size classed}
}
\value{
rendered HTML model selection menu object
}
\description{
Produce a size-based model-selection menu header object based 
  on model inputs
}
