% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_panels.R
\name{modelOutputPanel}
\alias{modelOutputPanel}
\title{Create a Model Output Tab Panel for the GenEst User Interface HTML}
\usage{
modelOutputPanel(outType)
}
\arguments{
\item{outType}{Toggle control for the model type of the panel. One of 
"SEFigures", "SEEstimates", "SEModComparison", "SEModSelection",
"CPFigures", "CPEstimates", "CPModComparison", "CPModSelection",
"gFigures", or "gSummary".}
}
\value{
HTML for the panel
}
\description{
This is a generalized function for creating a model output 
  panel used in the GenEst GUI, based on the output type 
  (\code{outType}).
}
