% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_widgets.R
\name{modelOutputWidget}
\alias{modelOutputWidget}
\title{Create a Model Output Widget for the GenEst User Interface HTML}
\usage{
modelOutputWidget(modType)
}
\arguments{
\item{modType}{Toggle control for the model type of the widget. One of 
"SE", "CP", "M", or "g".}
}
\value{
HTML for the model run widget.
}
\description{
This is a generalized function for creating a widget used in 
  the GenEst GUI to control the outputs based on the model type
  (\code{modType}).
}
