% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_widgets.R
\name{modelRunWidget}
\alias{modelRunWidget}
\title{Create a Model Run Widget for the GenEst User Interface HTML}
\usage{
modelRunWidget(modType)
}
\arguments{
\item{modType}{Toggle control for the model type of the widget. One of 
"SE", "CP", "M", or "g".}
}
\value{
HTML for the model run widget.
}
\description{
This is a generalized function for creating a model run 
  widget used in the GenEst GUI, based on the model type (\code{modType}).
  The widget includes the model run button and the model clear button 
  (once the model has finished running).
}
