% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{modelSetCells}
\alias{modelSetCells}
\title{Determine the cell table for a full model set}
\usage{
modelSetCells(modelSet)
}
\arguments{
\item{modelSet}{model set}
}
\value{
cell table for the model set
}
\description{
Determine the cell table for a full model set
}
