% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{modelSetModelCells}
\alias{modelSetModelCells}
\title{Determine the cells from each model in a model set}
\usage{
modelSetModelCells(modelSet)
}
\arguments{
\item{modelSet}{model set}
}
\value{
List of the cells from each model in a model set
}
\description{
Determine the cells from each model in a model set
}
