% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{modelSetModelPredictors}
\alias{modelSetModelPredictors}
\title{Determine the predictors from each model in a model set}
\usage{
modelSetModelPredictors(modelSet)
}
\arguments{
\item{modelSet}{model set}
}
\value{
List of the predictors from each model in a model set
}
\description{
Determine the predictors from each model in a model set
}
