% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{modelSetPredictors}
\alias{modelSetPredictors}
\title{Determine the predictors for a whole model set}
\usage{
modelSetPredictors(modelSet)
}
\arguments{
\item{modelSet}{model set}
}
\value{
vector of the predictors from a model set
}
\description{
Determine the predictors for a whole model set
}
