% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{modNamePaste}
\alias{modNamePaste}
\title{Paste the parts of a model's name back together}
\usage{
modNamePaste(parts, type = "SE", tab = FALSE)
}
\arguments{
\item{parts}{the component parts of the model's name}

\item{type}{"SE" or "CP"}

\item{tab}{logical for if it's the table output for CP}
}
\value{
the pasted name
}
\description{
Paste the component parts of a model's name back together
  for presentation
}
