% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{modNameSplit}
\alias{modNameSplit}
\title{Split model names into their components and remove only a desired 
  one}
\usage{
modNameSplit(modNames, pos)
}
\arguments{
\item{modNames}{names of the model to be split off}

\item{pos}{position in the name to split off}
}
\value{
vector of split-off model names
}
\description{
Split a model name to return a specific component. Splitting 
  is done based on the semicolon in the name
}
