% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{modTabColNamesSE}
\alias{modTabColNamesSE}
\title{Creates the names for the Searcher Efficiency model table columns}
\usage{
modTabColNamesSE(CL = 0.9)
}
\arguments{
\item{CL}{Confidence level}
}
\value{
column names
}
\description{
Based on confidence level of interest
}
