% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_M_functions.R
\name{MPanel}
\alias{MPanel}
\title{Analysis Mortality Main Panel UI element}
\usage{
MPanel()
}
\value{
HTML for the Analysis Mortality Estimation panel
}
\description{
create the HTML code for the Analysis Mortality panel
}
