% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgFracNote}
\alias{msgFracNote}
\title{Create the message for when an incorrect FFS is input}
\usage{
msgFracNote(fracNote)
}
\arguments{
\item{fracNote}{the note regarding the input}
}
\value{
a message regarding the input issue
}
\description{
Produces a notification for improper input for fraction 
  facility surveyed
}
