% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgList}
\alias{msgList}
\title{Produce the message list}
\usage{
msgList()
}
\value{
list of message elements
}
\description{
Produce the message list, currently restricted to a NULL value
  starting for each value
}
