% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgModDone}
\alias{msgModDone}
\title{Create a model done message}
\usage{
msgModDone(msgs, modelType, rv, clear = TRUE)
}
\arguments{
\item{msgs}{message list}

\item{modelType}{"SE", "CP", "g", or "M"}

\item{rv}{reactive values list}

\item{clear}{if all notifications should be cleared or not}
}
\value{
a model running message
}
\description{
Produces a model-done notification
}
