% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{msgModDoneM}
\alias{msgModDoneM}
\title{Creates the M finished message}
\usage{
msgModDoneM(ghatAj)
}
\arguments{
\item{ghatAj}{output from rghat}
}
\value{
M done message
}
\description{
Creates the M finished message
}
