% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_message_functions.R
\name{msgModPartialFail}
\alias{msgModPartialFail}
\title{Create the warning message for when only some models are fit 
  successfully}
\usage{
msgModPartialFail()
}
\value{
a partial model fail warning
}
\description{
Produces a notification for partial model failure
}
