% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_message_functions.R
\name{msgModRun}
\alias{msgModRun}
\title{Create a model running message}
\usage{
msgModRun(modelType)
}
\arguments{
\item{modelType}{"SE", "CP", "g", or "M"}
}
\value{
a model running message
}
\description{
Produces a model-running notification
}
