% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgModSENobs}
\alias{msgModSENobs}
\title{Create the SE data size notification}
\usage{
msgModSENobs(rv)
}
\arguments{
\item{rv}{reactive values list}
}
\value{
data size message
}
\description{
Produces a notification for SE data sizes (associated with k)
}
