% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgModWarning}
\alias{msgModWarning}
\title{Create the warning message for a model run (if needed)}
\usage{
msgModWarning(mods, type = "SE", rv = NULL)
}
\arguments{
\item{mods}{Set Size list of models}

\item{type}{"SE" or "CP"}

\item{rv}{reactive values list}
}
\value{
a partial model warning (if needed)
}
\description{
Produces a notification for partial model failures if needed
}
