% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{msgNobsSE}
\alias{msgNobsSE}
\title{Creates the SE data size notification}
\usage{
msgNobsSE(formula_k, kFixed, obsCols)
}
\arguments{
\item{formula_k}{formula for k}

\item{kFixed}{fixed value for k if used}

\item{obsCols}{vector of observation column names}
}
\value{
data size message
}
\description{
Creates the SE data size notification
}
