% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{msgSampleSize}
\alias{msgSampleSize}
\title{Creates the small sample size warning}
\usage{
msgSampleSize(mods)
}
\arguments{
\item{mods}{models to check for sample sizes}
}
\value{
a small sample size message
}
\description{
Creates the small sample size warning
}
