% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{msgsplitFail}
\alias{msgsplitFail}
\title{Creates the fail message for when splits aren't done correctly}
\usage{
msgsplitFail(type = NULL)
}
\arguments{
\item{type}{"setup" or "run"}
}
\value{
a split fail warning
}
\description{
Creates the fail message for when splits aren't done correctly
}
