% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_message_functions.R
\name{msgSSavgFail}
\alias{msgSSavgFail}
\title{Create the warning message for when the SS average doesn't work}
\usage{
msgSSavgFail()
}
\value{
an average SS fail warning
}
\description{
Produces a notification for when an average search schedule
  can't be created
}
