% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgSSinputFail}
\alias{msgSSinputFail}
\title{Create the warning message for when the SS based on inputs doesn't 
  work}
\usage{
msgSSinputFail(msgs, rv, clear = TRUE)
}
\arguments{
\item{msgs}{message list}

\item{rv}{reactive values list}

\item{clear}{if all notifications should be cleared or not}
}
\value{
an average SS fail warning
}
\description{
Produces a notification for when an input-based search 
  schedule can't be created
}
