% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_M_functions.R
\name{MSidebar}
\alias{MSidebar}
\title{Analysis Mortality sidebar panel UI element}
\usage{
MSidebar()
}
\value{
HTML for the Analysis Mortality sidebar
}
\description{
create the HTML code for the Analysis Mortality sidebar
}
