% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_M_functions.R
\name{MSummaryPanel}
\alias{MSummaryPanel}
\title{Analysis Mortality main panel summary UI element}
\usage{
MSummaryPanel()
}
\value{
HTML for the Analysis Mortality summary panel
}
\description{
create the HTML code for the Analysis Mortality
  main panel summary element
}
