% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{mucbinom}
\alias{mucbinom}
\title{Calculates the mean of a continuous binomial}
\usage{
mucbinom(prob)
}
\arguments{
\item{prob}{probability}
}
\value{
mean
}
\description{
uses internal-only data
}
\examples{
NA
}
