% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_utilities.R
\name{navbar}
\alias{navbar}
\title{Create the HTML for the navigation bar UI element}
\usage{
navbar()
}
\value{
HTML for the navbar element
}
\description{
create the HTML code for the navigation bar in the GenEst app.
}
