% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{obsCols_fta}
\alias{obsCols_fta}
\title{Select the columns from a data table that could be CP First Time
  Absent observations}
\usage{
obsCols_fta(data)
}
\arguments{
\item{data}{data table}
}
\value{
column names of columns that can be observations
}
\description{
Simple function to facilitate selection of columns that could
  be First Time Absent observations for a CP model
}
