% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{obsCols_ltp}
\alias{obsCols_ltp}
\title{Select the columns from a data table that could be CP Last Time
  Present observations}
\usage{
obsCols_ltp(data)
}
\arguments{
\item{data}{data table}
}
\value{
column names of columns that can be observations
}
\description{
Simple function to facilitate selection of columns that could
  be Last Time Present observations for a CP model
}
