% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{obsCols_SE}
\alias{obsCols_SE}
\title{Select the columns from a data table that could be SE observations}
\usage{
obsCols_SE(data)
}
\arguments{
\item{data}{data table}
}
\value{
column names of columns that can be observations
}
\description{
Simple function to facilitate selection of columns that could
  be observations for an SE model
}
