% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{ol}
\alias{ol}
\title{HTML ol function}
\usage{
ol(...)
}
\arguments{
\item{...}{attributes and children of the element}
}
\value{
HTML ordered list
}
\description{
Generate an ordered list HTML object. This simply pulls the
  function definition from the tag environment in htmltools
}
