% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{order_cp_models}
\alias{order_cp_models}
\title{Create the model selection order list for carcass persistence models.}
\usage{
order_cp_models(models)
}
\arguments{
\item{models}{Carcass persistence models.}
}
\value{
Model order list.
}
\description{
Create the model selection order list for carcass persistence models.
}
\examples{
NA
}
