% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{order_se_models}
\alias{order_se_models}
\title{Create the model selection order list for searcher efficiency models.}
\usage{
order_se_models(models)
}
\arguments{
\item{models}{Searcher efficiency models.}
}
\value{
Model order list.
}
\description{
Create the model selection order list for searcher efficiency models.
}
\examples{
NA
}
