% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkfunction}
\alias{pkfunction}
\title{Function optimized to fit the searcher efficiency models.}
\usage{
pkfunction(searches_missed, search_found, theta, n_theta_p, groups, maxmiss,
  facts, fix_k, fix_k_value)
}
\arguments{
\item{searches_missed}{Number of searches when carcass was present but
not found.}

\item{search_found}{Search on which carcass was found.}

\item{theta}{Parameters to be optimized.}

\item{n_theta_p}{Number of parameters associated with theta}

\item{groups}{Which cell each observation belongs to.}

\item{maxmiss}{Maximum possible number of misses for a carcass.}

\item{facts}{Combined pk model matrix.}

\item{fix_k}{Logical of whether k should be fixed.}

\item{fix_k_value}{Value of k if fixed.}
}
\value{
Negative log likelihood of the observations, given the parameters.
}
\description{
Function optimized to fit the searcher efficiency models.
}
\examples{
NA
}
