% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkLogLik}
\alias{pkLogLik}
\title{Calculate the negative log-likelihood of a searcher efficiency model.}
\usage{
pkLogLik(misses, found_on, beta, n_beta_p, carcass_cell, max_misses,
  mm_pk_cells, fixed_k = NULL)
}
\arguments{
\item{misses}{Number of searches when carcass was present but
not found.}

\item{found_on}{Search on which carcass was found.}

\item{beta}{Parameters to be optimized.}

\item{n_beta_p}{Number of parameters associated with p}

\item{carcass_cell}{Which cell each observation belongs to.}

\item{max_misses}{Maximum possible number of misses for a carcass.}

\item{mm_pk_cells}{Combined pk model matrix.}

\item{fixed_k_value}{Value of k if fixed.}
}
\value{
Negative log likelihood of the observations, given the parameters.
}
\description{
Calculate the negative log-likelihood of a searcher efficiency model.
}
\examples{
NA
}
