% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkm_set_aicc_tab}
\alias{pkm_set_aicc_tab}
\title{Create the  AICc tables for the searcher efficiency models}
\usage{
pkm_set_aicc_tab(pk_model_set)
}
\arguments{
\item{pk_model_set}{Set of searcher efficiency models fit to the same
observations}
}
\value{
AICc table
}
\description{
Create the  AICc tables for the searcher efficiency models
}
\examples{
NA
}
