% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmAllFail}
\alias{pkmAllFail}
\title{Check if all of the pkm models fail}
\usage{
pkmAllFail(pkmToCheck)
}
\arguments{
\item{pkmToCheck}{A \code{pkm} model or a set of them or a suite of sets
associated with multiple sizes}
}
\value{
A single logical value indicating if all of the models failed
}
\description{
Check if all of the pkm models fail
}
