% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmCheck}
\alias{pkmCheck}
\title{Verify that a suite of searcher efficiency models all fit successfully.}
\usage{
pkmCheck(pkmToCheck)
}
\arguments{
\item{pkmToCheck}{A \code{pkm} model or a set of them or a suite of sets
associated with multiple sizes}
}
\value{
A single (total) logcal
}
\description{
Verify that a suite of searcher efficiency models all fit successfully.
}
