% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmFail}
\alias{pkmFail}
\title{Check if a pk model is well-fit}
\usage{
pkmFail(pkmod)
}
\arguments{
\item{pkmod}{A \code{\link{pkm}} object to test}
}
\value{
logical value indicating a failed fit (TRUE) or successful (FALSE)
}
\description{
Run a check the arg is a well-fit pkm object
}
