% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{pkmSECellPlot}
\alias{pkmSECellPlot}
\title{Plot cell-specific decay curve for searcher efficiency}
\usage{
pkmSECellPlot(model, specificCell, col, axis_y = TRUE, axis_x = TRUE)
}
\arguments{
\item{model}{model of class pkm}

\item{specificCell}{name of the specific cell to plot (soon to be deprecated)}

\item{col}{color to use}

\item{axis_y}{logical of whether or not to plot the y axis}

\item{axis_x}{logical of whether or not to plot the x axis}
}
\value{
a cell plot panel
}
\description{
Plot cell-specific decay curve for searcher efficiency
}
