% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmSet}
\alias{pkmSet}
\title{Run a set of pkm models based on predictor inputs}
\usage{
pkmSet(formula_p, formula_k = NULL, data, obsCol = NULL,
  kFixed = NULL, kInit = 0.7, CL = 0.9, quiet = FALSE)
}
\arguments{
\item{formula_p}{Formula for p; an object of class "\code{\link{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted. Details of model specification are given under 
"Details".}

\item{formula_k}{Formula for k; an object of class "\code{\link{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted. Details of model specification are given under 
"Details".}

\item{data}{Dataframe with results from searcher efficiency trials and any
covariates included in \code{formula_p} or {formula_k} (required).}

\item{obsCol}{Vector of names of columns in \code{data} where results 
for each search occasion are stored (optional). If no \code{obsCol} are 
provided, \code{pkm} uses as \code{obsCol} all columns with names that 
begin with an \code{"s"} or \code{"S"} and end with a number, e.g., "s1",
"s2", "s3", etc. This option is included as a convenience for the user, 
but care must be taken that other data are not stored in columns with 
names matching that pattern. Alternatively, \code{obsCol} may be 
entered as a vector of names, like \code{c("s1", "s2", "s3")}, 
\code{paste0("s", 1:3)}, or \code{c("initialSearch", "anotherSearch", 
"lastSearch")}.}

\item{kFixed}{Parameter for user-specified \code{k} value (optional). If a
value is provided, \code{formula_k} is ignored and the model is fit under 
the assumption that the \code{k} parameter is fixed and known to be
\code{fix_k}.}

\item{kInit}{Initial value used for \code{k} in the optimization.}

\item{CL}{confidence level}

\item{quiet}{Logical indicator of whether or not to print messsages}
}
\value{
\code{pkmSet} returns a list of objects, each of class 
  "\code{pkm}", which each then a list whose components characterize the 
  fit of the specific model.
}
\description{
Run a set of \code{\link{pkm}} models based on all possible 
  models, given the predictor inputs. \code{pkmSet}'s inputs generally
  follow \code{\link{pkm}}, with all simpler models being run for all 
  included distributions and returned as a list of model objects
}
\examples{
  data(wind_RP)
  pkmSet(formula_p = p ~ Season, formula_k = k ~ Season, data = wind_RP$SE)

}
