% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmSetAllFail}
\alias{pkmSetAllFail}
\title{Check if all of the pkm models fail within a given set}
\usage{
pkmSetAllFail(pkmSetToCheck)
}
\arguments{
\item{pkmSetToCheck}{A \code{\link{pkmSet}} object to test}
}
\value{
A logical value indicating if all models failed in the set
}
\description{
Run a check on each model within a \code{\link{pkmSet}}
  object to determine if they all failed or not
}
