% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmSetFail}
\alias{pkmSetFail}
\title{Check if all of the pkm models fail}
\usage{
pkmSetFail(pkmSetToCheck)
}
\arguments{
\item{pkmSetToCheck}{A \code{pkmSet} object to test}
}
\value{
A vector of logical values indicating if each of the models failed
}
\description{
Check if all of the pkm models fail
}
