% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmSetFailRemove}
\alias{pkmSetFailRemove}
\title{Remove failed pkm models from a \code{pkmSet} object}
\usage{
pkmSetFailRemove(pkmSetToTidy)
}
\arguments{
\item{pkmSetToTidy}{A \code{\link{pkmSet}} object to tidy}
}
\value{
A \code{\link{pkmSet}} object with failed models removed
}
\description{
Remove all failed models within a \code{\link[=pkm]{pkmSet}} object
}
