% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmSetSizeFail}
\alias{pkmSetSizeFail}
\title{Check if all of the pkm models fail}
\usage{
pkmSetSizeFail(pkmSetSizeToCheck)
}
\arguments{
\item{pkmSetSizeToCheck}{A \code{pkmSetSize} object to test}
}
\value{
A list of vectors of logical values indicating if each of the 
  models failed
}
\description{
Check if all of the pkm models fail
}
