% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmSetSizeFailRemove}
\alias{pkmSetSizeFailRemove}
\title{Remove failed pkm models from a \code{pkmSetSize} object}
\usage{
pkmSetSizeFailRemove(pkmSetSizeToTidy)
}
\arguments{
\item{pkmSetSizeToTidy}{A list of \code{pkmSetSize} objects to tidy}
}
\value{
A list of \code{pkmSet} objects with failed models removed
}
\description{
Remove failed models from a \code{\link[=pkm]{pkmSetSize}} object
}
