% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{pkmSetSpecSECellPlot}
\alias{pkmSetSpecSECellPlot}
\title{Plot cell-specific decay curve for searcher efficiency for a specific model
  with comparison to the cellwise model}
\usage{
pkmSetSpecSECellPlot(modelSet, specificModel, specificCell, col_spec, col_full,
  axis_y = TRUE, axis_x = TRUE, n, seed_spec, seed_full)
}
\arguments{
\item{modelSet}{modelSet of class pkmSet}

\item{specificModel}{name of the specific submodel to plot}

\item{specificCell}{name of the specific cell to plot}

\item{col_spec}{color to use for the specific model}

\item{col_full}{color to use for the specific model}

\item{axis_y}{logical of whether or not to plot the y axis}

\item{axis_x}{logical of whether or not to plot the x axis}

\item{n}{number of draws to use to characterize the distributions}

\item{seed_spec}{random number seed to use for the specific model}

\item{seed_full}{random number seed to use for the full model}
}
\value{
a specific cell plot panel
}
\description{
Plot cell-specific decay curve for searcher efficiency for a specific model
  with comparison to the cellwise model
}
