% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{plot.cpm}
\alias{plot.cpm}
\title{Plot results of a single CP model}
\usage{
\method{plot}{cpm}(x, nsim = 500, seed = NULL, col = "black", ...)
}
\arguments{
\item{x}{model of class cpm}

\item{nsim}{number of draws to use to characterize the distributions}

\item{seed}{random number seed to use}

\item{col}{color to use}

\item{...}{to be passed down}
}
\description{
Plot a single \code{\link{cpm}} model
}
\examples{
  data(wind_RP)
  mod <- cpm(formula_l = l ~ Season, formula_s = s ~ Season,  
           data = wind_RP$CP, left = "Left", right = "Right"
          )
 plot(mod)

}
