% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{plot.cpmSet}
\alias{plot.cpmSet}
\title{Plot results of a cp model set}
\usage{
\method{plot}{cpmSet}(modelSet, specificModel = NULL, n = 500, seed = 1,
  col = c(exponential = rgb(0.8, 0.38, 0.56), weibull = rgb(1, 0.76, 0.15),
  loglogistic = rgb(0, 1, 1), lognormal = rgb(0, 0.41, 0.55)))
}
\arguments{
\item{modelSet}{pk model set of class pkmSet}

\item{specificModel}{the name of specific model(s) to restrict the plot}

\item{n}{number of draws to use to characterize the distributions}

\item{seed}{random number seed to use for the models}

\item{col}{color to use for the specific model}
}
\description{
Plot results of a cp model set
}
