% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_figure_functions.R
\name{plot.estM}
\alias{plot.estM}
\title{Plot total mortality estimation}
\usage{
\method{plot}{estM}(x, ..., CL = 0.95)
}
\arguments{
\item{x}{\code{estM} object}

\item{...}{arguments to pass down}

\item{CL}{confidence level}
}
\description{
\code{plot} defined for class \code{estM} objects
}
