% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_figure_functions.R
\name{plot.gGenericSize}
\alias{plot.gGenericSize}
\title{Plot results of a set of size-based generic ghat estimations}
\usage{
\method{plot}{gGenericSize}(x, CL = 0.95, ...)
}
\arguments{
\item{x}{\code{\link{gGenericSize}} output}

\item{CL}{confidence level to use}

\item{...}{to be passed down}
}
\value{
size-based detection probability plot
}
\description{
Plot method for a size-based generic \code{ghat} estimation
}
