% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_figure_functions.R
\name{plot.ghatGenericSize}
\alias{plot.ghatGenericSize}
\title{Plot results of a set of size-based generic ghat estimations}
\usage{
\method{plot}{ghatGenericSize}(x, CL = 0.9, ...)
}
\arguments{
\item{x}{ghatGenericSize output}

\item{CL}{confidence level to use}

\item{...}{to be passed down}
}
\value{
a plot
}
\description{
Plot results of a set of size-based generic ghat estimations
}
