% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{plot.pkmSet}
\alias{plot.pkmSet}
\title{Plot results of a pk model set}
\usage{
\method{plot}{pkmSet}(x, specificModel = NULL, n = 1000, seed_spec = 1,
  seed_full = 1, col_spec = "black", col_full = "grey", ...)
}
\arguments{
\item{x}{pk model set of class pkmSet}

\item{specificModel}{the name(s) of specific model(s) to restrict the plot}

\item{n}{number of draws to use to characterize the distributions}

\item{seed_spec}{random number seed to use for the specific model}

\item{seed_full}{random number seed to use for the full model}

\item{col_spec}{color to use for the specific model}

\item{col_full}{color to use for the specific model}

\item{...}{to be sent to subfunctions}
}
\value{
a set of plots
}
\description{
Plot results of a pk model set
}
