% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{plotCPCells}
\alias{plotCPCells}
\title{Plot the cellwise results of a single model in a set of CP models}
\usage{
plotCPCells(modelSet, specificModel, cols)
}
\arguments{
\item{modelSet}{cp model set of class cpmSet}

\item{specificModel}{the name of the specific model for the plot}

\item{cols}{named vector of the colors to use for the distributions}
}
\value{
a plot
}
\description{
Produce a set of cellwise figures for a specific CP model, as 
  fit by \code{\link{cpmSet}}
}
