% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{plotCPFigure}
\alias{plotCPFigure}
\title{Plot results of a single CP model in a set}
\usage{
plotCPFigure(modelSet, specificModel, cols = CPcols())
}
\arguments{
\item{modelSet}{cp model set of class \code{cpmSet}}

\item{specificModel}{the name of the specific model for the plot}

\item{cols}{named vector of the colors to use for the distributions}
}
\value{
a plot
}
\description{
Produce a figures for a specific CP model, as fit by
  \code{\link{cpmSet}}
}
