% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{plotCPHeader}
\alias{plotCPHeader}
\title{The CP plot header}
\usage{
plotCPHeader(modelSet, specificModel, cols = CPcols())
}
\arguments{
\item{modelSet}{cp model set of class cpmSet}

\item{specificModel}{the name of the specific model for the plot}

\item{cols}{named vector of the colors to use for the distributions}
}
\value{
a plot
}
\description{
Produce the header for a CP plot
}
