% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{plotNA}
\alias{plotNA}
\title{Produce a blank plot for unsucessful fits}
\usage{
plotNA(type = "model")
}
\arguments{
\item{type}{"model" or "split"}
}
\value{
dummy plot
}
\description{
Simply make a blank plot with descriptive text
}
