% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{plotSEBoxPlots}
\alias{plotSEBoxPlots}
\title{p and k box plots for an SE model set}
\usage{
plotSEBoxPlots(modelSet, specificModel, cols)
}
\arguments{
\item{modelSet}{modelSet of class pkmSet}

\item{specificModel}{name of the specific submodel to plot}

\item{cols}{named vector of colors to use for the specific and reference
models}
}
\value{
a set of parameter plot panels
}
\description{
Plot parameter box plots for each cell within a model for
  both p and k with comparison to the cellwise model (soon to be deprecated)
}
