% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{plotSEBoxTemplate}
\alias{plotSEBoxTemplate}
\title{template box plot}
\usage{
plotSEBoxTemplate(model_spec, col_spec)
}
\arguments{
\item{model_spec}{specific model object}

\item{col_spec}{color to use}
}
\value{
a template box plot
}
\description{
Plot template box plot
}
