% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{plotSECells}
\alias{plotSECells}
\title{Plot the cellwise results of a single model in a set of SE models}
\usage{
plotSECells(modelSet, specificModel, cols)
}
\arguments{
\item{modelSet}{pk model set of class pkmSet}

\item{specificModel}{the name of the specific model for the plot}

\item{cols}{named vector of colors to use for the specific and reference
models}
}
\value{
a plot
}
\description{
Produce a set of cellwise figures for a specific SE model, as
  fit by \code{\link{pkmSet}} (soon to be deprecated)
}
