% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{plotSEFigure}
\alias{plotSEFigure}
\title{Plot results of a single SE model in a set}
\usage{
plotSEFigure(modelSet, specificModel, app, cols)
}
\arguments{
\item{modelSet}{pk model set of class \code{pkmSet}}

\item{specificModel}{the name of the specific model for the plot}

\item{app}{logical indicating if the plot is for the app}

\item{cols}{named vector of colors to use for the specific and reference
models}
}
\value{
a plot
}
\description{
Produce a figures for a specific SE model, as fit by
  \code{\link{pkmSet}} (soon to be deprecated)
}
