% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{plotSEHeader}
\alias{plotSEHeader}
\title{The SE plot header}
\usage{
plotSEHeader(modelSet, specificModel, cols = SEcols())
}
\arguments{
\item{modelSet}{pk model set of class pkmSet}

\item{specificModel}{the name of the specific model for the plot}

\item{cols}{named vector of colors to use for the specific and reference
models}

\item{sizeclassName}{name of the size class if it is to be added to the
figure}
}
\value{
a plot
}
\description{
Produce the header for an SE plot
}
