% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{ppersist}
\alias{ppersist}
\title{Calculate the probability of persistence to detection}
\usage{
ppersist(distr, t_arrive0, t_arrive1, t_search, pdb, pda = NULL)
}
\arguments{
\item{distr}{Distribution used.}

\item{t_arrive0}{Beginning of arrival window.}

\item{t_arrive1}{End of arrival window.}

\item{t_search}{Search time.}

\item{pdb}{parameter b.}

\item{pda}{parameter a.}
}
\value{
Probability of persistence of detection to at t_search, 
         given arrival between t_arrive0 and t_arrive1
}
\description{
Calculate the probability of persistence to detection
}
\examples{
NA
}
