% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{ppersist_loglogistic}
\alias{ppersist_loglogistic}
\title{Calculate the probability of persistence to for a loglogistic}
\usage{
ppersist_loglogistic(t_arrive0, t_arrive1, t_search, pda, pdb)
}
\arguments{
\item{t_arrive0}{Beginning of arrival window.}

\item{t_arrive1}{End of arrival window.}

\item{t_search}{Search time.}

\item{pda}{parameter a.}

\item{pdb}{parameter b.}
}
\value{
Probability of persistence of detection to at t_search, 
         given arrival between t_arrive0 and t_arrive1
}
\description{
Calculate the probability of persistence to for a loglogistic
}
\examples{
NA
}
