% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{predsCols}
\alias{predsCols}
\title{Select the predictor-ok columns from a data table}
\usage{
predsCols(data)
}
\arguments{
\item{data}{data table}
}
\value{
column names of columns that can be predictors
}
\description{
Simple function to facilitate selection of columns that could
  be predictors for SE or CP models from a data table
}
