% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{prepPredictors}
\alias{prepPredictors}
\title{Prepares predictors based on inputs}
\usage{
prepPredictors(preds = NULL)
}
\arguments{
\item{preds}{input predictors}
}
\value{
prepared predictors (or 1 if no predictors)
}
\description{
Prepares predictors based on inputs
}
