% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{prepSizeclassText}
\alias{prepSizeclassText}
\title{Prepare text for size classes}
\usage{
prepSizeclassText(sizeclasses)
}
\arguments{
\item{sizeclasses}{names of the size classes}
}
\value{
prepared and render name text
}
\description{
Prepare text for size classes
}
