% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_functions.R
\name{prepSSCO}
\alias{prepSSCO}
\title{Prepare the SS and CO data}
\usage{
prepSSCO(data_SS, data_CO, dateSearchedCol = "DateSearched",
  dateFoundCol = "DateFound", unitCol = "Unit")
}
\arguments{
\item{data_SS}{Search Schedule data}

\item{data_CO}{Carcass observation data}

\item{dateSearchedCol}{Column name for the date searched data}

\item{dateFoundCol}{Column name for the date found data}

\item{unitCol}{Column name for the unit indicator}
}
\value{
list of [1] SS with days (starting with t = 0), [2] CO data with 
  days (starting with t = 0), and [3] a vector of the CO_data rows 
  indicating the carcasses found during cleanout searches.
}
\description{
Turns dates into days from the first day (cleanout search, set
  as time 0) and removes any carcasses observed on the first days
}
\examples{
NA
}
