% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{prettyModTabCP}
\alias{prettyModTabCP}
\title{Creates the pretty Carcass Persistence model table}
\usage{
prettyModTabCP(modTab, CL = 0.9)
}
\arguments{
\item{modTab}{model table}

\item{CL}{Confidence level}
}
\value{
column names
}
\description{
Based on confidence level of interest
}
