% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_table_functions.R
\name{prettySplitSpecTab}
\alias{prettySplitSpecTab}
\title{Create the pretty version of a specific split summary table}
\usage{
prettySplitSpecTab(splitSummarySpec)
}
\arguments{
\item{splitSummarySpec}{a specific split summary}
}
\value{
specific split pretty table
}
\description{
Format a reader-friendly version of a specific split summary 
  table from a mortality estimation
}
