% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{prettySplitTab}
\alias{prettySplitTab}
\title{Pretty a split summary table}
\usage{
prettySplitTab(splitSummary)
}
\arguments{
\item{splitSummary}{a split summary}
}
\value{
split pretty table
}
\description{
Pretty a split summary table
}
