% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{print.cpm}
\alias{print.cpm}
\title{Print a \code{\link{cpm}} model object}
\usage{
\method{print}{cpm}(x, ...)
}
\arguments{
\item{x}{a \code{\link{cpm}} model object}

\item{...}{to be passed down}
}
\description{
Print a \code{\link{cpm}} model object
}
