% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{print.pkm}
\alias{print.pkm}
\title{Print a \code{\link{pkm}} model object}
\usage{
\method{print}{pkm}(x, ...)
}
\arguments{
\item{x}{a \code{\link{pkm}} model object}

\item{...}{to be passed down}
}
\description{
Print a \code{\link{pkm}} model object
}
