% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{probability_persist_to_detection}
\alias{probability_persist_to_detection}
\title{Calculate the probability of persistence to detection}
\usage{
probability_persist_to_detection(persistence_distribution, t_arrive0, t_arrive1,
  t_search, pdb, pda = NULL)
}
\arguments{
\item{persistence_distribution}{Distribution used.}

\item{t_arrive0}{Beginning of arrival window.}

\item{t_arrive1}{End of arrival window.}

\item{t_search}{Search time.}

\item{pdb}{parameter b.}

\item{pda}{parameter a.}
}
\value{
Probability of persistence of detection to at t_search, 
         given arrival between t_arrive0 and t_arrive1
}
\description{
Calculate the probability of persistence to detection
}
\examples{
NA
}
