% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{rcp}
\alias{rcp}
\title{Simulate parameters from a fitted cp model.}
\usage{
rcp(n = 1, model, seed = NULL, type = "survreg")
}
\arguments{
\item{n}{the number of simulation draws}

\item{model}{A \code{\link{cpm}} object (which is returned from 
\code{cpm()})}

\item{seed}{optional input to set the seed of the RNG}

\item{type}{The type of parameters requested. \code{"survreg"} or 
\code{"ppersist"}}
}
\value{
list of two matrices of \code{n} simulated \code{l} and \code{s} 
  (if \code{type = "survreg"}) or \code{a} and \code{b} (if \code{type = 
  "ppersist"})for cells defined by the \code{model} object.
}
\description{
Simulate parameters from a fitted cp model.
}
\examples{
 NA
}
