% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{readCSV}
\alias{readCSV}
\title{Read in csv files in either format}
\usage{
readCSV(path)
}
\arguments{
\item{path}{file path}
}
\value{
read in data table
}
\description{
Handle reading in of a csv that is either comma-decimal or
  semicolon-comma separation style
}
