% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{readCSVs}
\alias{readCSVs}
\title{Read in csv files in either format}
\usage{
readCSVs(path)
}
\arguments{
\item{path}{file path}
}
\value{
read in data table
}
\description{
Handle reading in of a csv that is either comma-decimal or
  semicolon-comma separation style
}
