% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{removeCols}
\alias{removeCols}
\title{Remove selected columns from column names}
\usage{
removeCols(colNames, selCols)
}
\arguments{
\item{colNames}{column names from which some could be removed}

\item{selCols}{selected columns to be removed}
}
\value{
column names without selected columns
}
\description{
Simple function to facilitate removal of columns selected
}
