% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{removeSelCols}
\alias{removeSelCols}
\title{Removes selected columns from column names}
\usage{
removeSelCols(colNames, selCols)
}
\arguments{
\item{colNames}{column names}

\item{selCols}{selected columns to be removed}
}
\value{
column names without selected columns
}
\description{
Simple function to facilitate removal of columns selected
}
