% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_utilities.R
\name{renderDTns}
\alias{renderDTns}
\title{Render a data table without server-side processing}
\usage{
renderDTns(x)
}
\arguments{
\item{x}{HTML \code{datatable} widget output from 
\code{\link[DT]{datatable}}.}
}
\value{
Rendered \code{x}.
}
\description{
Simply render the data table without server-side processing.
}
