% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_utilities.R
\name{reNULL}
\alias{reNULL}
\title{Reset values of a list to NULL}
\usage{
reNULL(x, toNULL)
}
\arguments{
\item{x}{\code{list} object to have elements \code{toNULL} reset to 
\code{NULL}.}

\item{toNULL}{Names of elements in \code{x} to reset to \code{NULL}.}
}
\value{
Updated \code{x}.
}
\description{
Utility function for clearing and setting purposes.
}
