% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_resources_functions.R
\name{resourcesPanel}
\alias{resourcesPanel}
\title{Resources Main Panel UI element}
\usage{
resourcesPanel()
}
\value{
HTML for the Analysis panel
}
\description{
create the HTML code for the Resources panel
}
