% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{rghat}
\alias{rghat}
\title{Simulate ghat values for a set of carcasses from fitted pk and cp 
  models and search data}
\usage{
rghat(n = 1, data_carc, data_ss, model_SE, model_CP, seed_SE = 1,
  seed_CP = 1, kFill = NULL, unitColName = "Unit")
}
\arguments{
\item{n}{the number of simulation draws}

\item{data_carc}{Carcass Observation data}

\item{data_ss}{Search Schedule data}

\item{model_SE}{Searcher Efficiency model}

\item{model_CP}{Carcass Persistence model}

\item{seed_SE}{seed for random draws of the SE model}

\item{seed_CP}{seed for random draws of the CP model}

\item{kFill}{value to fill in for missing k when not existing in the model}

\item{unitColName}{Column name for the unit indicator}
}
\value{
matrix of n ghat estimates for each carcass
}
\examples{
NA
}
