% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{rghatCarcass}
\alias{rghatCarcass}
\title{Simulate ghat values and arrival intervals for a carcass from fitted pk 
  and cp models and search data}
\usage{
rghatCarcass(n = 1, data_carc, dist, data_SS, preds_SE, preds_CP, sim_SE,
  sim_CP, preds_static, preds_dynamic, unitCol, timeFoundCol, timeSearchedCol)
}
\arguments{
\item{n}{the number of simulation draws}

\item{data_carc}{Carcass Observation data for the carcass}

\item{dist}{distribution for the CP model}

\item{data_SS}{Search Schedule data}

\item{preds_SE}{Names of Searcher Efficiency predictors}

\item{preds_CP}{Names of Carcass Persistence predictors}

\item{sim_SE}{simulated Search Efficiency parameters}

\item{sim_CP}{simulated Carcass Persistence parameters}

\item{preds_static}{Names of predictors that don't change over time}

\item{preds_dynamic}{Names of predictors that change over time}

\item{unitCol}{Column name for the unit indicator}

\item{timeFoundCol}{Column name for the time found data}

\item{timeSearchedCol}{Column name for the time searched data}
}
\value{
list of [1] n ghat estimates for a carcass and [2] n arrival 
  interval (Aj) estimates for a carcass
}
\description{
Simulate ghat values and arrival intervals for a carcass from fitted pk 
  and cp models and search data
}
\examples{
NA
}
