% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{rghatGenericSize}
\alias{rghatGenericSize}
\title{Generic ghat estimation for a combination of SE model and CP model 
  under a given search schedule}
\usage{
rghatGenericSize(n = 1, data_SS, modelSetSize_SE, modelSetSize_CP,
  modelSizeSelections_SE, modelSizeSelections_CP, seed_SE = NULL,
  seed_CP = NULL, kFill = NULL)
}
\arguments{
\item{n}{the number of simulation draws}

\item{data_SS}{Search schedule data as a vector of days searched}

\item{modelSetSize_SE}{Searcher Efficiency model set for multiple sizes}

\item{modelSetSize_CP}{Carcass Persistence model set for multiple sizes}

\item{modelSizeSelections_SE}{vector of SE models to use, one for each size}

\item{modelSizeSelections_CP}{vector of CP models to use, one for each size}

\item{seed_SE}{seed for random draws of the SE model}

\item{seed_CP}{seed for random draws of the CP model}

\item{kFill}{value to fill in for missing k when not existing in the model}
}
\value{
list of ghat estimates, with one element in the list corresponding
  to each of the cells from the cross-model combination
}
\description{
The g estimated by \code{rghatGeneric} is a generic aggregate detection 
  probability and represents the probability of detecting a carcass that 
  arrives at a (uniform) random time during the period monitored, for each
  of the possible cell combinations, given the SE and CP models. This 
  is somethat different from the GenEst estimation of g when the purpose 
  is to estimate total mortality (M), in which case the detection 
  probability varies with carcass arrival interval and is difficult to 
  summarize statistically. The \code{rghatGeneric} estimate is a useful 
  "big picture" summary of detection probability, but would be difficult
  to work with for estimating M with precision.
}
