% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{rMhat}
\alias{rMhat}
\title{Draw values for Mhat given a ghat and DWP}
\usage{
rMhat(n = 1, ghat, DWP = 1, seed = 1)
}
\arguments{
\item{n}{number of random draws per carcass}

\item{ghat}{estimated detection probability}

\item{DWP}{Density weighted proportion associated with Mtilde}

\item{seed}{seed used to set the random number generator}
}
\value{
Mhat
}
\description{
Draw values for Mhat given a ghat and DWP
}
\examples{
NA
}
