% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{rMtilde}
\alias{rMtilde}
\title{Draw values for Mtilde given a ghat}
\usage{
rMtilde(n = 1, ghat, seed = 1)
}
\arguments{
\item{n}{number of random draws}

\item{ghat}{estimated detection probability}

\item{seed}{seed used to set the random number generator}
}
\value{
Mtilde
}
\description{
Draw values for Mtilde given a ghat
}
\examples{
NA
}
