% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{rpk}
\alias{rpk}
\title{Simulate p and k parameters from a fitted pk model.}
\usage{
rpk(nsim, pkmodel)
}
\arguments{
\item{nsim}{the number of simulation draws}

\item{pkmodel}{A \code{\link{pkm}} object (which is returned from \code{pkm()})}
}
\value{
Array of \code{nsim} simulated pairs of \code{p} and \code{k} for
 cells defined by the \code{pkmodel} object.
}
\description{
Simulate p and k parameters from a fitted pk model.
}
\examples{
pkmod1 <- pkm(p ~ 1, k ~ 1, data = pkmdat)
simulated_pk <- rpk(nsim = 10, pkmodel = pkmod1)
simulated_pk
boxplot(simulated_pk[,, 1])

pkmod2 <- pkm(p ~ visibility * season, k ~ site, data = pkmdat)
rpk(nsim = 10, pkmodel = pkmod2)
}
