% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{rtcbinom1}
\alias{rtcbinom1}
\title{Random draw from a continuous binomial random variable with
 size = 1/g and truncated at g}
\usage{
rtcbinom1(n, prob)
}
\arguments{
\item{n}{number of random draws}

\item{g}{Vector of probabilities}
}
\value{
mean
}
\description{
Random draw from a continuous binomial random variable with
 size = 1/g and truncated at g
}
