% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_functions.R
\name{runGenEst}
\alias{runGenEst}
\title{Launch the app}
\usage{
runGenEst()
}
\description{
Launches a local version of the GenEst application by running
  \code{shinyAppDir} pointed to the \code{app} subdirectory in the 
  local \code{GenEst} package folder.
}
