% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{se_aicc_table_create}
\alias{se_aicc_table_create}
\title{Create the  AICc tables for the searcher efficiency models}
\usage{
se_aicc_table_create(models)
}
\arguments{
\item{models}{Search efficiency models fit for each size class.}
}
\value{
List of AICc tables (one table per size class)
}
\description{
Create the  AICc tables for the searcher efficiency models
}
\examples{
NA
}
