% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{se_model_fit}
\alias{se_model_fit}
\title{Fit a single searcher efficiency model for a single size class.}
\usage{
se_model_fit(pequation, kequation, factor_combination_table, data, zeros, found,
  observation_columns, init_k_value, fix_k, fix_k_value)
}
\arguments{
\item{pequation}{Equation for p (character).}

\item{kequation}{Equation for k (character).}

\item{factor_combination_table}{Factor combination table.}

\item{data}{Searcher efficiency data restricted to a single size class.}

\item{zeros}{Vector of the number of searches on which each 
carcass was missed.}

\item{found}{Vector of the search number for the search on which
each carcass was found.}

\item{observation_columns}{Indicator vector of which columns in the data}

\item{init_k_value}{Initial value of k to use in optim.}

\item{fix_k}{Logical of whether k should be fixed.}

\item{fix_k_value}{Value of k if fixed.}
}
\value{
Model fit list.
}
\description{
Fit a single searcher efficiency model for a single size class.
}
\examples{
NA
}
