% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{se_model_set_across_sizes_fit}
\alias{se_model_set_across_sizes_fit}
\title{Fit all possible searcher efficiency models across all size classes.}
\usage{
se_model_set_across_sizes_fit(data, predictors, observation_columns,
  size_class_column, init_k_value, fix_k, fix_k_value)
}
\arguments{
\item{data}{Searcher efficiency data.}

\item{predictors}{Names of predictor variables to include.}

\item{observation_columns}{Indicator vector of which columns in the data}

\item{size_class_column}{Column header for the size class column.}

\item{init_k_value}{Initial value of k to use in optim.}

\item{fix_k}{Logical of whether k should be fixed.}

\item{fix_k_value}{Value of k if fixed.}
}
\value{
List of model fit lists.
}
\description{
Fit all possible searcher efficiency models across all size classes.
}
\examples{
NA
}
