% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{se_model_set_fit}
\alias{se_model_set_fit}
\title{Fit all possible searcher efficiency models for a single size class}
\usage{
se_model_set_fit(predictors, data, init_k_value, fix_k, fix_k_value,
  observation_columns)
}
\arguments{
\item{predictors}{Names of predictor variables to include.}

\item{data}{Searcher efficiency data restricted to a single size class.}

\item{init_k_value}{Initial value of k to use in optim.}

\item{fix_k}{Logical of whether k should be fixed.}

\item{fix_k_value}{Value of k if fixed.}

\item{observation_columns}{Indicator vector of which columns in the data}
}
\value{
List of model fit lists.
}
\description{
Fit all possible searcher efficiency models for a single size class
}
\examples{
NA
}
