% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{se_theta_create}
\alias{se_theta_create}
\title{Draw replicate samples of parameters from the searcher efficiency models 
across all size classes.}
\usage{
se_theta_create(data, predictors, size_class_column, model_fits, replicates,
  fix_k, fix_k_value)
}
\arguments{
\item{data}{Searcher efficiency data.}

\item{predictors}{Names of predictor variables to include.}

\item{size_class_column}{Column header for the size class column.}

\item{model_fits}{Searcher efficiency model fits.}

\item{replicates}{Number of replicate samples to draw from the 
distribution.}

\item{fix_k}{Logical of whether k should be fixed.}

\item{fix_k_value}{Value of k if fixed.}
}
\value{
Multidimensional array (replicates, 2, factor combination 
    cells, models, size classes) of searcher efficiency parameters (p, k).
}
\description{
Draw replicate samples of parameters from the searcher efficiency models 
across all size classes.
}
\examples{
NA
}
