% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{SEboxes}
\alias{SEboxes}
\title{Produce boxplots \code{p} and/or \code{k} for all cells for reference
 model and specific model}
\usage{
SEboxes(pk, cells_set, gdat_spc, gdat_ref, cols_SE = cols_SE)
}
\arguments{
\item{pk}{either \code{p} or \code{k}, depending on which type of boxplots
need to be inserted}
}
\value{
NULL inserts a boxplot panel into pkmSet plot
}
\description{
Each call to \code{SEboxes} produces a series of boxplots for
 either \code{p} or \code{k}. For interal use only, for producing figs for
 \code{plot.pkmSet}. Function requires that ncell, preds_set, lev1, lev2,
 boxW, bsep, gdat_spc, and gdat_ref are defined prior to the function call.
}
\keyword{internal}
