% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{SEcols}
\alias{SEcols}
\title{Produce a named vectory with standard SE plot colors}
\usage{
SEcols()
}
\description{
Produce a named vectory with standard SE plot colors. soon to be
 deprecated.
}
