% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_SE_functions.R
\name{SEEstimatesPanel}
\alias{SEEstimatesPanel}
\title{Analysis Searcher Efficiency main panel estimates UI element}
\usage{
SEEstimatesPanel()
}
\value{
HTML for the Analysis Searcher Efficiency model estimates panel
}
\description{
create the HTML code for the Analysis Searcher Efficiency
  main panel model estimates element
}
