% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{SEfig}
\alias{SEfig}
\title{Plot results of a single SE model in a set}
\usage{
SEfig(referenceModel, specificModel, gdat_spc, gdat_ref, found, available,
  cells_set, cols_SE, p, CL)
}
\description{
Produce a figures for a specific SE model, as fit by
  \code{\link{pkmSet}}
}
\keyword{internal}
