% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_SE_functions.R
\name{SEFiguresPanel}
\alias{SEFiguresPanel}
\title{Analysis Searcher Efficiency main panel figures UI element}
\usage{
SEFiguresPanel()
}
\value{
HTML for the Analysis Searcher Efficiency figures panel
}
\description{
create the HTML code for the Analysis Searcher Efficiency
  main panel figures element
}
