% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{selectData}
\alias{selectData}
\title{Select particular columns from a data set}
\usage{
selectData(data, cols)
}
\arguments{
\item{data}{data table to selec from}

\item{cols}{column names to select}
}
\value{
selected data
}
\description{
Select particular columns from a data set
}
