% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_panels.R
\name{selectedDataPanel}
\alias{selectedDataPanel}
\title{Create a Selected Data Tab Panel for the GenEst User Interface HTML}
\usage{
selectedDataPanel(modType)
}
\arguments{
\item{modType}{Toggle control for the model type of the panel. One of 
"SE", "CP", or "g".}
}
\value{
HTML for the panel
}
\description{
This is a generalized function for creating a data input 
  visualization panel used in the GenEst GUI, based on the data type 
  (\code{dataType}).
}
