% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_SE_functions.R
\name{SEMainPanel}
\alias{SEMainPanel}
\title{Analysis Searcher Efficiency main panel UI element}
\usage{
SEMainPanel()
}
\value{
HTML for the Analysis Searcher Efficiency main panel
}
\description{
create the HTML code for the Analysis Searcher Efficiency
  main
}
