% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_SE_functions.R
\name{SEModComparisonPanel}
\alias{SEModComparisonPanel}
\title{Analysis Searcher Efficiency main panel model comparison UI element}
\usage{
SEModComparisonPanel()
}
\value{
HTML for the Analysis Searcher Efficiency model comparison panel
}
\description{
create the HTML code for the Analysis Searcher Efficiency
  main panel model comparison element
}
