% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_SE_functions.R
\name{SEModSelectionPanel}
\alias{SEModSelectionPanel}
\title{Analysis Searcher Efficiency main panel model selection UI element}
\usage{
SEModSelectionPanel()
}
\value{
HTML for the Analysis Searcher Efficiency model selection panel
}
\description{
create the HTML code for the Analysis Searcher Efficiency
  main panel model selection element
}
