% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_SE_functions.R
\name{SEPanel}
\alias{SEPanel}
\title{Analysis Searcher Efficiency Main Panel UI element}
\usage{
SEPanel()
}
\value{
HTML for the Analysis SE panel
}
\description{
create the HTML code for the Analysis SE panel
}
