% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_SE_functions.R
\name{SESelectedDataPanel}
\alias{SESelectedDataPanel}
\title{Analysis Searcher Efficiency main panel selected data UI element}
\usage{
SESelectedDataPanel()
}
\value{
HTML for the Analysis Searcher Efficiency selected data panel
}
\description{
create the HTML code for the Analysis Searcher Efficiency
  main panel selected data element
}
