% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{SEsi0}
\alias{SEsi0}
\title{Calculate searcher efficiency after some searches for a single pk 
  combination}
\usage{
SEsi0(days, pk)
}
\arguments{
\item{days}{search days}

\item{pk}{pk}
}
\value{
searcher efficiency that matches the output of ppersist
}
\description{
Calculate searcher efficiency after some searches for a single pk 
  combination
}
\examples{
NA
}
