% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{SEsi_right}
\alias{SEsi_right}
\title{Calculate conditional probability of observation after a series of 
  searches}
\usage{
SEsi_right(nsi, pk)
}
\arguments{
\item{nsi}{number of searches after arrival}

\item{pk}{numeric array of searcher efficiency p and k parameters
(p = pk[ , 1] and k = pk[ , 2])}
}
\value{
numeric nsi x dim(pk)[1] array of probabilities of observing a
 carcass after 1:nsi searches (assuming that the carcass had not been
previously discovered or removed by scavengers
}
\description{
Calculate the conditional probability of observing a carcass 
  after i = 1:nsi searches (assuming carcass is not previous discovered by 
  searchers or removed by scavengers)
}
