% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_SE_functions.R
\name{SESidebar}
\alias{SESidebar}
\title{Analysis Searcher Efficiency sidebar panel UI element}
\usage{
SESidebar()
}
\value{
HTML for the Analysis Searcher Efficiency sidebar
}
\description{
create the HTML code for the Analysis Searcher Efficiency
  sidebar
}
