% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{setFigW}
\alias{setFigW}
\title{Set the figure width based on the number of cells}
\usage{
setFigW(modSet)
}
\arguments{
\item{modSet}{model set}
}
\value{
figure width
}
\description{
Set the figure width based on the number of cells
}
