% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{setkFillNeed}
\alias{setkFillNeed}
\title{Create the kFillNeed text}
\usage{
setkFillNeed(rv)
}
\arguments{
\item{rv}{reactive value list}
}
\value{
kFillNeed character of "yes" or "no"
}
\description{
Based on the number of observation columns, create text output
  of "yes" or "no"
}
