% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{setkNeed}
\alias{setkNeed}
\title{Create the kNeed text}
\usage{
setkNeed(rv)
}
\arguments{
\item{rv}{reactive value list}
}
\value{
kNeed character of "yes" or "no"
}
\description{
Based on the number of observation columns, create text output
  of "yes" or "no"
}
