% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_utilities.R
\name{setNotSuspending}
\alias{setNotSuspending}
\title{(Re)set outputOptions to not suspending for given elements}
\usage{
setNotSuspending(output, dontSuspend)
}
\arguments{
\item{output}{\code{output} list to have elements \code{dontSuspend} 
(re)set to having \code{suspendWhenHidden = FALSE}.}

\item{dontSuspend}{Names of elements in \code{output} to (re)set to 
having \code{suspendWhenHidden = FALSE}.}
}
\description{
Utility function for clearing and setting purposes.
}
