% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_figure_functions.R
\name{simpleMplot}
\alias{simpleMplot}
\title{Plot total mortality estimation for a simple situation}
\usage{
simpleMplot(M, ..., CL = 0.9)
}
\arguments{
\item{...}{arguments to pass down}

\item{CL}{confidence level}

\item{x}{Mortality object}
}
\description{
Hisogram and empirical PDF on top
}
