% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{small}
\alias{small}
\title{HTML small function}
\usage{
small(...)
}
\arguments{
\item{...}{attributes and children of the element}
}
\value{
HTML small text
}
\description{
Generate a small text HTML object. This simply pulls the
  function definition from the tag environment in htmltools
}
