% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solar_pv}
\alias{solar_pv}
\title{Synthetic solar facility data}
\format{A list with 5 items:
\describe{
  \item{SE}{Searcher efficiency trial data}
  \item{CP}{Carcass persistence trial data}
  \item{SS}{Search schedule data}
  \item{DWP}{Density weighted proportion of area searched data}
  \item{CO}{Carcass observations}  
}}
\source{
Mock
}
\usage{
solar_pv
}
\description{
A dataset containing SE, CP, SS, DWP, and CO data.
}
\keyword{datasets}
