% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{solarTroughData}
\alias{solarTroughData}
\title{Synthetic solar facility data}
\format{A list with 5 items:
\describe{
  \item{SearcherEfficiencyData}{Searcher efficiency trial data}
  \item{CarcassPersistenceData}{Carcass persistence trial data}
  \item{SearchData}{Search schedule data}
  \item{DensityWeightedProportionData}{Density weighted proportion of area 
    searched data}
  \item{CarcassObservationData}{Carcass observations}  
}}
\source{
Mock
}
\usage{
solarTroughData
}
\description{
A dataset containing SE, CP, SS, DWP, and CO data.
}
\keyword{datasets}
