% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits_functions.R
\name{SS}
\alias{SS}
\title{Create search schedule data into an SS object for convenient splits 
 analyses}
\usage{
SS(data_SS, dateCol = NULL, preds = NULL)
}
\arguments{
\item{data_SS}{data frame or matrix with search schedule parameters, 
including columns for search dates, covariates (describing characteristics
of the search intervals), and each unit (with 0s and 1s in rows to 
indicate whether the given unit was searched on the given date)}

\item{dateColumn}{name of the column with the search dates in it}

\item{covars}{vector of character strings giving the names of columns to be
interpreted as potential covariates.}
}
\value{
\code{SS} object that can be conveniently used in the splitting
 functions.
}
\description{
Since data_SS columns largely have a specific, required
  format, the \code{SS} function can often decipher the data automatically,
  but the user may specify explicit instructions for parsing the data for
  safety if desired. If the data are formatted properly, the automatic 
  parsing is reliable in most cases. There are two exceptions. (1) If 
  there is more than one column with possible dates (formatted as formal
  dates (as class \code{Date}, \code{POSIXlt} or \code{POSIXct}) or 
  character strings or factors that can be unambiguously interpreted as 
  dates (with assumed format "2018-05-15" or "2018/5/15"). In that case, 
  the user must specify the desired dates as \code{dateColumn}. (2) If
  there is a covariate column consisting entirely of 0s and 1s. In that
  case, the user must specify the column(s) in \code{covars}.
If no \code{dateColumn} is given, \code{SS} will attempt to find the date
  column based on data formats. If there is exactly one column that can be
  interpreted as dates, that column will be taken as the dates. Column 
  names that are not listed as \code{covars} will be interpreted as 
  "units" if their data consists solely of 0s and 1s. Otherwise, they will
  be included in the catch-all category of "covariates".
}
