% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{style}
\alias{style}
\title{HTML style function}
\usage{
style(...)
}
\arguments{
\item{...}{attributes and children of the element}
}
\value{
HTML style for use with a specific element
}
\description{
Generate HTML inline style for an element
}
