% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_functions.R
\name{summary.estM}
\alias{summary.estM}
\title{Summarize total mortality estimation}
\usage{
\method{summary}{estM}(object, ..., CL = 0.9)
}
\arguments{
\item{object}{\code{estM} object}

\item{...}{arguments to pass down}

\item{CL}{confidence level}
}
\description{
\code{summary} defined for class \code{estM} objects
}
