% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{summary.gGeneric}
\alias{summary.gGeneric}
\title{Summarize the gGeneric list to a simple table}
\usage{
\method{summary}{gGeneric}(object, ..., CL = 0.9)
}
\arguments{
\item{object}{gGeneric output list (each element is a named vector of 
gGeneric values for a cell in the model combinations)}

\item{...}{arguments to be passed down}

\item{CL}{confidence level}
}
\value{
a summary table of g values (medians and confidence bounds) for 
  each cell combination within the gGeneric list
}
\description{
Summarize the gGeneric list to a simple table
}
