% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{summary.gGenericSize}
\alias{summary.gGenericSize}
\title{Summarize the gGenericSize list to a list of simple tables}
\usage{
\method{summary}{gGenericSize}(object, ..., CL = 0.95)
}
\arguments{
\item{object}{gGenericSize output list (each element is a size-named 
list of named vectors of gGeneric values for a cell in the model 
combinations)}

\item{...}{arguments to be passed down}

\item{CL}{confidence level}
}
\value{
a list of summary tables of g values (medians and confidence 
  bounds) for each cell combination within the gGeneric list
}
\description{
Summarize the gGenericSize list to a list of simple tables
}
