% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{summary.ghatGeneric}
\alias{summary.ghatGeneric}
\title{Summarize the ghatGeneric list to a simple table}
\usage{
\method{summary}{ghatGeneric}(object, ..., CL = 0.9)
}
\arguments{
\item{object}{ghatGeneric output list (each element is a named vector of 
ghatGeneric values for a cell in the model combinations)}

\item{...}{arguments to be passed down}

\item{CL}{confidence level}
}
\value{
a summary table of ghat values (means and confidence bounds) for 
  each cell combination within the ghatGeneric list
}
\description{
Summarize the ghatGeneric list to a simple table
}
