% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{summary.ghatGenericSize}
\alias{summary.ghatGenericSize}
\title{Summarize the ghatGenericSize list to a list of simple tables}
\usage{
\method{summary}{ghatGenericSize}(object, ..., CL = 0.9)
}
\arguments{
\item{object}{ghatGenericSize output list (each element is a size-named 
list of named vectors of ghatGeneric values for a cell in the model 
combinations)}

\item{...}{arguments to be passed down}

\item{CL}{confidence level}
}
\value{
a list of summary tables of ghat values (medians and confidence 
  bounds) for each cell combination within the ghatGeneric list
}
\description{
Summarize the ghatGenericSize list to a list of simple tables
}
