% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{tidyModelSetCP}
\alias{tidyModelSetCP}
\title{Tidy a CP model set}
\usage{
tidyModelSetCP(modelSet)
}
\arguments{
\item{modelSet}{cp model set of class cpmSet}
}
\value{
a trimmed model set
}
\description{
Remove bad fit models
}
