% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{tidyModelSetSE}
\alias{tidyModelSetSE}
\title{Tidy an SE model set}
\usage{
tidyModelSetSE(modelSet)
}
\arguments{
\item{modelSet}{pk model set of class pkmSet}
}
\value{
a trimmed model set
}
\description{
Remove bad fit models
}
