% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splits_functions.R
\name{transposeSplits}
\alias{transposeSplits}
\title{Transpose a \code{splitFull} array (preserving attributes)}
\usage{
transposeSplits(splits)
}
\arguments{
\item{splits}{a \code{splitFull} object, which is a list of \code{n}
\code{m} x \code{k} arrays with attributes describing characteristics of 
the splits}
}
\value{
a list of \code{m} \code{n} x \code{k} arrays as a \code{splitFull}
 object
}
\description{
Transpose a \code{splitFull} array (preserving attributes)
}
