% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{u}
\alias{u}
\title{HTML u function}
\usage{
u(...)
}
\arguments{
\item{...}{attributes and children of the element}
}
\value{
HTML underlined
}
\description{
Generate an HTML underlined text element. This simply pulls 
  the function definition from the tag environment in htmltools
}
