% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{Create the UI}
\usage{
ui(request)
}
\arguments{
\item{request}{request as required for bookmarking}
}
\value{
UI
}
\description{
Render the active UI for the GenEst application
}
