% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{ul}
\alias{ul}
\title{HTML ul function}
\usage{
ul(...)
}
\arguments{
\item{...}{attributes and children of the element}
}
\value{
HTML unordered list
}
\description{
Generate an unordered list HTML object. This simply pulls the
  function definition from the tag environment in htmltools
}
