% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_clear_all}
\alias{update_input_clear_all}
\title{Update the inputs when everything is cleared out}
\usage{
update_input_clear_all(rv, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when everything is cleared
}
