% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_cols_CP}
\alias{update_input_cols_CP}
\title{Update the remaining columns when CP data columns are selected}
\usage{
update_input_cols_CP(rv, input, session, x = "ltp")
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}

\item{session}{session}

\item{x}{specific column}
}
\description{
Update the inputs when the CP data columns are selected
}
