% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_cols_fta}
\alias{update_input_cols_fta}
\title{Update the predictor options and Last Time Present options when the 
  First Time Absent column is selected}
\usage{
update_input_cols_fta(rv, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the First Time Absent column is 
  selected
}
