% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_cols_SE}
\alias{update_input_cols_SE}
\title{Update the remaining columns when SE data columns are selected}
\usage{
update_input_cols_SE(rv, input, session, x = "obsCols")
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}

\item{session}{session}

\item{x}{specific column}
}
\description{
Update the inputs when the SE data columns are selected
}
