% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_cols_SE_preds}
\alias{update_input_cols_SE_preds}
\title{Update the observation options when SE predictor columns are 
  selected}
\usage{
update_input_cols_SE_preds(rv, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the SE predictor columns are selected
}
