% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_data_CO}
\alias{update_input_data_CO}
\title{Update the inputs when CO data are read in}
\usage{
update_input_data_CO(rv, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the CO data file is input
}
