% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_data_CP}
\alias{update_input_data_CP}
\title{Update the inputs when CP data are read in}
\usage{
update_input_data_CP(rv, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the CP data file is input
}
