% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_data_SS}
\alias{update_input_data_SS}
\title{Update the inputs when SS data are read in}
\usage{
update_input_data_SS(rv, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the SS data file is input
}
