% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_data_SS_clear}
\alias{update_input_data_SS_clear}
\title{Update the inputs when SS data are cleared out}
\usage{
update_input_data_SS_clear(rv, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the SS data file input is cleared
}
