% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_file_DWP}
\alias{update_input_file_DWP}
\title{Update the inputs when DWP data are read in}
\usage{
update_input_file_DWP(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the DWP data file is input
}
