% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_file_SE}
\alias{update_input_file_SE}
\title{Update the inputs when SE data are read in}
\usage{
update_input_file_SE(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the SE data file is input
}
