% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_ltp}
\alias{update_input_ltp}
\title{Update the predictor options and First Time Absent options when the 
  Last Time Present column is selected}
\usage{
update_input_ltp(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the Last Time Present column is 
  selected
}
