% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_obsSE}
\alias{update_input_obsSE}
\title{Update the predictor options when SE observation columns are 
  selected}
\usage{
update_input_obsSE(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the SE observation columns are selected
}
