% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_outsc_CP}
\alias{update_input_outsc_CP}
\title{Update the CP output dropdown selections when the size class is 
  chosen}
\usage{
update_input_outsc_CP(rv, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the CP output dropdown selections when the size 
  class is chosen
}
