% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_outSEclass}
\alias{update_input_outSEclass}
\title{Update the SE output dropdown selections when the size class is 
  chosen}
\usage{
update_input_outSEclass(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the SE output dropdown selections when the size 
  class is chosen
}
