% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_predsCP}
\alias{update_input_predsCP}
\title{Update the observation options when CP predictor columns are 
  selected}
\usage{
update_input_predsCP(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the CP predictor columns are selected
}
