% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_run_CP}
\alias{update_input_run_CP}
\title{Update the CP output dropdown selections when the model is run}
\usage{
update_input_run_CP(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the CP output dropdown selections when the model is run
}
