% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_run_CP_clear}
\alias{update_input_run_CP_clear}
\title{Update the inputs when CP model is cleared}
\usage{
update_input_run_CP_clear(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the CP model is cleared
}
