% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_run_M}
\alias{update_input_run_M}
\title{Update the M output dropdown selections and the Fraction Surveyed
  when the model is run}
\usage{
update_input_run_M(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the M output dropdown selections and the Fraction 
  Surveyed when the model is run
}
