% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_run_SE}
\alias{update_input_run_SE}
\title{Update the SE output dropdown selections when the model is run}
\usage{
update_input_run_SE(rv, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the SE output dropdown selections when the model is run
}
