% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_sizeclassCol}
\alias{update_input_sizeclassCol}
\title{Update the inputs when the size class column is selected}
\usage{
update_input_sizeclassCol(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}

\item{session}{session}
}
\description{
Update the inputs when the SE data file is input
}
