% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_split_M_clear}
\alias{update_input_split_M_clear}
\title{Update the inputs when M split is cleared}
\usage{
update_input_split_M_clear(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the inputs when the M split is cleared
}
