% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_input_update_functions.R
\name{update_input_useSSdata}
\alias{update_input_useSSdata}
\title{Update the SS average dropdown selections when the SS data are used}
\usage{
update_input_useSSdata(rv, input, session)
}
\arguments{
\item{rv}{reactive values list}

\item{session}{session}
}
\description{
Update the SS average dropdown selections when the SS data 
  are used
}
