% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_update_output.R
\name{update_output}
\alias{update_output}
\title{Update the outputs when an event occurs}
\usage{
update_output(eventName, rv, output, input)
}
\arguments{
\item{eventName}{Character name of the event. One of "clear_all",
"file_SE", "file_SE_clear", "file_CP", "file_CP_clear", "file_SS",
"file_SS_clear", "file_DWP", "file_DWP_clear", "file_CO", 
"file_CO_clear", "class", "obsSE", "predsSE", "run_SE", "run_SE_clear",
"outSEclass", "outSEp", "outSEk", "ltp", "fta", "predsCP", "run_CP",
"run_CP_clear", "outCPclass", "outCPdist", "outCPl", "outCPs",
"run_M", "run_M_clear", "split_M", "split_M_clear", "transpose_split",
"run_g", "run_g_clear", or "outgclass".}

\item{rv}{Reactive values list for the GenEst GUI.}

\item{output}{\code{output} list for the GenEst GUI.}

\item{input}{\code{input} lisst for the GenEst GUI}
}
\value{
Updated \code{output} list.
}
\description{
When an event occurs in the GenEst GUI, the output values may
  need to be updated. This function contains all of the possible updates
  based on the event options (or lacks any updates if the event doesn't
  require any).
}
