% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_clear_all}
\alias{update_output_clear_all}
\title{Update the reactive value list when everything is cleared out}
\usage{
update_output_clear_all(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}
}
\value{
an updated output list
}
\description{
Update the rv list when everything is cleared
}
