% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_data_SE}
\alias{update_output_data_SE}
\title{Update the output list when SE data are read in}
\usage{
update_output_data_SE(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}
}
\value{
an updated output list
}
\description{
Update the output list when the SE data file is input
}
