% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_obsSE}
\alias{update_output_obsSE}
\title{Update the selected SE data when a column is selected}
\usage{
update_output_obsSE(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}
}
\value{
an updated output list
}
\description{
Update the output table when the SE columns are selected
}
