% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_outdls_CP}
\alias{update_output_outdls_CP}
\title{Update the CP output when a distribuition or l or s equation is 
  chosen}
\usage{
update_output_outdls_CP(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}
}
\value{
an updated output list
}
\description{
Update the SE output when a distribution or l or s equation is 
  chosen
}
