% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_outpk_SE}
\alias{update_output_outpk_SE}
\title{Update the SE output when a p or k equation is chosen}
\usage{
update_output_outpk_SE(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}

\item{session}{session}
}
\value{
an updated output list
}
\description{
Update the SE output when a p or k equation is chosen
}
