% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_outsc_g}
\alias{update_output_outsc_g}
\title{Update the output when a generic g size class is chosen}
\usage{
update_output_outsc_g(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}

\item{session}{session}
}
\value{
an updated output list
}
\description{
Update the output table when a generic g size class is chosen
}
