% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_outSEp}
\alias{update_output_outSEp}
\title{Update the SE output when a p or k equation is chosen}
\usage{
update_output_outSEp(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}
}
\value{
an updated output list
}
\description{
Update the SE output when a p or k equation is chosen
}
