% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_run_CP}
\alias{update_output_run_CP}
\title{Update the output when an CP model has been run}
\usage{
update_output_run_CP(rv, output, session)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}

\item{session}{session}
}
\value{
an updated output list
}
\description{
Update the output table when an CP model has been run
}
