% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_run_CP_clear}
\alias{update_output_run_CP_clear}
\title{Update the output list when CP model is cleared}
\usage{
update_output_run_CP_clear(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}
}
\value{
an updated output list
}
\description{
Update the output list when the CP model is cleared
}
