% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_sizeclassCol}
\alias{update_output_sizeclassCol}
\title{Update the selected CP and SE data when a size class column is 
  selected}
\usage{
update_output_sizeclassCol(rv, input, output)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}

\item{output}{output list}
}
\value{
an updated output list
}
\description{
Update the output tables when the size class column is
  selected
}
