% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_SS}
\alias{update_output_SS}
\title{Update the SS text output when the SS is updated}
\usage{
update_output_SS(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}

\item{session}{session}
}
\value{
an updated output list
}
\description{
Update the SS text output when the SS is updated
}
