% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_update_functions.R
\name{update_output_transpose_split}
\alias{update_output_transpose_split}
\title{Update the output when the M split is transposed}
\usage{
update_output_transpose_split(rv, output)
}
\arguments{
\item{rv}{reactive values list}

\item{output}{output list}

\item{session}{session}
}
\value{
an updated output list
}
\description{
Update the output when the M split is transposed
}
