% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_update_rv.R
\name{update_rv}
\alias{update_rv}
\title{Update the reactive value list when an event occurs}
\usage{
update_rv(eventName, rv, input)
}
\arguments{
\item{eventName}{Character name of the event. One of "clear_all",
"file_SE", "file_SE_clear", "file_CP", "file_CP_clear", "file_SS",
"file_SS_clear", "file_DWP", "file_DWP_clear", "file_CO", 
"file_CO_clear", "class", "obsSE", "predsSE", "run_SE", "run_SE_clear",
"outSEclass", "outSEp", "outSEk", "ltp", "fta", "predsCP", "run_CP",
"run_CP_clear", "outCPclass", "outCPdist", "outCPl", "outCPs",
"run_M", "run_M_clear", "split_M", "split_M_clear", "transpose_split",
"run_g", "run_g_clear", "outgclass", "load_RP", "load_RPbat",
"load_cleared", "load_PV", "load_trough", "load_powerTower", or "load_mock"}

\item{rv}{Reactive values list for the GenEst GUI, created by 
\code{\link{initialReactiveValues}}, which calls 
\code{\link[shiny]{reactiveValues}}}

\item{input}{\code{input} list for the GenEst GUI.}
}
\value{
Updated \code{rv} list.
}
\description{
When an event occurs in the GenEst GUI, the reactive values 
  need to be updated. This function contains all of the possible updates
  based on the event options.
}
