% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_clear_all}
\alias{update_rv_clear_all}
\title{Update the reactive value list when everything is cleared out}
\usage{
update_rv_clear_all(rv, input)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}
}
\value{
an updated reactive values list
}
\description{
Update the rv list when everything is cleared
}
