% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_cols_ltp}
\alias{update_rv_cols_ltp}
\title{Update the reactive value list when a CP Last Time Present column is 
  selected}
\usage{
update_rv_cols_ltp(rv, input)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}
}
\value{
an updated reactive values list
}
\description{
Update the rv list when a CP Last Time Present column is 
  selected
}
