% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_cols_SE_preds}
\alias{update_rv_cols_SE_preds}
\title{Update the reactive value list when an SE predictor column is 
  selected}
\usage{
update_rv_cols_SE_preds(rv, input)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}
}
\value{
an updated reactive values list
}
\description{
Update the rv list when an SE predictor column is selected
}
