% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_data_CP}
\alias{update_rv_data_CP}
\title{Update the reactive value list when CP data are read in}
\usage{
update_rv_data_CP(rv, input)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}
}
\value{
an updated reactive values list
}
\description{
Update the rv list when the CP data file is input
}
