% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_data_DWP_clear}
\alias{update_rv_data_DWP_clear}
\title{Update the reactive value list when DWP data are cleared out}
\usage{
update_rv_data_DWP_clear(rv, input)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}
}
\value{
an updated reactive values list
}
\description{
Update the rv list when the DWP data file input is cleared
}
