% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_data_SS}
\alias{update_rv_data_SS}
\title{Update the reactive value list when SS data are read in}
\usage{
update_rv_data_SS(rv, input)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}
}
\value{
an updated reactive values list
}
\description{
Update the rv list when the SS data file is input
}
