% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_file_SE}
\alias{update_rv_file_SE}
\title{Update the reactive value list when SE data are read in}
\usage{
update_rv_file_SE(rv, input)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}
}
\value{
an updated reactive values list
}
\description{
Update the rv list when the SE data file is input
}
