% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_outdls_CP}
\alias{update_rv_outdls_CP}
\title{Update the CP reactive values when a distribution or l or s model 
  is chosen}
\usage{
update_rv_outdls_CP(rv, input)
}
\arguments{
\item{rv}{the reactive values list}

\item{input}{the input list}
}
\value{
an updated reactive values list
}
\description{
Update the CP reactive values when a distribution or l or s 
  model is chosen
}
