% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_outsc_SE}
\alias{update_rv_outsc_SE}
\title{Update the SE reactive values when the size class is chosen}
\usage{
update_rv_outsc_SE(rv, input)
}
\arguments{
\item{rv}{the reactive values list}

\item{input}{the input list}
}
\value{
an updated reactive values list
}
\description{
Update the SE reactive values when the size class is chosen
}
