% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_outSEp}
\alias{update_rv_outSEp}
\title{Update the SE reactive values when a p or k model is chosen}
\usage{
update_rv_outSEp(rv, input)
}
\arguments{
\item{rv}{the reactive values list}

\item{input}{the input list}
}
\value{
an updated reactive values list
}
\description{
Update the SE reactive values when a p or k model is chosen
}
