% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_run_g_clear}
\alias{update_rv_run_g_clear}
\title{Update the reactive value list when g model is cleared}
\usage{
update_rv_run_g_clear(rv, input)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}
}
\value{
an updated reactive values list
}
\description{
Update the rv list when the g model is cleared
}
