% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_run_M}
\alias{update_rv_run_M}
\title{Run the M Model}
\usage{
update_rv_run_M(rv, input)
}
\arguments{
\item{rv}{the reactive values list}

\item{input}{the input list}
}
\value{
an updated reactive values list
}
\description{
Use the inputs to run the M model requested by the UI
}
