% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_split_M}
\alias{update_rv_split_M}
\title{Update the M reactive values when M is split}
\usage{
update_rv_split_M(rv, input)
}
\arguments{
\item{rv}{the reactive values list}

\item{input}{the input list}
}
\value{
an updated reactive values list
}
\description{
Update the M reactive values when M is split
}
