% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_split_M_clear}
\alias{update_rv_split_M_clear}
\title{Update the reactive value list when M split is cleared}
\usage{
update_rv_split_M_clear(rv, input)
}
\arguments{
\item{rv}{reactive values list}

\item{input}{input list}
}
\value{
an updated reactive values list
}
\description{
Update the rv list when the M split is cleared
}
