% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_transpose_split}
\alias{update_rv_transpose_split}
\title{Update the M reactive values when M split is transposed}
\usage{
update_rv_transpose_split(rv)
}
\arguments{
\item{rv}{the reactive values list}
}
\value{
an updated reactive values list
}
\description{
Update the M reactive values when M split is transposed
}
