% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_useSSdata}
\alias{update_rv_useSSdata}
\title{Update the SS reactive values when the SS are chosen}
\usage{
update_rv_useSSdata(rv)
}
\arguments{
\item{rv}{the reactive values list}
}
\value{
an updated reactive values list
}
\description{
Update the SS reactive values when the SS are chosen
}
