% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_rv_update_functions.R
\name{update_rv_useSSinputs}
\alias{update_rv_useSSinputs}
\title{Update the SS reactive values when the average SS is chosen}
\usage{
update_rv_useSSinputs(rv, input)
}
\arguments{
\item{rv}{the reactive values list}

\item{input}{the input list}
}
\value{
an updated reactive values list
}
\description{
Update the SS reactive values when the average SS is chosen
}
