% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{updateColNames_all}
\alias{updateColNames_all}
\title{Updates the string of column names that are in all the needed tables}
\usage{
updateColNames_all(colnames_SE, colnames_CP, colnames_CO)
}
\arguments{
\item{colnames_SE}{column names for the searcher efficiency}

\item{colnames_CP}{column names for the carcass persistence}

\item{colnames_CO}{column names for the carcass observations}
}
\value{
possible column names
}
\description{
Updates the string of column names that are in all the needed tables
}
