% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{updateColNames_size}
\alias{updateColNames_size}
\title{Update the string of column names that are in all the needed tables}
\usage{
updateColNames_size(rv)
}
\arguments{
\item{rv}{reactive values list}
}
\value{
possible column names
}
\description{
Determine the overlap between the column names in the SE, CP,
   and CO data tables.
}
