% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{updateSizeclassCol}
\alias{updateSizeclassCol}
\title{Updates the name of the size class column based on available names}
\usage{
updateSizeclassCol(sizeclassCol, colNames_all)
}
\arguments{
\item{sizeclassCol}{current size class column name}

\item{colNames_all}{updated vector of column names in all needed tables}
}
\value{
updated sizeclassCol
}
\description{
if the existing size class column name is no longer in the
  set of available names, a NULL is returned to reset the column name
}
