% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{updateSizeclasses}
\alias{updateSizeclasses}
\title{Update the size classes}
\usage{
updateSizeclasses(data, sizeclassCol)
}
\arguments{
\item{data}{data table to draw sizes from}

\item{sizeclassCol}{size class column name}
}
\value{
unique size classes
}
\description{
Update the size classes
}
