% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_resources_functions.R
\docType{data}
\name{USGSpath}
\alias{USGSpath}
\title{Resources Main Panel UI element}
\format{An object of class \code{character} of length 1.}
\usage{
USGSpath
}
\value{
HTML for the Analysis panel
}
\description{
create the HTML code for the Resources panel
}
\keyword{datasets}
