% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{whichCleanout}
\alias{whichCleanout}
\title{Determine which carcasses were from cleanout searches}
\usage{
whichCleanout(data_carc, data_ss, unitColName = "Unit")
}
\arguments{
\item{data_carc}{Carcass observation data}

\item{data_ss}{Search schedule data}

\item{unitColName}{Column name for the unit indicator}
}
\value{
index values of which carcasses were taken on the first search
}
\description{
Determine which carcasses were from cleanout searches
}
\examples{
NA
}
