% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_widgets.R
\name{widgetMaker}
\alias{widgetMaker}
\title{Input Widget Maker}
\usage{
widgetMaker(Condition, Name, Fun, Label, Args)
}
\arguments{
\item{Condition}{Condition under which the widget is present to the user.}

\item{Name}{Name (id) of the widget created.}

\item{Fun}{Function name (as character) used to create the widget.}

\item{Label}{Label presented to the user in the UI for the widget.}

\item{Args}{List of any additional arguments to be passed to the widget 
creating function.}
}
\value{
HTML for the widget.
}
\description{
Basic generalized function for creating an input widget based
  on the condition of the widget being presented, the name of the widget, 
  the function used to create it, it's label on the UI, and any additional 
  arguments.
}
