% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind_RP}
\alias{wind_RP}
\title{Synthetic data set based on 120 m radius road and pad searches of all 100
  turbines at a theoretical site.}
\format{A list with 5 items:
\describe{
  \item{SE_trials}{Searcher efficiency trial data}
  \item{CP_trials}{Carcass persistence trial data}
  \item{SS_params}{Search schedule parameters}
  \item{DWP_data}{Density weighted proportion of area searched}
  \item{CO_data}{Carcass observations}
}}
\source{
RP
}
\usage{
wind_RP
}
\description{
Complete data set for fatality estimation at a wind power facility. 
  Included are SE, CP, SS, DWP, and CO data.
}
\keyword{datasets}
