% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind_RPbat}
\alias{wind_RPbat}
\title{Synthetic data set based on 120 m radius road and pad searches of all 100
  turbines at a theoretical site. Bat data only}
\format{A list with 5 elements:
\describe{
  \item{SE}{Searcher efficiency trial data}
  \item{CP}{Carcass persistence trial data}
  \item{SS}{Search schedule parameters}
  \item{DWP}{Density weighted proportion of area searched}
  \item{CO}{Carcass observations}
}}
\source{
RP
}
\usage{
wind_RPbat
}
\description{
Complete data set for fatality estimation at a wind power facility.
  Included are SE, CP, SS, DWP, and CO data.
}
\keyword{datasets}
