% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{yyyymmdd}
\alias{yyyymmdd}
\title{Converts the standard Excel date format (m/d/YYYY) to YYYY-mm-dd}
\usage{
yyyymmdd(x)
}
\arguments{
\item{x}{date(s) to convert}
}
\value{
converted dates
}
\description{
only transforms m/d/Y dates, returns untransformed dates that
  were already YYYY-mm-dd
}
\examples{
NA
}
