% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{ZeroAfterOne}
\alias{ZeroAfterOne}
\title{Check if any 0s come after a 1 in a string}
\usage{
ZeroAfterOne(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
logical if any 0s come after a 1
}
\description{
Do any 0s come after a 1? Only applicable when only 1 1 is 
  observed.
}
