% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_about_functions.R
\name{aboutPanel}
\alias{aboutPanel}
\title{About Main Panel UI element}
\usage{
aboutPanel()
}
\value{
HTML for the about panel
}
\description{
create the HTML code for the About main panel
}
