% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{aicc}
\alias{aicc}
\title{Generic S3 function for summarizing AICc}
\usage{
aicc(x, ...)
}
\arguments{
\item{x}{Model or list of models to extract AICc values from.}

\item{...}{further arguments passed to or from other methods}
}
\value{
list of models sorted by AICc
}
\description{
Extract AICc values from \code{pkm}, \code{pkmSet},
 \code{pkmSetSize}, \code{cpm}, \code{cpmSet}, and \code{cpmSetSize}.
}
