% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{alogit}
\alias{alogit}
\title{Compute the anti-logit.}
\usage{
alogit(x)
}
\arguments{
\item{x}{A number.}
}
\value{
The anti-logit of \code{x}.
}
\description{
Compute the anti-logit.
}
\examples{
alogit(0)
}
