% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{cButtonStyle}
\alias{cButtonStyle}
\title{Define the style for the clear buttons}
\usage{
cButtonStyle(type = "single")
}
\arguments{
\item{type}{"single" or "all"}
}
\value{
character element defining the style
}
\description{
Define the style tag for clear buttons used throughout GenEst.
}
