% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_utilities.R
\name{cButtonStyle}
\alias{cButtonStyle}
\title{Define Clear Button Style}
\usage{
cButtonStyle(buttonType = "single")
}
\arguments{
\item{buttonType}{"single" (for clearing a single component) or "all" (for
clearing everything).}
}
\value{
Character element defining the style.
}
\description{
Define the style tag for clear buttons used throughout GenEst.
}
