% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{checkComponents}
\alias{checkComponents}
\title{Check if all component terms and interactions are included in a formula}
\usage{
checkComponents(formula)
}
\arguments{
\item{formula_in}{A formula object

Terms are automatically alphabatized within interactions.}
}
\description{
Check if all component terms and interactions are included in a formula
}
