% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{clearNotifications}
\alias{clearNotifications}
\title{Clear specific notifications}
\usage{
clearNotifications(msg_NobsSE = NULL, msg_SampleSizeSE = NULL,
  msg_ModFailSE = NULL, msg_ModFailCP = NULL, msg_SampleSizeCP = NULL,
  msg_ModFailM = NULL, msg_ModFailg = NULL, msg_avgSSfail = NULL)
}
\arguments{
\item{msg_NobsSE}{Number of search occasions message for SE}

\item{msg_SampleSizeSE}{Sample size message for SE}

\item{msg_ModFailSE}{Fail message for SE}

\item{msg_ModFailCP}{Fail message for CP}

\item{msg_SampleSizeCP}{Sample size message for CP}

\item{msg_ModFailM}{Fail message for mortality estimation}

\item{msg_ModFailg}{Fail message for generic g estimation}

\item{msg_avgSSfail}{Fail message for averaging search schedule}
}
\value{
Nothing
}
\description{
Clear notifications within the app.
}
