% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{combinePredsAcrossModels}
\alias{combinePredsAcrossModels}
\title{Combine predictors across models}
\usage{
combinePredsAcrossModels(preds_CP, preds_SE, data_CP, data_SE)
}
\arguments{
\item{preds_CP}{Carcass persistence predictor variables.}

\item{preds_SE}{Searcher efficiency predictor variables.}

\item{data_CP}{Carcass persistence data.}

\item{data_SE}{Searcher efficiency data.}
}
\value{
Factor combination table.
}
\description{
Create the factor combination table for a set of searcher
 efficiency and carcass persistence analyses.
}
