% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmSetAICcTab}
\alias{cpmSetAICcTab}
\title{Create the AICc tables for a set of carcass persistence models}
\usage{
cpmSetAICcTab(cpmset, quiet = FALSE)
}
\arguments{
\item{cpmset}{Set of carcass persistence models fit to the same
observations}

\item{quiet}{Logical indicating if messages should be printed}
}
\value{
AICc table
}
\description{
Generates model comparison tables based on AICc values for
  a set of CP models generated by \code{\link{cpmSet}}
}
\examples{
  data(wind_RP)
  mod <- cpmSet(formula_l = l ~ Season, formula_l = s ~ Season,  
           data = wind_RP$CP, left = "Left", right = "Right"
          )
 cpmSetAICcTab(mod)

}
