% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmSetFailRemove}
\alias{cpmSetFailRemove}
\title{Remove failed cpm models from a \code{\link{cpmSet}} object}
\usage{
cpmSetFailRemove(cpmSetToTidy)
}
\arguments{
\item{cpmSetToTidy}{A \code{\link{cpmSet}} object to tidy}
}
\value{
A \code{\link{cpmSet}} object with failed models removed
}
\description{
Remove all failed models within a \code{\link{cpmSet}} object
}
