% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{cpmSetSizeFailRemove}
\alias{cpmSetSizeFailRemove}
\title{Remove failed cpm models}
\usage{
cpmSetSizeFailRemove(cpmSetSizeToTidy)
}
\arguments{
\item{cpmSetSizeToTidy}{A list of \code{cpmSetSize} objects to tidy}
}
\value{
A list of \code{cpmSet} objects with failed models removed
}
\description{
Remove failed cpm models
}
