% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{cpmSetSpecCPCellPlot}
\alias{cpmSetSpecCPCellPlot}
\title{Plot cell-specific decay curve for carcass persistence}
\usage{
cpmSetSpecCPCellPlot(modelSet, specificModel, specificCell, cols, axes)
}
\arguments{
\item{modelSet}{modelSet of class cpmSet}

\item{specificModel}{name of the specific submodel to plot}

\item{specificCell}{name of the specific cell to plot}

\item{cols}{named vector of the colors to use for the distributions}

\item{axes}{named vector of logical values indicating whether or not to 
plot the x axis and the y axis}
}
\value{
a specific cell plot panel
}
\description{
Produce the figure panel for a specific cell (factor 
  level combination) including the specific fitted decay curves.
}
