% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_panels.R
\name{dataTabPanel}
\alias{dataTabPanel}
\title{Create a Data Tab Panel for the GenEst User Interface HTML}
\usage{
dataTabPanel(dataType)
}
\arguments{
\item{dataType}{Toggle control for the model type of the panel. One of 
"SE", "CP", "SS", "DWP", or "CO".}
}
\value{
HTML for the panel
}
\description{
This is a generalized function for creating a data input 
  visualization panel used in the GenEst GUI, based on the data type 
  (\code{dataType}).
}
