% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{dateCols}
\alias{dateCols}
\title{Select the date columns from a data table}
\usage{
dateCols(data)
}
\arguments{
\item{data}{data table potentially containing columns that could be
coerced (via \code{checkDate()}) into a properly formatted date}
}
\value{
column names of columns that can be coerced to dates
}
\description{
Simple function to facilitate selection of date columns from
  a data table
}
