% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dateToDay}
\alias{dateToDay}
\title{Calculate day of study from calendar date}
\usage{
dateToDay(date, ref = NULL)
}
\arguments{
\item{date}{A date or vector of dates to convert to days.}

\item{ref}{Reference date.}
}
\value{
Numeric value(s) of days from \code{ref}.
}
\description{
Convert calendar date to integer day from a reference date
  (\code{ref}).
}
\examples{
  x <- c("2018-01-01", "2018-02-01")
  dateToDay(x, x[1])

}
