% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{dateToDay}
\alias{dateToDay}
\title{Calculate day of study from calendar date}
\usage{
dateToDay(date, ref = NULL)
}
\arguments{
\item{date}{date to convert}

\item{ref}{reference date}
}
\value{
converted days from reference
}
\description{
Convert calendar date to day from reference
}
\examples{
  x <- c("2018-01-01", "2018-02-01")
  dateToDay(x, x[1])

}
