% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_disclaimers_functions.R
\name{disclaimers}
\alias{disclaimers}
\title{Help Disclaimers UI element}
\usage{
disclaimers(type)
}
\arguments{
\item{type}{"base" (for local versions) or "deploy" (for hosted version)}
}
\value{
HTML for the Help Disclaimers text
}
\description{
create the HTML code for the Help Disclaimers
}
