% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadData}
\alias{downloadData}
\title{Download a zipped data set}
\usage{
downloadData(set, csvformat = NULL)
}
\arguments{
\item{set}{the name of the data set to download (character string)}

\item{csvformat}{Format for .csv file to be downloaded. For comma-separated
fields ( , ) and period ( . ) for decimal separator, \code{csvformat = ""}
or \code{csvformat = 2}. For semicolon separated fields ( ; ) and a comma
for decimal separator, \code{csvformat = 2}}
}
\value{
a download handler function
}
\description{
Handle the downloading of a data set
}
