% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadData}
\alias{downloadData}
\title{Download a zipped data set}
\usage{
downloadData(set, csvformat = NULL)
}
\arguments{
\item{set}{the name of the data set to download}

\item{csvformat}{Format of .csv files to download. For comma field 
separator and period decimal separator, use \code{csvformat = NULL} or "".
For semicolon field separator and comma decimal separator, use
\code{csvformat = 2}.}
}
\value{
a download handler function
}
\description{
Handle the downloading of a data set
}
