% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadMFig}
\alias{downloadMFig}
\title{Download the M figure}
\usage{
downloadMFig(rv, split = TRUE)
}
\arguments{
\item{rv}{the reactive values list}

\item{split}{logical indicator to use the split or not}
}
\value{
a download handler function
}
\description{
Handle the M figure downloading
}
