% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadSEFig}
\alias{downloadSEFig}
\title{Download the SE figure}
\usage{
downloadSEFig(rv)
}
\arguments{
\item{rv}{the reactive values list}
}
\value{
a download handler function
}
\description{
Handle the SE figure downloading
}
