% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_downloads_functions.R
\name{downloadsPanel}
\alias{downloadsPanel}
\title{Downloads Main Panel UI element}
\usage{
downloadsPanel()
}
\value{
HTML for the help panel
}
\description{
create the HTML code for the Downloads main panel
}
