% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_download_functions.R
\name{downloadTable}
\alias{downloadTable}
\title{Download a table}
\usage{
downloadTable(filename, tablename, csvformat)
}
\arguments{
\item{filename}{the name for the file writing out}

\item{tablename}{the name of the table in the rv list}

\item{csvformat}{format for .csv file: "" or NULL for comma-separated, 2 
for semi-colon separated}
}
\value{
a download handler function
}
\description{
Handle the downloading of a table
}
