% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{DWPbyCarcass}
\alias{DWPbyCarcass}
\title{Expand the density weighted proportion table to a value for each carcass}
\usage{
DWPbyCarcass(data_DWP, data_CO, unitCol = "Unit", sizeclassCol = NULL,
  data_SS = NULL, dateFoundCol = "DateFound",
  dateSearchedCol = "DateSearched")
}
\arguments{
\item{data_DWP}{Survey unit (rows) by size (columns) density weighted 
proportion table}

\item{data_CO}{Carcass observation data}

\item{unitCol}{Column name for the unit indicator}

\item{sizeclassCol}{Name of colum in \code{data_CO} where the size 
classes are recorded}

\item{data_SS}{Search Schedule data}

\item{dateFoundCol}{Column name for the date found data (if cleanout 
searches are to be removed)}

\item{dateSearchedCol}{Column name for the date searched data (if cleanout 
searches are to be removed)}
}
\value{
DWP value for each carcass
}
\description{
Expand the density weighted proportion table to a value for each carcass
}
\examples{
NA
}
