% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{DWPbyCarcass}
\alias{DWPbyCarcass}
\title{Expand the density weighted proportion table to a value for each carcass}
\usage{
DWPbyCarcass(data_DWP, data_carc, data_ss, unitCol = "Unit",
  dateFoundCol = "DateFound", dateSearchedCol = "DateSearched",
  sizeclassCol = NULL)
}
\arguments{
\item{data_DWP}{Survey unit (rows) by size (columns) density weighted 
proportion table}

\item{data_carc}{Carcass observation data}

\item{data_ss}{Search Schedule data}

\item{unitCol}{Column name for the unit indicator}

\item{dateFoundCol}{Column name for the date found data}

\item{dateSearchedCol}{Column name for the date searched data}

\item{sizeclassCol}{Name of colum in \code{data_carc} where the size 
classes are recorded}
}
\value{
DWP value for each carcass
}
\description{
Expand the density weighted proportion table to a value for each carcass
}
\examples{
NA
}
