% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fatality_functions.R
\name{DWPbyCarcass}
\alias{DWPbyCarcass}
\title{Expand the density weighted proportion table to a value for each carcass
  (within a size class) based on the unit where they were found}
\usage{
DWPbyCarcass(data_DWP, data_CO, data_SS, dateFoundCol = "DateFound",
  dateSearchedCol = "DateSearched", DWPCol = NULL, unitCol = "Unit")
}
\arguments{
\item{data_DWP}{Survey unit (rows) by size (columns) density weighted 
proportion table}

\item{data_CO}{Carcass observation data}

\item{data_SS}{Search Schedule data}

\item{dateFoundCol}{Column name for the date found data}

\item{dateSearchedCol}{Column name for the date searched data}

\item{DWPCol}{Column name for the DWP values in the DWP table when no
size class is used}

\item{unitCol}{Column name for the unit indicator}
}
\value{
DWP value for each carcass
}
\description{
Expand the density weighted proportion table to a value for each carcass
  (within a size class) based on the unit where they were found
}
\examples{
NA
}
