% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_functions.R
\name{DWPbyCarcass}
\alias{DWPbyCarcass}
\title{Assign DWP Value to Each Carcass}
\usage{
DWPbyCarcass(data_DWP, data_CO, sizeclassCol = NULL, unitCol = NULL,
  DWPCol = NULL)
}
\arguments{
\item{data_DWP}{Survey unit (rows) by size (columns) density weighted 
proportion table}

\item{data_CO}{Carcass observation data}

\item{sizeclassCol}{Name of colum in \code{data_CO} where the size classes
are recorded. Optional.}

\item{unitCol}{Column name for the unit indicator (optional)}

\item{DWPcol}{Name of column where DWP values are stored (optional). Used
when there is more than one DWP column in \code{data_DWP} but analysis is
intended for a single class (i.e., no "size" is specified in data_CO).}
}
\value{
DWP value for each carcass
}
\description{
Expand the density weighted proportion table to a value for 
  each carcass (across multiple classes if desired) based on the unit where 
  they were found
}
\examples{
NA
}
