% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{Ecbinom}
\alias{Ecbinom}
\title{Expected value of a continuous binomial with size = 1/g}
\usage{
Ecbinom(prob)
}
\arguments{
\item{prob}{Vector of probabilities.}
}
\value{
Mean value of the probabilities.
}
\description{
Calculates the expected value of a continuous binomial random
 variable with size = 1/g. Uses internal-only data.
}
