% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{estg}
\alias{estg}
\title{Estimate g values and arrival intervals for a set of carcasses 
  from fitted pk and cp models and search data}
\usage{
estg(nsim = 1, data_CO, data_SS, model_SE, model_CP, seed_SE = NULL,
  seed_CP = NULL, seed_g = NULL, kFill = NULL, unitCol = "Unit",
  dateFoundCol = "DateFound", dateSearchedCol = "DateSearched",
  sizeclassCol = NULL, cleanoutCarcs = NULL)
}
\arguments{
\item{nsim}{the number of simulation draws}

\item{data_CO}{Carcass Observation data}

\item{data_SS}{Search Schedule data}

\item{model_SE}{Searcher Efficiency model (or list of models if there are 
multiple size classes)}

\item{model_CP}{Carcass Persistence model (or list of models if there are 
multiple size classes)}

\item{seed_SE}{seed for random draws of the SE model}

\item{seed_CP}{seed for random draws of the CP model}

\item{seed_g}{seed for random draws of the gs}

\item{kFill}{value(s) to fill in for missing k when not existing in the 
model(s)}

\item{unitCol}{Column name for the unit indicator}

\item{dateFoundCol}{Column name for the date found data}

\item{dateSearchedCol}{Column name for the date searched data}

\item{sizeclassCol}{Name of colum in \code{data_CO} where the size classes
are recorded}

\item{cleanoutCarcs}{of which carcasses (if any) were found on cleanout 
searches}
}
\value{
list of [1] g estimates, [2] arrival interval (Aj) estimates,
  [3]  estimates of SE parameters (pk), and [4] estimates CP parameters 
  (ab) for each of the carcasses. The row names of the Aj matrix are the
  names of the units where each carcass was found.
}
\description{
Estimate g values and arrival intervals for a set of carcasses 
  from fitted pk and cp models and search data
}
\examples{
NA
}
