% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_utilities.R
\name{ftpLink}
\alias{ftpLink}
\title{Create a Link to the FTP-Housed Document of Interest}
\usage{
ftpLink(doc = "UserGuide")
}
\arguments{
\item{doc}{"UserGuide" or "Models".}
}
\value{
Character element of the link to the document.
}
\description{
The GenEst User Guide and Models Document live on the USGS
  FTP and can be linked to, rather than stored in the app. This function
  provides the link of interest.
}
