% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_utilities.R
\name{GenEstInlineCSS}
\alias{GenEstInlineCSS}
\title{Generate the Inline CSS Definition}
\usage{
GenEstInlineCSS(...)
}
\arguments{
\item{...}{Arguments to be passed down to \code{\link[shinyjs]{inlineCSS}}.}
}
\description{
Defines the inline CSS code for the GenEst app.
}
