% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_utilities.R
\name{GenEstShinyJS}
\alias{GenEstShinyJS}
\title{Shiny Java Script (via shinyjs) Enabler}
\usage{
GenEstShinyJS(...)
}
\arguments{
\item{...}{Arguments to be passed down to 
\code{\link[shinyjs]{useShinyjs}}.}
}
\description{
Enable the use of the \code{shinyjs} package within GenEst.
  This function is a simple wrapper to bring 
  \code{\link[shinyjs]{useShinyjs}} into the namespace.
}
