% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_utilities.R
\name{initialOutput}
\alias{initialOutput}
\title{Update the output list upon initiation of the app}
\usage{
initialOutput(rv, output)
}
\arguments{
\item{rv}{Reactive values list for the GenEst GUI.}

\item{output}{\code{output} list for the GenEst GUI.}
}
\value{
Updated \code{output} list.
}
\description{
Update the output list when the app is initialized.
}
