% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{isNeverDecreasing}
\alias{isNeverDecreasing}
\title{Is a vector never decreasing?}
\usage{
isNeverDecreasing(x, tiesOK = TRUE, na.rm = TRUE)
}
\arguments{
\item{x}{Vector of numeric values.}

\item{tiesOK}{Logical indicator if ties are ok or not.}

\item{na.rm}{Logical indicator if NAs are to be removed or not.}
}
\value{
Logical value.
}
\description{
Check if a vector is never decreasing.
}
