% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_widgets.R
\name{kFixedWidget}
\alias{kFixedWidget}
\alias{kFixedWidgetHeader}
\alias{kFixedWidgetRow}
\title{Make a kFixed Widget}
\usage{
kFixedWidget(sizeclasses)

kFixedWidgetHeader(sizeclasses)

kFixedWidgetRow(sizeclasses, sci)
}
\arguments{
\item{sizeclasses}{Vector of size class names (from the reactive values 
list).}

\item{sci}{Numeric size class element index.}
}
\value{
Rendered HTML kFixed input widget.
}
\description{
Produce a kFixed input widget based on the size classes.
}
\details{
\code{kFixedWidgetHeader} creates the widget header based on
  the number of size classes.

\code{kFixedWidgetRow} creates a row of the widget (input
  for one size class).
}
