% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_element_utility_functions.R
\name{makekFillInput}
\alias{makekFillInput}
\title{Make a kFill selection menu}
\usage{
makekFillInput(rv)
}
\arguments{
\item{rv}{reactive values list}
}
\value{
rendered HTML k selection menu object
}
\description{
Produce a size-based kFill selection menu object based on
  column selection
}
