% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{makeMenu}
\alias{makeMenu}
\title{Make a model menu}
\usage{
makeMenu(mods, sizeclasses, type)
}
\arguments{
\item{mods}{models}

\item{sizeclasses}{size class names}

\item{type}{"SE" or "CP"}
}
\value{
rendered model menu object
}
\description{
Make a model menu
}
