% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{matchCells}
\alias{matchCells}
\title{Match cells between a (potentially) reduced and full model}
\usage{
matchCells(model_spec, model_full)
}
\arguments{
\item{model_spec}{specific model compared against the full}

\item{model_full}{full model to compare to the specific}
}
\value{
vector of length equal to the number of cells in the full model, 
 each element containing the related cell name in the reduced model
}
\description{
Terms are automatically alphabatized to insure matching across models.
}
