% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{matchCells}
\alias{matchCells}
\title{Match cells between models.}
\usage{
matchCells(specific, modelSet)
}
\arguments{
\item{specific}{specific model compared against the full set}

\item{modelSet}{full model set to compare to the specific}
}
\value{
vector of length equal to the number of cells in the full model, 
  each element containing the related cell name in the reduced model
}
\description{
Match cells between a (potentially) reduced and the full model
  set. Terms are automatically alphabatized to insure matching across 
  models.
}
