% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_welcome_functions.R
\name{modalTextWest}
\alias{modalTextWest}
\title{Create West text for the modal welcome}
\usage{
modalTextWest()
}
\value{
HTML element of the West text
}
\description{
Create West text for the modal welcome
}
