% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_widgets.R
\name{modelInputWidget}
\alias{modelInputWidget}
\title{Create a Model Input Widget for the GenEst User Interface HTML}
\usage{
modelInputWidget(inType)
}
\arguments{
\item{inType}{Toggle control for the input type of the widget. One of 
"nsim", "CL", "class", "obsSE", "predsSE", "kFixed", "ltp", "fta", 
"predsCP", "dist", "frac", "DWPCol", "COdate", "gSearchInterval", 
"gSearchMax", "useSSinputs", or "useSSdata".}
}
\value{
HTML for the model input widget.
}
\description{
This is a generalized function for creating a model input 
  widget used in the GenEst GUI, based on the input type (\code{inType}).
}
