% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgModDone}
\alias{msgModDone}
\title{Create a model done message}
\usage{
msgModDone(msgs, rv, type = "SE", clear = TRUE)
}
\arguments{
\item{msgs}{message list}

\item{rv}{reactive values list}

\item{type}{"SE", "CP", "M", "split", or "g"}

\item{clear}{if all notifications should be cleared or not}
}
\value{
an SE model done message
}
\description{
Produces a model-done notification
}
