% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{msgModFail}
\alias{msgModFail}
\title{Creates the error message for when no models are fit successfully}
\usage{
msgModFail(mods)
}
\arguments{
\item{mods}{(fully failed) model object}
}
\value{
a model fail error message
}
\description{
Creates the error message for when no models are fit successfully
}
