% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgModFail}
\alias{msgModFail}
\title{Create the error message for when no models are fit successfully}
\usage{
msgModFail(msgs, clear = TRUE)
}
\arguments{
\item{msgs}{message list}

\item{clear}{if all notifications should be cleared or not}
}
\value{
a model fail error message
}
\description{
Produces a notification for complete model failure
}
