% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_message_functions.R
\name{msgModFail}
\alias{msgModFail}
\title{Create the error message for when no models are fit successfully}
\usage{
msgModFail(mods)
}
\arguments{
\item{mods}{(fully failed) model object}
}
\value{
a model fail error message
}
\description{
Produces a notification for complete model failure
}
