% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgModFail}
\alias{msgModFail}
\title{Create the error message for when no models are fit successfully}
\usage{
msgModFail(mods, type = "SE", special = NULL)
}
\arguments{
\item{mods}{(fully failed) model object}

\item{type}{"SE", "CP", "M", or "g"}

\item{special}{indicator of a special type of message}
}
\value{
a model fail error message
}
\description{
Produces a notification for complete model failure
}
