% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{msgModPartialFail}
\alias{msgModPartialFail}
\title{Creates the warning message for when only some models are fit successfully}
\usage{
msgModPartialFail()
}
\value{
a partial model fail warning
}
\description{
Creates the warning message for when only some models are fit successfully
}
