% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgModPartialFail}
\alias{msgModPartialFail}
\title{Create the warning message text for when only some models are fit 
  successfully}
\usage{
msgModPartialFail(mods, type = "SE")
}
\arguments{
\item{mods}{Set Size list of models}

\item{type}{"SE" or "CP"}
}
\value{
a partial model fail warning text
}
\description{
Produces text for a notification for partial model failure
}
