% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgModRun}
\alias{msgModRun}
\title{Create a model running message}
\usage{
msgModRun(msgs, modelType, clear = TRUE)
}
\arguments{
\item{msgs}{message list}

\item{modelType}{"SE", "CP", "g", or "M"}

\item{clear}{if all notifications should be cleared or not}
}
\value{
a model running message
}
\description{
Produces a model-running notification
}
