% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_message_functions.R
\name{msgSampleSize}
\alias{msgSampleSize}
\title{Create the small sample size warning}
\usage{
msgSampleSize(mods)
}
\arguments{
\item{mods}{models to check for sample sizes}
}
\value{
a small sample size message
}
\description{
Produces a notification for small sample sizes
}
