% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgSampleSize}
\alias{msgSampleSize}
\title{Create the warning message text for small sample sizes}
\usage{
msgSampleSize(mods)
}
\arguments{
\item{mods}{Set Size list of models}
}
\value{
small sample sizes warning text (if needed)
}
\description{
Produces text for a notification for small sample sizes
}
