% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgSplitFail}
\alias{msgSplitFail}
\title{Create the fail message for when splits aren't done correctly}
\usage{
msgSplitFail(type = NULL)
}
\arguments{
\item{type}{"setup" or "run"}
}
\value{
a split fail warning
}
\description{
Produces a notification for failed mortality splits
}
