% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{msgSSavgFail}
\alias{msgSSavgFail}
\title{Creates the warning message for when the SS average doesn't work}
\usage{
msgSSavgFail()
}
\value{
an average SS fail warning
}
\description{
Creates the warning message for when the SS average doesn't work
}
