% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_msg_functions.R
\name{msgSSavgFail}
\alias{msgSSavgFail}
\title{Create the warning message for when the SS average doesn't work}
\usage{
msgSSavgFail(msgs, rv, clear = TRUE)
}
\arguments{
\item{msgs}{message list}

\item{rv}{reactive values list}

\item{clear}{if all notifications should be cleared or not}
}
\value{
an average SS fail warning
}
\description{
Produces a notification for when an average search schedule
  can't be created
}
