% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{pickSizeclass}
\alias{pickSizeclass}
\title{Locate the sizeclass selected by the inputs}
\usage{
pickSizeclass(sizeclasses, choice)
}
\arguments{
\item{sizeclasses}{size class options}

\item{choice}{size class chosen}
}
\value{
location of the size class chosen
}
\description{
Locate the selection of a size class from the size class 
  column, retuning the first option from the size classes if the selection
  is not available.
}
