% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{pickSizeclass}
\alias{pickSizeclass}
\title{Locate the sizeclass selected by the inputs}
\usage{
pickSizeclass(sizeclasses, choice)
}
\arguments{
\item{sizeclasses}{carcass class options}

\item{choice}{carcass class chosen}
}
\value{
location of the carcass class chosen
}
\description{
Locate the selection of a carcass class from the carcass class
  column, retuning the first option from the carcass classes if the selection
  is not available.
}
