% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkm}
\alias{pkm}
\title{Fit a single searcher efficiency model.}
\usage{
pkm(pformula, kformula = NULL, data, obs_cols = NULL, fixed_k = NULL,
  k_init = 0.7)
}
\arguments{
\item{pformula}{Formula for p; an object of class "\code{\link{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted. Details of model specification are given under 
'Details'.}

\item{kformula}{Formula for k; an object of class "\code{\link{formula}}"
(or one that can be coerced to that class): a symbolic description of the
model to be fitted. Details of model specification are given under}

\item{data}{Dataframe with results from searcher efficiency trials and any
covariates included in \code{pformula} or {kformula} (required).}

\item{obs_cols}{Vector of names of columns in \code{data} where results 
for each search occasion are stored (optional). If no \code{obs_cols} are 
provided, \code{pkm} uses as \code{obs_cols} all columns with names that 
begin with an \code{'s'} or \code{'S'} and end with a number, e.g., 's1',
's2', 's3', etc. This option is included as a convenience for the user, 
but care must be taken that other data are not stored in columns with 
names matching that pattern. Alternatively, \code{obs_cols} may be 
entered as a vector of names, like \code{c('s1', 's2', 's3')}, 
\code{paste0('s', 1:3)}, or \code{c('initialSearch', 'anotherSearch', 
'lastSearch')}.}

\item{fixed_k}{Parameter for user-specified \code{k} value (optional). If a
value is provided, \code{kformula} is ignored and the model is fit under 
the assumption that the \code{k} parameter is fixed and known to be
\code{fix_k}.}

\item{init_k}{Initial value used for \code{k} in the optimization.}
}
\value{
\code{pkm} returns an object of class "\code{pkm}", which is a list
whose components characterize the fit of the model. Due to the large number
and complexity of components, only a subset of them is printed 
automatically; the rest can be viewed/accessed directly via the \code{$}
operator if desired.

The following components are displayed automatically:

\describe{
 \item{\code{call}}{the function call to fit the model}
 \item{\code{predictors}}{list of covariates of \code{p} and/or \code{k}}
 \item{\code{cellwise_pk}}{summary statistics for estimated cellwise 
   \code{p} and \code{k}, including the medians and upper & lower bounds
   on 95% CIs for each parameter, indexed by cell (or combination of
   covariate levels).}
 \item{\code{AIC}}{the 
   \href{https://en.wikipedia.org/wiki/Akaike_information_criterion}{AIC}
   value for the fitted model}
 \item{\code{AICc}}{the AIC value as corrected for small sample size}
 \item{\code{convergence}}{convergence status of the numerical optimization 
   to find the maximum likelihood estimates of \code{p} and \code{k}. A 
   value of \code{0} indicates that the model was fit successfully. For 
   help in deciphering other values, see \code{\link{optim}}.}
}

The following components are not printed automatically but can be accessed
via the \code{$} operator:
\describe{
  \item{\code{p_formula}}{the model formula for the \code{p} parameter}
  \item{\code{k_formula}}{the model formula for the \code{k} parameter}
  \item{\code{beta_hat_p}}{parameter estimates for the terms in the 
    regression model for for \code{p} (logit scale)}
  \item{\code{beta_hat_k}}{parameter estimates for the terms in the 
    regression model for for \code{k} (logit scale). If \code{k} is fixed 
    and known, \code{betahat_k} is not calculated.}
  \item{\code{beta_var}}{the variance-covariance matrix of the estimators
    for \code{c(betahat_p, betahat_k}.}
  \item{\code{mm_p_cells}}{a cellwise design matrix for covariate structure
    of \code{p_formula}}
  \item{\code{mm_k_cells}}{a cellwise design matrix for covariate structure 
    of \code{k_formula}}
  \item{\code{p_levels}}{all levels of each covariate of \code{p}}
  \item{\code{k_levels}}{all levels of each covariate of \code{k}}
  \item{\code{n_beta_p, n_beta_k}}{number of parameters to fit the \code{p}
    and \code{k} models}
  \item{\code{cells}}{cell structure of the pk-model, i.e., combinations of
    all levels for each covariate of \code{p} and \code{k}. For example, if
    \code{covar1} has levels \code{"a"}, \code{"b"}, and \code{"c"}, and
    \code{covar2} has levels \code{"X"} and \code{"Y"}, then the cells 
    would consist of \code{a.X}, \code{a.Y}, \code{b.X}, \code{b.Y}, 
    \code{c.X}, and \code{c.Y}.}
  \item{\code{n_cells}}{total number of cells}
 \item{\code{p_predictors}}{list of covariates of \code{p}}
 \item{\code{k_predictors}}{list of covariates of \code{k}}
 \item{\code{observations}}{observations used to fit the model}
 \item{\code{fixed_k}}{the input \code{fixed_k}}
}
}
\description{
Searcher efficiency is modeled as a function of the number of times a 
carcass has been missed in previous searches and any number of covariates. 
Format and usage parallel that of common \code{R} functions such as 
\code{lm}, \code{glm}, and \code{gam}. However, the input data 
(\code{data}) is structured differently to accommodate the multiple-search
searcher efficiency trials (see 'Details'), and model formulas may be 
entered for both \code{p} (akin to an intercept) and \code{k} (akin to a 
slope).
}
\details{
The probability of finding a carcass that is present at the time of search
is \code{p} on the first search after carcass arrival and is assumed to
decrease by a factor of \code{k} each time the carcass is missed in 
searches. Both \code{p} and \code{k} may depend on covariates such as 
ground cover, season, species, etc., and a separate model format 
(\code{pformula} and \code{kformula}) may be entered for each. The models 
are entered as they would be in the familiar \code{lm} or \code{glm} 
functions in R. For example, \code{p} might vary with \code{visibility}, 
\code{season}, and \code{site}, while \code{k} varies only with 
\code{visibility}. A user might then enter \code{p ~ visibility + season + 
site} for \code{pformula} and \code{k ~ visibility} for \code{kformula}. 
Other R conventions for defining formulas may also be used, with 
\code{covar1:covar2} for the interaction between covariates 1 and 2 and 
\code{covar1 * covar2} as short-hand for \code{covar1 + covar2 + 
covar1:covar2}.

Search trial \code{data} must be entered in a data frame with data in 
each row giving the fate of a single carcass in the field trials. There
must be a column for each search occassion, with 0, 1, or NA depending on 
whether the carcass was missed, found, or not available (typically because 
it was found and removed on a previous search, had been earlier removed by 
scavengers, or was not searched for) on the given search occasion. 
Additional columns with values for categorical covariates (e.g., 
visibility = E, M, or D) may also be included.
}
\examples{
data(pkmdat)
pkm(p ~ visibility, k ~ 1, data = pkmdat)
pkm(p ~ visibility * season, k ~ site, data = pkmdat)
pkm(p ~ visibility, k = 0.7, data = pkmdat)
}
