% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{pkmParamPlot}
\alias{pkmParamPlot}
\title{Plot parameter box plots for each cell for either p or k}
\usage{
pkmParamPlot(model, pk = "p", col)
}
\arguments{
\item{model}{model of class pkm}

\item{pk}{character of "p" or "k" to delineate between parameter graphed}

\item{col}{color to use}
}
\value{
a parameter plot panel
}
\description{
Boxplot for pk model cells (soon to be deprecated)
}
