% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmSetAICcTab}
\alias{pkmSetAICcTab}
\title{Create the AICc tables for a set of searcher efficiency models}
\usage{
pkmSetAICcTab(pkmset, quiet = FALSE)
}
\arguments{
\item{pkmset}{Set of searcher efficiency models fit to the same
observations}

\item{quiet}{Logical indicating if messages should be printed}
}
\value{
AICc table
}
\description{
Generates model comparison tables based on AICc values for
  a set of CP models generated by \code{\link{pkmSet}}
}
\examples{
  data(wind_RP)
  mod <- pkmSet(formula_p = p ~ Season, formula_k = k ~ Season, 
           data = wind_RP$SE
         )
 pkmSetAICcTab(mod)

}
