% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmSetAICcTab}
\alias{pkmSetAICcTab}
\title{Create the  AICc tables for the searcher efficiency models}
\usage{
pkmSetAICcTab(pkmset, quiet = FALSE)
}
\arguments{
\item{pkmset}{Set of searcher efficiency models fit to the same
observations}

\item{quiet}{Logical indicating if messages should be printed}
}
\value{
AICc table
}
\description{
Create the  AICc tables for the searcher efficiency models
}
\examples{
NA
}
