% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmSetFail}
\alias{pkmSetFail}
\title{Check if pkm models fail}
\usage{
pkmSetFail(pkmSetToCheck)
}
\arguments{
\item{pkmSetToCheck}{A \code{\link{pkmSet}} object to test}
}
\value{
A vector of logical values indicating if each of the models failed
}
\description{
Run a check on each model within a \code{\link{pkmSet}} 
  object to determine if it failed or not
}
