% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{pkmSetSizeFail}
\alias{pkmSetSizeFail}
\title{Check if all of the pkm models fail}
\usage{
pkmSetSizeFail(pkmSetSizeToCheck)
}
\arguments{
\item{pkmSetSizeToCheck}{A \code{pkmSetSize} object to test}
}
\value{
A list of logical vectors indicating which models failed
}
\description{
Run a check on each model within a \code{\link[=pkm]{pkmSetSize}}
  object to determine if they all failed or not
}
