% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{pkmSetSpecParamPlot}
\alias{pkmSetSpecParamPlot}
\title{p or k box plots for an SE model set}
\usage{
pkmSetSpecParamPlot(modelSet, specificModel, pk = "p", cols)
}
\arguments{
\item{modelSet}{modelSet of class pkmSet}

\item{specificModel}{name of the specific submodel to plot}

\item{pk}{character of "p" or "k" to delineate between parameter graphed}

\item{cols}{named vector of colors to use for the specific and reference
models}
}
\value{
a specific parameter plot panel
}
\description{
Plot parameter box plots for each cell within a model for
  either p or k with comparison to the cellwise model (soor to be deprecated)
}
