% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{pkmSetSpecParamPlot}
\alias{pkmSetSpecParamPlot}
\title{Plot parameter box plots for each cell within a model for either p or k 
  with comparison to the cellwise model}
\usage{
pkmSetSpecParamPlot(modelSet, specificModel, pk = "p", n, seed_spec,
  seed_full, col_spec, col_full)
}
\arguments{
\item{modelSet}{modelSet of class pkmSet}

\item{specificModel}{name of the specific submodel to plot}

\item{pk}{character of "p" or "k" to delineate between parameter graphed}

\item{n}{number of draws to use to characterize the distributions}

\item{seed_spec}{random number seed to use for the specific model}

\item{seed_full}{random number seed to use for the full model}

\item{col_spec}{color to use for the specific model}

\item{col_full}{color to use for the specific model}
}
\description{
Plot parameter box plots for each cell within a model for either p or k 
  with comparison to the cellwise model
}
