% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{pkmSetSpecSECellPlot}
\alias{pkmSetSpecSECellPlot}
\title{Plot cell-specific decay curve for searcher efficiency for a specific model
  with comparison to the cellwise model}
\usage{
pkmSetSpecSECellPlot(modelSet, specificModel, specificCell, cols, axes)
}
\arguments{
\item{modelSet}{modelSet of class pkmSet (soon to be deprecated)}

\item{specificModel}{name of the specific submodel to plot}

\item{specificCell}{name of the specific cell to plot}

\item{cols}{named vector of colors to use for the specific and reference
models}

\item{axes}{named vector of logical values indicating whether or not to
plot the x axis and the y axis}
}
\value{
a specific cell plot panel
}
\description{
Plot cell-specific decay curve for searcher efficiency for a specific model
  with comparison to the cellwise model
}
