% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{plot.cpm}
\alias{plot.cpm}
\title{Plot results of a single cp model}
\usage{
\method{plot}{cpm}(model, n = 500, seed = 1, col = "black")
}
\arguments{
\item{model}{model of class cpm}

\item{n}{number of draws to use to characterize the distributions}

\item{seed}{random number seed to use}

\item{col}{color to use}
}
\description{
Plot results of a single cp model
}
