% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_figure_functions.R
\name{plot.cpmSet}
\alias{plot.cpmSet}
\title{Plot results of a set of CP models}
\usage{
\method{plot}{cpmSet}(x, specificModel = NULL, nsim = 500, seed = 1,
  col = c(exponential = rgb(0.8, 0.38, 0.56), weibull = rgb(1, 0.76, 0.15),
  loglogistic = rgb(0, 1, 1), lognormal = rgb(0, 0.41, 0.55)),
  sizeclassName = NULL, ...)
}
\arguments{
\item{x}{pk model set of class pkmSet}

\item{specificModel}{the name(s) or index number(s) of specific model(s) to 
restrict the plot}

\item{nsim}{number of draws to use to characterize the distributions}

\item{seed}{random number seed to use for the models}

\item{col}{color to use for the specific model}

\item{sizeclassName}{name of the size class if it is to be added to the
figure}

\item{...}{to be passed down}
}
\description{
Produce a set of figures for a set of CP models, as fit by
  \code{\link{cpmSet}}
}
\examples{
  data(wind_RP)
  mod <- cpmSet(formula_l = l ~ Season, formula_s = s ~ Season,  
           data = wind_RP$CP, left = "Left", right = "Right"
          )
 plot(mod)

}
