% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_figure_functions.R
\name{plot.gGeneric}
\alias{plot.gGeneric}
\title{Plot results of a single generic ghat estimation}
\usage{
\method{plot}{gGeneric}(x, sizeclassName = NULL, CL = 0.9, ...)
}
\arguments{
\item{x}{ghatGeneric output}

\item{sizeclassName}{name of the size class if it is to be added to the
figure}

\item{CL}{confidence level to use}

\item{...}{to be passed down}
}
\value{
a plot
}
\description{
Plot results of a single generic ghat estimation
}
