% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_figure_functions.R
\name{plot.gGenericSize}
\alias{plot.gGenericSize}
\title{Plot results of a set of size-based generic g estimations}
\usage{
\method{plot}{gGenericSize}(x, CL = 0.9, ...)
}
\arguments{
\item{x}{gGenericSize output}

\item{CL}{confidence level to use}

\item{...}{to be passed down}
}
\value{
a plot
}
\description{
Plot results of a set of size-based generic g estimations
}
