% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_figure_functions.R
\name{plot.ghatGeneric}
\alias{plot.ghatGeneric}
\title{Plot results of a single pk model}
\usage{
\method{plot}{ghatGeneric}(x, ..., CL = 0.9)
}
\arguments{
\item{x}{ghatGeneric output}

\item{...}{to be passed down}

\item{CL}{confidence level to use}
}
\value{
a plot
}
\description{
Plot results of a single pk model
}
