% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{plot.pkm}
\alias{plot.pkm}
\title{Plot results of a single pk model}
\usage{
\method{plot}{pkm}(x, col = NULL, CL = NULL, ...)
}
\arguments{
\item{x}{model of class pkm}

\item{col}{color to use}

\item{CL}{confidence level to show in boxplots and confidence bounds}

\item{...}{arguments to be passed to sub functions}
}
\value{
a plot
}
\description{
Plot a single \code{\link{pkm}} model
}
\examples{
  data(wind_RP)
  mod <- pkm(formula_p = p ~ Season, formula_k = k ~ 1, data = wind_RP$SE)
  plot(mod)

}
