% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{plot.pkm}
\alias{plot.pkm}
\title{Plot results of a single pk model}
\usage{
\method{plot}{pkm}(x, n = 1000, seed = 1, col = "black", ...)
}
\arguments{
\item{x}{model of class pkm}

\item{n}{number of draws to use to characterize the distributions}

\item{seed}{random number seed to use}

\item{col}{color to use}

\item{...}{arguments to be passed to sub functions}
}
\description{
Plot results of a single pk model
}
