% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{plot.pkmSet}
\alias{plot.pkmSet}
\title{Plot results of a set of SE models}
\usage{
\method{plot}{pkmSet}(x, specificModel = NULL, nsim = 1000,
  seed_spec = NULL, seed_full = NULL, col_spec = "black",
  col_full = "grey", sizeclassName = NULL, ...)
}
\arguments{
\item{x}{pk model set of class pkmSet}

\item{specificModel}{the name(s) or index number(s) of specific model(s) to
restrict the plot}

\item{nsim}{number of draws to use to characterize the distributions}

\item{seed_spec}{random number seed to use for the specific model}

\item{seed_full}{random number seed to use for the full model}

\item{col_spec}{color to use for the specific model}

\item{col_full}{color to use for the specific model}

\item{sizeclassName}{name of the size class if it is to be added to the
figure}

\item{...}{to be sent to subfunctions}
}
\value{
a set of plots
}
\description{
Produce a set of figures for a set of SE models, as fit by
  \code{\link{pkmSet}}
}
\examples{
  data(wind_RP)
  mod <- pkmSet(formula_p = p ~ Season, formula_k = k ~ Season,
           data = wind_RP$SE
         )
  plot(mod)

}
