% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{plotSEBoxTemplate}
\alias{plotSEBoxTemplate}
\title{template box plot}
\usage{
plotSEBoxTemplate(modelSet, specificModel, cols)
}
\arguments{
\item{modelSet}{modelSet of class pkmSet}

\item{specificModel}{name of the specific submodel to plot}

\item{cols}{named vector of colors to use for the specific and reference
models}
}
\value{
a template box plot
}
\description{
Plot template box plot (soon to be deprecated)
}
