% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carcass_persistence_functions.R
\name{ppersist}
\alias{ppersist}
\title{Calculate the probability of persistence to detection}
\usage{
ppersist(pda, pdb, dist, t_arrive0, t_arrive1, t_search)
}
\arguments{
\item{pda}{parameter a.}

\item{pdb}{parameter b.}

\item{dist}{Distribution used.}

\item{t_arrive0}{Beginning of arrival window.}

\item{t_arrive1}{End of arrival window.}

\item{t_search}{Search time.}
}
\value{
Probability of persistence of detection to at t_search, 
         given arrival between t_arrive0 and t_arrive1
}
\description{
Given a set of CP parameters (of \code{"ppersist"} type), 
  calculate the probability of persistence to detection for a carcass.
}
