% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{prepPredictors}
\alias{prepPredictors}
\title{Prepare predictors based on inputs}
\usage{
prepPredictors(preds = NULL)
}
\arguments{
\item{preds}{predictors, as input to the app}
}
\value{
prepared predictors (or 1 if no predictors)
}
\description{
Prepare predictor inputs from the app for use in the model
  function
}
