% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{prepSizeclassText}
\alias{prepSizeclassText}
\title{Prepare text for carcass classes}
\usage{
prepSizeclassText(sizeclasses)
}
\arguments{
\item{sizeclasses}{names of the carcass classes}
}
\value{
prepared and render name text
}
\description{
Prepare and render text of the carcass class names
}
