% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_table_functions.R
\name{prettyModTabCP}
\alias{prettyModTabCP}
\title{Create the pretty version of the Carcass Persistence model table}
\usage{
prettyModTabCP(modTab)
}
\arguments{
\item{modTab}{\code{descCP} object or NULL}
}
\value{
pretty version of the CP model table in a data frame with point
 and interval estimates for medianCP and rI statistics. Output table is
 ready for rendering in shiny and posting in the GUI
}
\description{
Format a reader-friendly version of the parameter table from
 a carcass persistence model showing CIs for medianCP and for rI's for
 intervals of Ir
}
