% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_table_functions.R
\name{prettyModTabCP}
\alias{prettyModTabCP}
\title{Create the pretty version of the Carcass Persistence model table}
\usage{
prettyModTabCP(modTabs, CL = 0.9)
}
\arguments{
\item{modTabs}{model tables}

\item{CL}{Confidence level}
}
\value{
pretty version of the CP model table
}
\description{
Format a reader-friendly version of the parameter table from
  a Carcass Persistence model, based on confidence level of interest
}
