% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_table_functions.R
\name{prettyModTabSE}
\alias{prettyModTabSE}
\title{Create the pretty versions of model and summary tables}
\usage{
prettyModTabSE(modTab, CL = 0.9)
}
\arguments{
\item{modTab}{model table}

\item{CL}{Confidence level}
}
\value{
pretty version of the SE model table
}
\description{
Format reader-friendly versions of results summary tables for
 searcher efficiency (GUI display and download), carcass persistence, and
 splits.
}
