% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_table_functions.R
\name{prettyModTabSE}
\alias{prettyModTabSE}
\title{Create the pretty version of the Searcher Efficiency model table}
\usage{
prettyModTabSE(modTab, CL = 0.9)
}
\arguments{
\item{modTab}{model table}

\item{CL}{Confidence level}
}
\value{
pretty version of the SE model table
}
\description{
Format a reader-friendly version of the parameter table from
  a Searcher Efficiency model, based on confidence level of interest
}
