% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{prettySplitSpecTab}
\alias{prettySplitSpecTab}
\title{Pretty a specific split summary table}
\usage{
prettySplitSpecTab(splitSummarySpec)
}
\arguments{
\item{splitSummarySpec}{a specific split summary}
}
\value{
specific split pretty table
}
\description{
Pretty a specific split summary table
}
