% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_table_functions.R
\name{prettySplitTab}
\alias{prettySplitTab}
\title{Create the pretty version of the split summary table}
\usage{
prettySplitTab(splitSummary)
}
\arguments{
\item{splitSummary}{a split summary}
}
\value{
split pretty table
}
\description{
Format a reader-friendly version of the split summary table
  a mortality estimation
}
