% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{print.corpus_frame}
\alias{print.corpus_frame}
\title{Generic S3 function for printing corpus_frame}
\usage{
\method{print}{corpus_frame}(x, ...)
}
\arguments{
\item{x}{data frame to print}

\item{...}{other arguments}
}
\value{
prints data frame
}
\description{
Generic S3 function for printing corpus_frame
}
