% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{rghatCarcass}
\alias{rghatCarcass}
\title{Simulate ghat values for a carcasses from fitted pk and cp models 
  and search data}
\usage{
rghatCarcass(n = 1, data_carc, dist, data_ss, preds_SE, preds_CP, pkSim,
  cpSim, preds_static, preds_dynamic, unitColName)
}
\arguments{
\item{n}{the number of simulation draws}

\item{data_carc}{Carcass Observation data for the carcass}

\item{dist}{distribution for the CP model}

\item{data_ss}{Search Schedule data}

\item{preds_SE}{Names of Searcher Efficiency predictors}

\item{preds_CP}{Names of Carcass Persistence predictors}

\item{pkSim}{simulated pk parameters}

\item{cpSim}{simulated cp parameters}

\item{preds_static}{Names of predictors that don't change over time}

\item{preds_dynamic}{Names of predictors that change over time}

\item{unitColName}{Column name for the unit indicator}
}
\value{
n ghat estimates for a carcass
}
\description{
Simulate ghat values for a carcasses from fitted pk and cp models 
  and search data
}
\examples{
NA
}
