% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{rpk}
\alias{rpk}
\title{Simulate p and k parameters from a fitted pk model.}
\usage{
rpk(n = 1, model, seed = NULL, kFill = NA)
}
\arguments{
\item{n}{the number of simulation draws}

\item{model}{A \code{\link{pkm}} object (which is returned from 
\code{pkm()})}

\item{seed}{optional input to set the seed of the RNG}

\item{kFill}{what value to fill k with if k was not estimated}
}
\value{
list of two matrices of \code{n} simulated \code{p} and \code{k} 
  for cells defined by the \code{model} object.
}
\description{
Simulate p and k parameters from a fitted pk model.
}
\examples{
  NA
}
