% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{rpk}
\alias{rpk}
\title{Simulate p and k parameters from a fitted pk model.}
\usage{
rpk(n = 1, pk_model, seed = NULL)
}
\arguments{
\item{n}{the number of simulation draws}

\item{pk_model}{A \code{\link{pkm}} object (which is returned from 
\code{pkm()})}

\item{seed}{optional input to set the seed of the RNG}
}
\value{
list of two matrices of \code{n} simulated \code{p} and \code{k} 
 for cells defined by the \code{pk_model} object.
}
\description{
Simulate p and k parameters from a fitted pk model.
}
\examples{
data(pkmdat)
pk_mod_1 <- pkm(p ~ 1, k ~ 1, data = pkmdat)
simulated_pk <- rpk(n_sim = 10, pk_model = pk_mod_1)
simulated_pk

pk_mod_2 <- pkm(p ~ visibility * season, k ~ site, data = pkmdat)
rpk(n_sim = 10, pk_model = pk_mod_2)
}
