% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{selectData}
\alias{selectData}
\title{Select particular columns from a data set}
\usage{
selectData(data, cols)
}
\arguments{
\item{data}{data table to select from}

\item{cols}{column names to select}
}
\value{
selected data
}
\description{
Convenience function for selecting specific columns from a 
  data table
}
