% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_figure_functions.R
\name{SEpanel}
\alias{SEpanel}
\title{Produce a single panel in an SE summary/diagnostic plot}
\usage{
SEpanel(found, available, y_spc, y_ref, xends, cols_SE)
}
\arguments{
\item{found}{vector of number carcasses found on the ith attempt}

\item{available}{vector of number carcasses found on the ith attempt}

\item{y_spc}{vector of model fits for the specific model}

\item{y_ref}{vector of model fits for the reference model}

\item{xends}{x-axis buffer (numeric scalar) on sides of figs}

\item{cols_SE}{named vector of colors (character)}
}
\value{
NULL inserts a panel with no labels into a preformatted figure
}
\description{
Each call to \code{SEpanel} produces a single panel showing
 searcher efficiency as a function of number of searches. Includes raw data
 (\code{found} and \code{available}) and model fits for a specific model
 (\code{y_spc}) and for the reference model (\code{y_ref}) for the
 \code{pkmSet} object from the reference model was extracted. For interal use
 only, for producing figs for \code{plot.pkmSet}.
}
