% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{SEsi}
\alias{SEsi}
\title{Calculate decayed searcher efficiency}
\usage{
SEsi(days, pk)
}
\arguments{
\item{days}{search days}

\item{pk}{\code{p} and \code{k} values}
}
\value{
searcher efficiency that matches the output of ppersist
}
\description{
Calculate searcher efficiency after some searches under 
  pk values
}
