% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searcher_efficiency_functions.R
\name{SEsi0}
\alias{SEsi0}
\title{Calculate decayed searcher efficiency for a single pk}
\usage{
SEsi0(days, pk)
}
\arguments{
\item{days}{search days}

\item{pk}{pk combination}
}
\value{
searcher efficiency that matches the output of ppersist
}
\description{
Calculate searcher efficiency after some searches for a 
  single pk combination
}
