% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_figure_functions.R
\name{setFigH}
\alias{setFigH}
\title{Set the figure height based on the number of cells}
\usage{
setFigH(modSet, minH, type = "SE")
}
\arguments{
\item{modSet}{model set}

\item{minH}{minimum height}

\item{type}{"SE" or "CP"}
}
\value{
figure height
}
\description{
Convenience function for determining the needed figure height
}
