% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_figure_size_functions.R
\name{setFigH}
\alias{setFigH}
\title{Set the figure height based on the number of cells}
\usage{
setFigH(modelSet, type = "SE")
}
\arguments{
\item{type}{"SE" or "CP"}

\item{modelelSet}{model set}

\item{minH}{minimum height}
}
\value{
figure height
}
\description{
Convenience function for determining the needed figure height
}
