% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_figure_size_functions.R
\name{setFigW}
\alias{setFigW}
\title{Set the figure width based on the number of cells}
\usage{
setFigW(modelSet)
}
\arguments{
\item{modelSet}{model set}
}
\value{
figure width
}
\description{
Convenience function for determining the needed figure width
}
