% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_output_utilities.R
\name{setFigW}
\alias{setFigW}
\alias{setFigH}
\title{Set Figure width and height based on the number of cells}
\usage{
setFigW(modelSet)

setFigH(modelSet, modType = "SE")
}
\arguments{
\item{modelSet}{Model set of class \code{cpmSet} or \code{pkmSet}.}

\item{modType}{"SE" or "CP"}
}
\value{
\code{setFigW}: Figure width.

\code{setFigH}: Figure height.
}
\description{
Convenience functions for determining the needed figure sizes.
}
