% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{setkFillNeed}
\alias{setkFillNeed}
\title{Creates the kFillNeed text}
\usage{
setkFillNeed(obsCols)
}
\arguments{
\item{obsCols}{vector of observation column names}
}
\value{
kFillNeed
}
\description{
Creates the kFillNeed text
}
