% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{setkFix}
\alias{setkFix}
\title{Update the fixed k value}
\usage{
setkFix(kFixedChoice, kFixed)
}
\arguments{
\item{kFixedChoice}{choice to fix k (1) or not (anything else)}

\item{kFixed}{existing kFixed value}
}
\value{
new kFixed value
}
\description{
Update the value for \code{kFixed} is chosen and available
}
