% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_figure_functions.R
\name{simpleMplot}
\alias{simpleMplot}
\title{Plot a total mortality estimation for a simple situation}
\usage{
simpleMplot(M, ..., Xmin = 0, CL = 0.9)
}
\arguments{
\item{M}{Mortality object}

\item{...}{arguments to pass down}

\item{Xmin}{minimum number of observable carcasses}

\item{CL}{confidence level}
}
\description{
Function underneath \code{\link{plot.estM}}, which 
  defines the plot method for a mortality object, composed of a hisogram 
  with the empirical PDF and summary statistics
}
