% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{sizeCols}
\alias{sizeCols}
\title{Select the potential carcass class columns from a data table}
\usage{
sizeCols(data)
}
\arguments{
\item{data}{data table}
}
\value{
column names of columns that can be carcass class values
}
\description{
Simple function to facilitate selection of columns that could
  be carcass class values from a data table.
}
