% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_ui_utilities.R
\name{style}
\alias{style}
\alias{ol}
\alias{ul}
\alias{li}
\alias{b}
\alias{u}
\alias{small}
\alias{big}
\alias{center}
\title{HTML tag functions}
\usage{
style(...)

ol(...)

ul(...)

li(...)

b(...)

u(...)

small(...)

big(text = NULL)

center(text = NULL)
}
\arguments{
\item{...}{attributes and children of the element}

\item{text}{text to wrap in the tag}
}
\value{
HTML tagged elements
}
\description{
This suite of functions are used for producing specific
  HTML tags, in complement to those imported from the htmltools package.
  \cr \cr \code{style}: in-line style.

\code{ol}: ordered list.

\code{ul}: unordered list.

\code{li}: list element.

\code{b}: bolded text.

\code{u}: underlined text.

\code{small}: small text.

\code{big}: big text.

\code{center}: center-justified text.
}
