% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_utility_functions.R
\name{trimSetSize}
\alias{trimSetSize}
\title{Trim a Model-Set-Size Complex to a Single Model Per Size}
\usage{
trimSetSize(modSetSize, mods)
}
\arguments{
\item{modSetSize}{modSetSize complex (cpm or pkm)}

\item{mods}{named (according to carcass classes) vector of model names to use}
}
\value{
modSetSize reduced to a single model per carcass class
}
\description{
Select a single model from each carcass class (based on the model
  names).
}
