% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trueLength}
\alias{trueLength}
\title{Get the length of real things}
\usage{
trueLength(x)
}
\arguments{
\item{x}{vector of values}
}
\value{
integer of how many non-NA values in x
}
\description{
Length of non-missing values in a vector
}
\examples{
  x <- c(1, 2, NA, 3)
  length(x)
  trueLength(x)

}
