% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{trueLength}
\alias{trueLength}
\title{Get the length of real (non-NA) things}
\usage{
trueLength(x)
}
\arguments{
\item{x}{Vector of values, some of which many be \code{NA}.}
}
\value{
Integer count of how many non-\code{NA} values in \code{x}.
}
\description{
Length of non-missing (non-\code{NA}) values in a vector.
}
\examples{
  x <- c(1, 2, NA, 3)
  length(x)
  trueLength(x)

}
