% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{updateColNames_all}
\alias{updateColNames_all}
\title{Update the string of column names that are in all the needed tables}
\usage{
updateColNames_all(rv)
}
\arguments{
\item{rv}{reactive values list}
}
\value{
possible column names
}
\description{
Determine the overlap between the column names in the SE, CP,
   and CO data tables.
}
