% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utility_functions.R
\name{updateColNames_all}
\alias{updateColNames_all}
\title{Updates the string of column names that are in all the needed tables}
\usage{
updateColNames_all(colNames_all, colNames_new)
}
\arguments{
\item{colNames_all}{current vector of column names in all needed tables}

\item{colNames_new}{column names of the new table to bring in}
}
\value{
updated colNames_all
}
\description{
Updates the string of column names that are in all the needed tables
}
