% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{updateSizeclasses}
\alias{updateSizeclasses}
\title{Update the size classes}
\usage{
updateSizeclasses(data, sizeCol)
}
\arguments{
\item{data}{data table to draw sizes from}

\item{sizeCol}{size class column name}
}
\value{
unique size classes
}
\description{
Determine the options for size classes, based on a data table
  and column name, returning \code{NULL} if no size class column is
  provided
}
