% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_utilities.R
\name{updateSizeclasses}
\alias{updateSizeclasses}
\title{Update the carcass classes}
\usage{
updateSizeclasses(data, sizeCol)
}
\arguments{
\item{data}{data table to draw sizes from}

\item{sizeCol}{carcass class column name}
}
\value{
unique carcass classes
}
\description{
Determine the options for carcass classes, based on a data table
  and column name, returning \code{NULL} if no carcass class column is
  provided
}
