% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{whichCleanout}
\alias{whichCleanout}
\title{Determine which carcasses were from cleanout searches}
\usage{
whichCleanout(data_carc, data_ss, unitCol, dateFoundCol, dateSearchedCol)
}
\arguments{
\item{data_carc}{Carcass observation data}

\item{data_ss}{Search schedule data}

\item{unitCol}{Column name for the unit indicator}

\item{dateFoundCol}{Column name for the date found data}

\item{dateSearchedCol}{Column name for the date searched data}
}
\value{
index values of which carcasses were taken on the first search
}
\description{
Determine which carcasses were from cleanout searches
}
\examples{
NA
}
