% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection_probability_functions.R
\name{whichCleanout}
\alias{whichCleanout}
\title{Determine which carcasses were from cleanout searches}
\usage{
whichCleanout(data_CO, data_SS, unitCol, timeFoundCol, timeSearchedCol)
}
\arguments{
\item{data_CO}{Carcass observation data}

\item{data_SS}{Search schedule data}

\item{unitCol}{Column name for the unit indicator}

\item{timeFoundCol}{Column name for the time found data}

\item{timeSearchedCol}{Column name for the time searched data}
}
\value{
index values of which carcasses were taken on the first search
}
\description{
Determine which carcasses were from cleanout searches
}
\examples{
NA
}
