% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind_cleared}
\alias{wind_cleared}
\title{Synthetic data set based on searches of 60 m cleared plots of 23 out of 100
  turbines at a theoretical site.}
\format{A list with 5 elements:
\describe{
  \item{SE}{Searcher efficiency trial data}
  \item{CP}{Carcass persistence trial data}
  \item{SS}{Search schedule parameters}
  \item{DWP}{Density weighted proportion of area searched}
  \item{CO}{Carcass observations}
}}
\source{
cleared
}
\usage{
wind_cleared
}
\description{
Complete data set for fatality estimation at a wind power facility. 
  Included are SE, CP, SS, DWP, and CO data.
}
\keyword{datasets}
