% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wind_cleared}
\alias{wind_cleared}
\title{A complete example data set for estimating fatalities from 60 m cleared plots
at 23 out of 100 searches at a wind power facility}
\format{\code{wind_cleared} is a list with 5 elements:
\describe{
  \item{\code{SE}}{Searcher efficiency trial data}
  \item{\code{CP}}{Carcass persistence trial data}
  \item{\code{SS}}{Search schedule parameters}
  \item{\code{DWP}}{Density weighted proportion of area searched}
  \item{\code{CO}}{Carcass observations}
}}
\source{
cleared
}
\usage{
wind_cleared
}
\description{
Data on carcass observations (CO) from road and pad searches out to 120 m
from each of 100 turbines at a theoretical site, field trials for estimating
carcass persistence (CP) and searcher efficiency (SE), search schedule (SS)
parameters (for example, which turbines were searched on which days), and
density weighted proportion (DWP) of area searched at each turbine (which is
an area adjustment factor to account for incomplete search coverage).
}
\section{General Description}{
NA
}

\keyword{datasets}
