% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{windClearedPlotData}
\alias{windClearedPlotData}
\title{Description of the dataset}
\format{A list with 5 items:
\describe{
  \item{SE_trials}{Searcher efficiency trial data}
  \item{CP_trials}{Carcass persistence trial data}
  \item{SS_23cleared60}{Search schedule data}
  \item{DWP_23cleared60}{Density weighted proportion of area searched data}
  \item{CO_23cleared60}{Carcass observations}  
}}
\source{
Source name
}
\usage{
windClearedPlotData
}
\description{
A dataset containing SE, CP, SS, DWP, and CO data.
}
\keyword{datasets}
